Ganeti.Compat

Description

Compatibility helper module.

This module holds definitions that help with supporting multiple library versions or transitions between versions.

Documentation

filePath' :: Event -> FilePath

Wrappers converting ByteString filepaths to Strings and vice versa

hinotify 0.3.10 switched to using RawFilePaths instead of FilePaths, the former being Data.ByteString and the latter String.

maybeFilePath' :: Event -> Maybe FilePath

toInotifyPath :: FilePath -> RawFilePath

type Pid = CPid

Process 1.6.3. introduced the getPid function, for older versions provide an implemention here (https:/github.comhaskellprocesspull/109)

getPid' :: ProcessHandle -> IO (Maybe Pid)