Ganeti.Confd.ClientFunctions

Description

Some utility functions, based on the Confd client, providing data in a ready-to-use way.

Documentation

getInstances :: String -> Maybe String -> Maybe Int -> ResultT String IO ([Instance], [Instance])

Get the list of instances the given node is ([primary], [secondary]) for. The server address and the server port parameters are mainly intended for testing purposes. If they are Nothing, the default values will be used.

getDisks :: Instance -> Maybe String -> Maybe Int -> ResultT String IO [Disk]

Get the list of disks that belong to a given instance The server address and the server port parameters are mainly intended for testing purposes. If they are Nothing, the default values will be used.

getInstanceDisks :: String -> Maybe String -> Maybe Int -> ResultT String IO [(Instance, [Disk])]

Get the list of instances on the given node along with their disks The server address and the server port parameters are mainly intended for testing purposes. If they are Nothing, the default values will be used.