Ganeti.Cpu.Types

Description

CPUload data types

This module holds the definition of the data types describing the CPU load according to information collected periodically from procstat.

Documentation

data CPUavgload

This is the format of the report produced by the cpu load collector.

Constructors

CPUavgload

Fields

cavCpuNumber :: Int

cavCpus :: [Double]

cavCpuTotal :: Double

loadCPUavgload :: JSValue -> Result CPUavgload

saveCPUavgload :: CPUavgload -> JSValue

data CPUstat

This is the format of the data parsed by the input file.

Constructors

CPUstat

Fields

csName :: String

csUser :: Int

csNice :: Int

csSystem :: Int

csIdle :: Int

csIowait :: Int

csIrq :: Int

csSoftirq :: Int

csSteal :: Int

csGuest :: Int

csGuestNice :: Int

loadCPUstat :: JSValue -> Result CPUstat

saveCPUstat :: CPUstat -> JSValue