Ganeti.Daemon.Utils

Description

Utility functions for complex operations carried out by several daemons.

Documentation

verifyMasterVotes :: IO (Result Bool)

Gather votes from all nodes and verify that we we are the master. Return True if the voting is won, False if not enough

verifyMaster :: Int -> IO (Result ())

Verify, by voting, that this node is the master. Bad if we're not. Allow the given number of retries to wait for not available nodes.

handleMasterVerificationOptions :: DaemonOptions -> IO (Either ExitCode ())

Verify master position according to the options provided, usually by carrying out a voting. Either return unit on success, or a suggested exit code.