Ganeti.DataCollectors.CPUload

Description

procstat data collector.

Documentation

defaultFile :: FilePath

The default path of the CPU status file. It is hardcoded because it is not likely to change.

bufferSize :: Int

The buffer size of the values kept in the map.

windowSize :: Integer

The window size of the values that will export the average load.

defaultCharNum :: Int

The default setting for the maximum amount of not parsed character to print in case of error. It is set to use most of the screen estate on a standard 80x25 terminal. TODO: add the possibility to set this with a command line parameter.

dcName :: String

The name of this data collector.

dcVersion :: DCVersion

The version of this data collector.

dcFormatVersion :: Int

The version number for the data format of this data collector.

dcCategory :: Maybe DCCategory

The category of this data collector.

dcKind :: DCKind

The kind of this data collector.

dcReport :: Maybe CollectorData -> IO DCReport

The data exported by the data collector, taken from the default location.

type Buffer = Seq (ClockTime, [Int])

Data stored by the collector in mond's memory.

computeLoad :: CPUstat -> Int

Compute the load from a CPU.

dcCollectFromFile :: FilePath -> IO (ClockTime, [Int])

Reads and Computes the load for each CPU.

dcCollect :: IO Buffer

Returns the collected data in the appropriate type.

formatData :: [Double] -> CPUavgload

Formats data for JSON transformation.

updateEntry :: Buffer -> Buffer -> Buffer

Update a Map Entry.

dcUpdate :: Maybe CollectorData -> IO CollectorData

Updates the given Collector data.

computeAverage :: Buffer -> Integer -> Integer -> Result [Double]

Computes the average load for every CPU and the overall from data read from the map. Returns Bad if there are not enough values to compute it.

buildJsonReport :: Buffer -> IO JSValue

This function computes the JSON representation of the CPU load.

buildDCReport :: Buffer -> IO DCReport

This function computes the DCReport for the CPU load.