Ganeti.DataCollectors.Diskstats

Description

procdiskstats data collector.

Documentation

dcName :: String

The name of this data collector.

dcVersion :: DCVersion

The version of this data collector.

dcFormatVersion :: Int

The version number for the data format of this data collector.

dcCategory :: Maybe DCCategory

The category of this data collector.

dcKind :: DCKind

The kind of this data collector.

dcReport :: IO DCReport

The data exported by the data collector, taken from the default location.

Command line options

options :: IO [OptType]

arguments :: [ArgCompletion]

The list of arguments supported by the program.

main :: Options -> [String] -> IO ()

Main function.