Ganeti.DataCollectors.InstStatus

Description

Instance status data collector.

Documentation

dcName :: String

The name of this data collector.

dcVersion :: DCVersion

The version of this data collector.

dcFormatVersion :: Int

The version number for the data format of this data collector.

dcCategory :: Maybe DCCategory

The category of this data collector.

dcKind :: DCKind

The kind of this data collector.

dcReport :: IO DCReport

The report of this data collector.

Command line options

options :: IO [OptType]

arguments :: [ArgCompletion]

The list of arguments supported by the program.

getReasonTrail :: String -> IO ReasonTrail

Try to get the reason trail for an instance. In case it is not possible, log the failure and return an empty list instead.

computeStatusField :: AdminState -> ActualState -> DCStatus

Determine the value of the status field for the report of one instance

buildStatus :: Map String Domain -> Map Int UptimeInfo -> RealInstanceData -> IO InstStatus

computeGlobalStatus :: [InstStatus] -> DCStatus

Compute the status code and message, given the current DRBD data The final state will have the code corresponding to the worst code of all the devices, and the error message given from the concatenation of the non-empty error messages.

buildInstStatusReport :: Maybe String -> Maybe Int -> IO DCReport

Build the report of this data collector, containing all the information about the status of the instances.

main :: Options -> [String] -> IO ()

Main function.