Ganeti.DataCollectors.Lv

Description

Logical Volumes data collector.

Documentation

defaultCharNum :: Int

The default setting for the maximum amount of not parsed character to print in case of error. It is set to use most of the screen estate on a standard 80x25 terminal. TODO: add the possibility to set this with a command line parameter.

dcName :: String

The name of this data collector.

dcVersion :: DCVersion

The version of this data collector.

dcFormatVersion :: Int

The version number for the data format of this data collector.

dcCategory :: Maybe DCCategory

The category of this data collector.

dcKind :: DCKind

The kind of this data collector.

dcReport :: IO DCReport

The data exported by the data collector, taken from the default location.

Command line options

options :: IO [OptType]

arguments :: [ArgCompletion]

The list of arguments supported by the program.

getLvInfo :: Maybe FilePath -> IO [LVInfo]

Get information about logical volumes from file (if specified) or by actually running the command to get it from a live cluster.

getInstDiskList :: Options -> IO [(RealInstanceData, [Disk])]

Get the list of real instances on the current node along with their disks, either from a provided file or by querying Confd.

addInstNameToOneLv :: [(RealInstanceData, [Disk])] -> LVInfo -> LVInfo

Adds the name of the instance to the information about one logical volume.

addInstNameToLv :: [(RealInstanceData, [Disk])] -> [LVInfo] -> [LVInfo]

Adds the name of the instance to the information about logical volumes.

buildJsonReport :: Options -> IO JSValue

This function computes the JSON representation of the LV status.

buildDCReport :: Options -> IO DCReport

This function computes the DCReport for the logical volumes.

main :: Options -> [String] -> IO ()

Main function.