Ganeti.DataCollectors.XenCpuLoad

Description

xentop CPU data collector

Documentation

dcName :: String

The name of this data collector.

dcVersion :: DCVersion

The version of this data collector.

dcFormatVersion :: Int

The version number for the data format of this data collector.

dcCategory :: Maybe DCCategory

The category of this data collector.

dcKind :: DCKind

The kind of this data collector.

dcUpdate :: Maybe CollectorData -> IO CollectorData

Updates the given Collector data.

dcReport :: Maybe CollectorData -> IO DCReport

The data exported by the data collector, taken from the default location.