Ganeti.HTools.AlgorithmParams

Description

Algorithm Options for HTools

This module describes the parameters that influence the balancing algorithm in htools.

Documentation

data AlgorithmOptions

Constructors

AlgorithmOptions

Fields

algDiskMoves :: Bool

Whether disk moves are allowed

algInstanceMoves :: Bool

Whether instance moves are allowed

algRestrictedMigration :: Bool

Whether migration is restricted

algIgnoreSoftErrors :: Bool

Whether to always ignore soft errors

algEvacMode :: Bool

Consider only eavacation moves

algMinGain :: Double

Minimal gain per balancing step

algMinGainLimit :: Double

Limit below which minimal gain is used

algCapacity :: Bool

Whether to check capacity properties, like global N+1 redundancy

algCapacityIgnoreGroups :: [Gdx]

Groups to ignore in capacity checks

algRestrictToNodes :: Maybe [String]

nodes to restrict allocation to

algAcceptExisting :: Bool

accept existing violations in capacity checks

fromCLIOptions :: Options -> AlgorithmOptions

Obtain the relevant algorithmic option from the commandline options

defaultOptions :: AlgorithmOptions

Default options for the balancing algorithm