Ganeti.HTools.Backend.MonD

Description

Monitoring daemon backend

This module holds implements the querying of the monitoring daemons for dynamic utilisation data.

General definitions

Node-total CPU load average data collector

Xen instance CPU-usage collector

Collector choice

Querying infrastructure

pMonDData :: String -> Result [MonDData]

Parse MonD data file contents.

queryAllMonDDCs :: ClusterData -> Options -> WriterT All IO ClusterData

Query all MonDs for all Data Collector. Return the cluster enriched by dynamic data, as well as a bit indicating wether all collectors could be queried successfully.