Ganeti.HTools.Backend.Rapi

Description

Implementation of the RAPI client interface.

Documentation

filePrefix :: String

File method prefix.

getUrl :: MonadFail m => String -> IO (m String)

Read an URL via curl and return the body if successful.

connTimeout :: Long

Connection timeout (when using non-file methods).

queryTimeout :: Long

The default timeout for queries (when using non-file methods).

curlOpts :: [CurlOption]

The curl options we use.

ioErrToResult :: IO a -> IO (Result a)

Helper to convert I/O errors in Bad values.

formatHost :: String -> String

Append the default port if not passed in.

getInstances :: NameAssoc -> String -> Result [(String, Instance)]

Parse a instance list in JSON format.

getNodes :: NameAssoc -> String -> Result [(String, Node)]

Parse a node list in JSON format.

getGroups :: String -> Result [(String, Group)]

Parse a group list in JSON format.

parseInstance :: NameAssoc -> JSRecord -> Result (String, Instance)

Construct an instance from a JSON object.

parseNode :: NameAssoc -> JSRecord -> Result (String, Node)

Construct a node from a JSON object.

parseGroup :: JSRecord -> Result (String, Group)

Construct a group from a JSON object.

parseCluster :: JSObject JSValue -> Result ([String], IPolicy, String, Hypervisor)

Parse cluster data from the info resource.

readDataHttp

Arguments

:: String
Cluster or URL to use as source

-> IO (Result String, Result String, Result String, Result String)

Loads the raw cluster data from an URL.

readDataFile

Arguments

:: String
Path to the directory containing the files

-> IO (Result String, Result String, Result String, Result String)

Loads the raw cluster data from the filesystem.

readData

Arguments

:: String
URL to use as source

-> IO (Result String, Result String, Result String, Result String)

Loads data via either readDataFile or readDataHttp.

parseData :: (Result String, Result String, Result String, Result String) -> Result ClusterData

Builds the cluster data from the raw Rapi content.

loadData

Arguments

:: String
Cluster or URL to use as source

-> IO (Result ClusterData)

Top level function for data loading.