Ganeti.HTools.Backend.Simu

Description

Parsing data from a simulated description of the cluster.

This module holds the code for parsing a cluster description.

Documentation

parseData

Arguments

:: [String]
Cluster description in text format

-> Result ClusterData

Builds the cluster data from node/instance files.

loadData

Arguments

:: [String]
Cluster description in text format

-> IO (Result ClusterData)

Builds the cluster data from node/instance files.