Ganeti.HTools.Cluster.Metrics

Description

Implementation of the cluster metric

Documentation

reservedMemRtotalCoeff :: Double

Coefficient for the total reserved memory in the cluster metric. We use a (local) constant here, as it is also used in the computation of the best possible cluster score.

detailedCVInfoExt :: [((Double, String), ([AggregateComponent] -> Statistics, Bool))]

The names and weights of the individual elements in the CV list, together with their statistical accumulation function and a bit to decide whether it is a statistics for online nodes.

optimalCVScore :: List -> Double

Compute the lower bound of the cluster score, i.e., the sum of the minimal values for all cluster score values that are not 0 on a perfectly balanced cluster.

detailedCVInfo :: [(Double, String)]

The names and weights of the individual elements in the CV list.

detailedCVWeights :: [Double]

Holds the weights used by compCVNodes for each metric.

detailedCVAggregation :: [([AggregateComponent] -> Statistics, Bool)]

The aggregation functions for the weights

detailedCVOnlineStatus :: [Bool]

The bit vector describing which parts of the statistics are for online nodes.

compDetailedCVNode :: Node -> [AggregateComponent]

Compute statistical measures of a single node.

compClusterStatistics :: [Node] -> [Statistics]

Compute the statistics of a cluster.

updateClusterStatistics :: [Statistics] -> (Node, Node) -> [Statistics]

Update a cluster statistics by replacing the contribution of one node by that of another.

updateClusterStatisticsTwice :: [Statistics] -> (Node, Node) -> (Node, Node) -> [Statistics]

Update a cluster statistics twice.

compDetailedCV :: [Node] -> [Double]

Compute cluster statistics

compCVfromStats :: [Statistics] -> Double

Compute the cluster score from its statistics

compCVNodes :: [Node] -> Double

Compute the total variance.

compCV :: List -> Double

Wrapper over compCVNodes for callers that have a List.

printStats :: String -> List -> String

Shows statistics for a given node list.