Ganeti.HTools.ExtLoader

Description

External data loader.

This module holds the external data loading, and thus is the only one depending (via the specialized Text/Rapi/Luxi modules) on the actual libraries implementing the low-level protocols.

Documentation

loadExternalData :: Options -> IO ClusterData

External tool data loader from a variety of sources.

maybeSaveData

Arguments

:: Maybe FilePath
The file prefix to save to

-> String
The suffix (extension) to add

-> String
Informational message

-> ClusterData
The cluster data

-> IO ()

Function to save the cluster data to a file.