Ganeti.HTools.Group

Description

Module describing a node group.

Type declarations

data Group

The node group type.

Constructors

Group

Fields

name :: String

The node name

uuid :: GroupID

The UUID of the group

idx :: Gdx

Internal index for book-keeping

allocPolicy :: AllocPolicy

The allocation policy for this group

networks :: [NetworkID]

The networks connected to this group

iPolicy :: IPolicy

The instance policy for this group

allTags :: [String]

The tags for this group

type AssocList = [(Gdx, Group)]

A simple name for the int, node association list.

type List = Container Group

A simple name for a node map.

Initialization functions

create

Arguments

:: String
The node name

-> GroupID
The UUID of the group

-> AllocPolicy
The allocation policy for this group

-> [NetworkID]
The networks connected to this group

-> IPolicy
The instance policy for this group

-> [String]
The tags for this group

-> Group

Create a new group.

setIdx :: Group -> Gdx -> Group

Sets the group index.

This is used only during the building of the data structures.

isAllocable :: Group -> Bool

Checks if a group is allocable.

setUnallocable :: Group -> Group

Makes the group unallocatable