Ganeti.HTools.Loader

Description

Generic data loader.

This module holds the common code for parsing the input data after it has been loaded from external sources.

Types

data RqType

The iallocator request type.

This type denotes what request we got from Ganeti and also holds request-specific fields.

Constructors

Allocate Instance AllocDetails (Maybe [String])
A new instance allocation, maybe with allocation restrictions

AllocateSecondary Idx
Find a suitable secondary node for disk conversion

Relocate Idx Int [Ndx]
Choose a new secondary node

NodeEvacuate [Idx] EvacMode
node-evacuate mode

ChangeGroup [Gdx] [Idx]
Multi-relocate mode

MultiAllocate [(Instance, AllocDetails)]
Multi-allocate mode

data Request

A complete request, as received from Ganeti.

Constructors

Request RqType ClusterData

isAllocationRequest :: RqType -> Maybe (Maybe String)

Decide whether a request asks to allocate new instances; if so, also return the desired node group, if a unique node group is specified. That is, return Nothing if the request is not an allocation request, `Just Nothing`, if it is an Allocation request, but there is no unique group specified, and return `Just (Just g)` if it is an allocation request uniquely requesting Group g.

data ClusterData

The cluster state.

Constructors

ClusterData

Fields

cdGroups :: List

The node group list

cdNodes :: List

The node list

cdInstances :: List

The instance list

cdTags :: [String]

The cluster tags

cdIPolicy :: IPolicy

The cluster instance policy

emptyCluster :: ClusterData

An empty cluster.

Functions

lookupNode :: MonadFail m => NameAssoc -> String -> String -> m Ndx

Lookups a node into an assoc list.

lookupInstance :: MonadFail m => NameAssoc -> String -> m Idx

Lookups an instance into an assoc list.

lookupGroup :: MonadFail m => NameAssoc -> String -> String -> m Gdx

Lookups a group into an assoc list.

assignIndices :: Element a => [(String, a)] -> (NameAssoc, Container a)

Given a list of elements (and their names), assign indices to them.

setMaster :: MonadFail m => NameAssoc -> List -> String -> m List

Given am indexed node list, and the name of the master, mark it as such.

setLocationScore :: List -> Instance -> Instance

Given the nodes with the location tags already set correctly, compute the location score for an instance.

fixNodes :: List -> Instance -> List

For each instance, add its index to its primary and secondary nodes.

setNodePolicy :: List -> Node -> Node

Set the node's policy to its group one. Note that this requires the group to exist (should have been checked before), otherwise it will abort with a runtime error.

updateExclTags :: [String] -> Instance -> Instance

Update instance with exclusion tags list.

updateDesiredLocationTags :: [String] -> Instance -> Instance

Update instance with desired location tags list.

updateMovable

Arguments

:: [String]
Selected instances (if not empty)

-> [String]
Excluded instances

-> Instance
Target Instance

-> Instance
Target Instance with updated attribute

Update the movable attribute.

disableSplitMoves :: List -> Instance -> Instance

Disables moves for instances with a split group.

setArPolicy

Arguments

:: [String]
Cluster tags

-> List
List of node groups

-> List
List of nodes

-> List
List of instances

-> ClockTime
Current timestamp, to evaluate ArSuspended

-> List
Updated list of instances

Set the auto-repair policy for an instance.

getArPolicy :: [String] -> ClockTime -> Maybe AutoRepairPolicy

Get the auto-repair policy from a list of tags.

This examines the ganeti:watcher:autorepair and ganeti:watcher:autorepair:suspend tags to determine the policy. If none of these tags are present, Nothing (and not ArNotEnabled) is returned.

longestDomain :: [String] -> String

Compute the longest common suffix of a list of strings that starts with a dot.

extractExTags :: [String] -> [String]

Extracts the exclusion tags from the cluster configuration.

extractDesiredLocations :: [String] -> [String]

Extracts the desired locations from the instance tags.

commonSuffix :: List -> List -> String

Extracts the common suffix from node/instance names.

addMigrationTags

Arguments

:: [String]
cluster tags

-> Node

-> Node

Set the migration-related tags on a node given the cluster tags; this assumes that the node tags are already set on that node.

addLocationTags

Arguments

:: [String]
cluster tags

-> Node

-> Node

Set the location tags on a node given the cluster tags; this assumes that the node tags are already set on that node.

mergeData

Arguments

:: [(String, DynUtil)]
Instance utilisation data

-> [String]
Exclusion tags

-> [String]
Selected instances (if not empty)

-> [String]
Excluded instances

-> ClockTime
The current timestamp

-> ClusterData
Data from backends

-> Result ClusterData
Fixed cluster data

Initializer function that loads the data from a node and instance list and massages it into the correct format.

clearDynU :: ClusterData -> Result ClusterData

In a cluster description, clear dynamic utilisation information.

setStaticKvmNodeMem

Arguments

:: List
Nodes to update

-> Int
Static node size

-> List
Updated nodes

Update cluster data to use static node memory on KVM.

updateMemStat :: Node -> List -> Node

Update node memory stat based on instance list.

updateMissing

Arguments

:: List
All nodes in the cluster

-> List
All instances in the cluster

-> Int
Static node memory for KVM

-> ([String], List)
Pair of errors, update node list

Check the cluster for memory/disk allocation consistency and update stats.

nodeImem :: Node -> List -> Int

Compute the amount of memory used by primary instances on a node.

nodeIdsk :: Node -> List -> Int

Compute the amount of disk used by instances on a node (either primary or secondary).

eitherLive :: MonadFail m => Bool -> a -> m a -> m a

Get live information or a default value