Ganeti.HTools.Program.Hbal

Description

Cluster rebalancer.

Documentation

options :: IO [OptType]

Options list and functions.

arguments :: [ArgCompletion]

The list of arguments supported by the program.

annotateOpCode :: Maybe String -> Timestamp -> Annotator

Wraps an OpCode in a MetaOpCode while also adding a comment about what generated the opcode.

iterateDepth

Arguments

:: Bool
Whether to print moves

-> AlgorithmOptions
Algorithmic options to apply

-> Table
The starting table

-> Int
Remaining length

-> Int
Max node name len

-> Int
Max instance name len

-> [MoveJob]
Current command list

-> Score
Score at which to stop

-> IO (Table, [MoveJob])
The resulting table and commands

Start computing the solution at the given depth and recurse until we find a valid solution or we exceed the maximum depth.

printStats :: List -> List -> IO ()

Displays the cluster stats.

maybeExecJobs :: Options -> [a] -> List -> List -> [JobSet] -> IO (Result ())

Executes the jobs, if possible and desired.

selectGroup :: Options -> List -> List -> List -> IO (String, (List, List))

Select the target node group.

checkCluster :: Int -> List -> List -> IO ()

Do a few checks on the cluster data.

checkGroup :: Bool -> Int -> String -> List -> List -> IO ()

Do a few checks on the selected group data.

checkNeedRebalance :: Options -> Score -> Score -> IO ()

Check that we actually need to rebalance.

main :: Options -> [String] -> IO ()

Main function.