Ganeti.HTools.Program.Hspace

Description

Cluster space sizing

Documentation

options :: IO [OptType]

Options list and functions.

arguments :: [ArgCompletion]

The list of arguments supported by the program.

data Phase

The allocation phase we're in (initial, after tiered allocs, or after regular allocation).

Constructors

PInitial

PFinal

PTiered

data SpecType

The kind of instance spec we print.

Constructors

SpecNormal

SpecTiered

htsPrefix :: String

Prefix for machine readable names

specPrefix :: SpecType -> String

What we prefix a spec with.

specDescription :: SpecType -> String

The description of a spec.

specName :: SpecType -> String

The "name" of a SpecType.

effFn :: (CStats -> Integer) -> (CStats -> Double) -> CStats -> Double

Efficiency generic function.

memEff :: CStats -> Double

Memory efficiency.

dskEff :: CStats -> Double

Disk efficiency.

cpuEff :: CStats -> Double

Cpu efficiency.

spnEff :: CStats -> Double

Spindles efficiency.

statsData :: [(String, CStats -> String)]

Holds data for converting a CStats structure into detailed statistics.

specData :: [(String, RSpec -> String)]

List holding RSpec formatting information.

specDataSpn :: [(String, RSpec -> String)]

RSpec formatting information including spindles.

clusterData :: [(String, CStats -> String)]

List holding CStats formatting information.

clusterDataSpn :: [(String, CStats -> String)]

CStats formatting information including spindles

printStats :: Phase -> CStats -> [(String, String)]

Function to print stats for a given phase.

printFRScores :: List -> List -> [(FailMode, Int)] -> IO ()

Print failure reason and scores

printResults :: Bool -> List -> List -> Int -> Int -> [(FailMode, Int)] -> IO ()

Print final stats and related metrics.

printFinalHTS :: Bool -> IO ()

Prints the final OK marker in machine readable output.

tieredSpecMap :: [Instance] -> [(RSpec, Int)]

Compute the tiered spec counts from a list of allocated instances.

formatSpecMap :: [(RSpec, Int)] -> [String]

Formats a spec map to strings.

formatRSpec :: String -> AllocInfo -> [(String, String)]

Formats "key-metrics" values.

printAllocationStats :: List -> List -> IO ()

Shows allocations stats.

printKeysHTS :: [(String, String)] -> IO ()

Format a list of key/values as a shell fragment.

printInstance :: List -> Instance -> [String]

Converts instance data to a list of strings.

printAllocationMap :: Int -> String -> List -> [Instance] -> IO ()

Optionally print the allocation map.

formatResources :: a -> [(String, a -> String)] -> String

Formats nicely a list of resources.

printCluster :: Bool -> CStats -> Int -> Bool -> IO ()

Print the cluster resources.

printISpec :: Bool -> RSpec -> SpecType -> DiskTemplate -> Bool -> IO ()

Prints the normal instance spec.

printTiered :: Bool -> [(RSpec, Int)] -> List -> List -> [(FailMode, Int)] -> IO ()

Prints the tiered results.

printClusterScores :: List -> List -> IO ()

Displays the initial/final cluster scores.

printClusterEff :: CStats -> Bool -> IO ()

Displays the cluster efficiency.

failureReason :: [(FailMode, Int)] -> String

Computes the most likely failure reason.

sortReasons :: [(FailMode, Int)] -> [(FailMode, Int)]

Sorts the failure reasons.

runAllocation

Arguments

:: ClusterData
Cluster data

-> Maybe AllocResult
Optional stop-allocation

-> Result AllocResult
Allocation result

-> RSpec
Requested instance spec

-> DiskTemplate
Requested disk template

-> SpecType
Allocation type

-> Options
CLI options

-> IO (FailStats, List, Int, [(RSpec, Int)])

Runs an allocation algorithm and saves cluster state.

instFromSpec :: RSpec -> DiskTemplate -> Int -> Instance

Create an instance from a given spec. For values not implied by the resorce specification (like distribution of of the disk space to individual disks), sensible defaults are guessed (e.g., having a single disk).

combineTiered :: AlgorithmOptions -> Maybe Int -> AllocNodes -> AllocResult -> Instance -> Result AllocResult

main :: Options -> [String] -> IO ()

Main function.