Ganeti.HTools.Types

Description

Some common types.

Documentation

type Idx = Int

The instance index type.

type Ndx = Int

The node index type.

type Gdx = Int

The group index type.

type NameAssoc = Map String Int

The type used to hold name-to-idx mappings.

type Score = Double

A separate name for the cluster score type.

type Weight = Double

A separate name for a weight metric.

type GroupID = String

The Group UUID type.

defaultGroupID :: GroupID

Default group UUID (just a string, not a real UUID).

data MirrorType

Mirroring type.

Constructors

MirrorNone
No mirroring/movability

MirrorInternal
DRBD-type mirroring

MirrorExternal
Shared-storage type mirroring

templateMirrorType :: DiskTemplate -> MirrorType

Correspondence between disk template and mirror type.

data RSpec

The resource spec type.

Constructors

RSpec

Fields

rspecCpu :: Int

Requested VCPUs

rspecMem :: Int

Requested memory

rspecDsk :: Int

Requested disk

rspecSpn :: Int

Requested spindles

data AllocInfo

Allocation stats type. This is used instead of RSpec (which was used at first), because we need to track more stats. The actual data can refer either to allocated, or available, etc. values depending on the context. See also computeAllocationDelta.

Constructors

AllocInfo

Fields

allocInfoVCpus :: Int

VCPUs

allocInfoNCpus :: Double

Normalised CPUs

allocInfoMem :: Int

Memory

allocInfoDisk :: Int

Disk

allocInfoSpn :: Int

Spindles

type AllocStats = (AllocInfo, AllocInfo, AllocInfo)

Currently used, possibly to allocate, unallocable.

type NetworkID = String

The network UUID type.

data ISpec

Instance specification type.

Constructors

ISpec

Fields

iSpecMemorySize :: Int

iSpecCpuCount :: Int

iSpecDiskSize :: Int

iSpecDiskCount :: Int

iSpecNicCount :: Int

iSpecSpindleUse :: Int

defMinISpec :: ISpec

The default minimum ispec.

defStdISpec :: ISpec

The default standard ispec.

maxDisks :: Int

maxNics :: Int

defMaxISpec :: ISpec

The default max ispec.

data MinMaxISpecs

Minimum and maximum instance specs type.

Constructors

MinMaxISpecs

Fields

minMaxISpecsMinSpec :: ISpec

minMaxISpecsMaxSpec :: ISpec

data IPolicy

Instance policy type.

Constructors

IPolicy

Fields

iPolicyMinMaxISpecs :: [MinMaxISpecs]

iPolicyStdSpec :: ISpec

iPolicyDiskTemplates :: [DiskTemplate]

iPolicyVcpuRatio :: Double

iPolicySpindleRatio :: Double

rspecFromISpec :: ISpec -> RSpec

Converts an ISpec type to a RSpec one.

defIPolicy :: IPolicy

The default instance policy.

data DynUtil

The dynamic resource specs of a machine (i.e. load or load capacity, as opposed to size).

Constructors

DynUtil

Fields

cpuWeight :: Weight

Standardised CPU usage

memWeight :: Weight

Standardised memory load

dskWeight :: Weight

Standardised disk I/O usage

netWeight :: Weight

Standardised network usage

zeroUtil :: DynUtil

Initial empty utilisation.

baseUtil :: DynUtil

Base utilisation (used when no actual utilisation data is supplied).

addUtil :: DynUtil -> DynUtil -> DynUtil

Sum two utilisation records.

subUtil :: DynUtil -> DynUtil -> DynUtil

Substracts one utilisation record from another.

type Placement = (Idx, Ndx, Ndx, IMove, Score)

The description of an instance placement. It contains the instance index, the new primary and secondary node, the move being performed and the score of the cluster after the move.

data IMove

An instance move definition.

Constructors

Failover
Failover the instance (f)

FailoverToAny Ndx
Failover to a random node (fa:np), for shared storage

ReplacePrimary Ndx
Replace primary (f, r:np, f)

ReplaceSecondary Ndx
Replace secondary (r:ns)

ReplaceAndFailover Ndx
Replace secondary, failover (r:np, f)

FailoverAndReplace Ndx
Failover, replace secondary (f, r:ns)

type MoveJob = ([Ndx], Idx, IMove, [String])

Formatted solution output for one move (involved nodes and commands.

unknownField :: String

Unknown field in table output.

type JobSet = [MoveJob]

A list of command elements.

defReservedDiskRatio :: Double

Default max disk usage ratio.

unitMem :: Int

Base memory unit.

unitDsk :: Int

Base disk unit.

unitCpu :: Int

Base vcpus unit.

unitSpindle :: Int

Base spindles unit.

data FailMode

Reason for an operation's falure.

Constructors

FailMem
Failed due to not enough RAM

FailDisk
Failed due to not enough disk

FailCPU
Failed due to not enough CPU capacity

FailN1
Failed due to not passing N1 checks

FailTags
Failed due to tag exclusion

FailMig
Failed due to migration restrictions

FailDiskCount
Failed due to wrong number of disks

FailSpindles
Failed due to wrong/missing spindles

FailInternal
Internal error

type FailStats = [(FailMode, Int)]

List with failure statistics.

type OpResult = GenericResult FailMode

Either-like data-type customized for our failure modes.

The failure values for this monad track the specific allocation failures, so this is not a general error-monad (compare with the Result data type). One downside is that this type cannot encode a generic failure mode, hence our way to build a FailMode from string will instead raise an exception.

opToResult :: OpResult a -> Result a

Conversion from OpResult to Result.

class Element a where

A generic class for items that have updateable names and indices.

Minimal complete definition

nameOf, allNames, idxOf, setAlias, setIdx

Methods

nameOf :: a -> String

Returns the name of the element

allNames :: a -> [String]

Returns all the known names of the element

idxOf :: a -> Int

Returns the index of the element

setAlias :: a -> String -> a

Updates the alias of the element

computeAlias :: String -> a -> a

Compute the alias by stripping a given suffix (domain) from the name

setIdx :: a -> Int -> a

Updates the index of the element

data AutoRepairType

The repair modes for the auto-repair tool.

Constructors

ArFixStorage

ArMigrate

ArFailover

ArReinstall

autoRepairTypeFromRaw :: forall m. Monad m => String -> m AutoRepairType

autoRepairTypeToRaw :: AutoRepairType -> String

data AutoRepairResult

The possible auto-repair results.

Constructors

ArEnoperm

ArSuccess

ArFailure

autoRepairResultFromRaw :: forall m. Monad m => String -> m AutoRepairResult

autoRepairResultToRaw :: AutoRepairResult -> String

data AutoRepairPolicy

The possible auto-repair policy for a given instance.

Constructors

ArEnabled AutoRepairType
Auto-repair explicitly enabled

ArSuspended AutoRepairSuspendTime
Suspended temporarily, or forever

ArNotEnabled
Auto-repair not explicitly enabled

data AutoRepairSuspendTime

The suspend timeout for ArSuspended.

Constructors

Forever
Permanently suspended

Until ClockTime
Suspended up to a certain time

data AutoRepairStatus

The possible auto-repair states for any given instance.

Constructors

ArHealthy (Maybe AutoRepairData)
No problems detected with the instance

ArNeedsRepair AutoRepairData
Instance has problems, no action taken

ArPendingRepair AutoRepairData
Repair jobs ongoing for the instance

ArFailedRepair AutoRepairData
Some repair jobs for the instance failed

data AutoRepairData

The data accompanying a repair operation (future, pending, or failed).

Constructors

AutoRepairData

Fields

arType :: AutoRepairType

arUuid :: String

arTime :: ClockTime

arJobs :: [JobId]

arResult :: Maybe AutoRepairResult

arTag :: String