Ganeti.Hs2Py.GenOpCodes

Description

GenOpCodes handles Python opcode generation.

GenOpCodes contains the helper functions that generate the Python opcodes as strings from the Haskell opcode description.

Documentation

pyClassDoc :: String -> String

Generates the Python class docstring.

pyClassField :: OpCodeField -> String

Generates an opcode parameter in Python.

intercalateIndent :: [String] -> String

Comma intercalates and indents opcode parameters in Python.

showPyClass :: OpCodeDescriptor -> String

Generates an opcode as a Python class.

showPyClasses :: String

Generates all opcodes as Python classes.