Ganeti.Hypervisor.Xen

Description

Module to access the information provided by the Xen hypervisor.

Documentation

getDomainsInfo :: IO (Result (Map String Domain))

Get information about the current Xen domains as a map where the domain name is the key. This only includes the information made available by Xen itself.

inferDomInfos :: Map String Domain -> Domain -> Domain

Given a domain and a map containing information about multiple domains, infer additional information about that domain (specifically, whether it is hung).

getInferredDomInfo :: IO (Result (Map String Domain))

Get information about the current Xen domains as a map where the domain name is the key. This includes information made available by Xen itself as well as further information that can be inferred by querying Xen multiple times and comparing the results.

getUptimeInfo :: IO (Map Int UptimeInfo)

Get information about the uptime of domains, as a map where the domain ID is the key.