Ganeti.Hypervisor.Xen.Types

Description

Data types for Xen-specific hypervisor functionalities.

Documentation

data LispConfig

Data type representing configuration data as produced by the xm list --long command.

Constructors

LCList [LispConfig]

LCString String

LCDouble Double

data Domain

Data type representing a Xen Domain.

Constructors

Domain

Fields

domId :: Int

domName :: String

domCpuTime :: Double

domState :: ActualState

domIsHung :: Maybe Bool

class FromLispConfig a where

Class representing all the types that can be extracted from LispConfig.

Methods

fromLispConfig :: LispConfig -> Result a

data UptimeInfo

Constructors

UptimeInfo

Fields

uInfoName :: String

uInfoID :: Int

uInfoUptime :: String

data ActualState

Constructors

ActualRunning
The instance is running

ActualBlocked
The instance is not running or runnable

ActualPaused
The instance has been paused

ActualShutdown
The instance is shut down

ActualCrashed
The instance has crashed

ActualDying
The instance is in process of dying

ActualHung
The instance is hung

ActualUnknown
Unknown state. Parsing error.