Ganeti.Locking.Locks

Description

Ganeti lock structure

Documentation

data GanetiLocks

The type of Locks available in Ganeti. The order of this type is the lock oder.

Constructors

ClusterLockSet

BGL

InstanceLockSet

Instance String

NodeGroupLockSet

NodeGroup String

NodeLockSet

Node String

NodeResLockSet

NodeRes String

NetworkLockSet

Network String

ConfigLock
A lock used for a transitional period when WConfd keeps the state of the configuration, but all the operations are still performed on the Python side.

lockName :: GanetiLocks -> String

Provide the String representation of a lock

data LockLevel

The levels, the locks belong to.

Constructors

LevelCluster

LevelInstance

LevelNodeGroup

LevelNode

LevelNodeRes

LevelNetwork

LevelConfig
A transitional level for internal configuration locks

lockLevel :: GanetiLocks -> LockLevel

For a lock, provide its level.

data ClientType

Type of entities capable of owning locks. Usually, locks are owned by jobs. However, occassionally other tasks need locks (currently, e.g., to lock the configuration). These are identified by a unique name, reported to WConfD as a strig.

Constructors

ClientOther String

ClientJob JobId

data ClientId

A client is identified as a job id, thread id, a path to its process identifier file, and its process id.

The JobId isn't enough to identify a client as the master daemon also handles client calls that aren't jobs, but which use the configuration. These taks are identified by a unique name, reported to WConfD as a string.

Constructors

ClientId

Fields

ciIdentifier :: ClientType

ciLockFile :: FilePath

ciPid :: ProcessID

type GanetiLockWaiting = LockWaiting GanetiLocks ClientId Integer

The type of lock Allocations in Ganeti. In Ganeti, the owner of locks are jobs.