Ganeti.Logging.WriterLog

Description

A pure implementation of MonadLog using MonadWriter

The data type of the monad transformer

type LogSeq = Seq (Priority, String)

type WriterSeq = WriterT LogSeq

newtype WriterLogT m a

A monad transformer that adds pure logging capability.

Constructors

WriterLogT

Fields

unwrapWriterLogT :: WriterSeq m a

type WriterLog = WriterLogT Identity

runWriterLogT :: WriterLogT m a -> m (a, LogSeq)

runWriterLog :: WriterLog a -> (a, LogSeq)

execWriterLogT :: MonadLog m => WriterLogT m a -> m a

Runs a WriterLogT, and when it finishes, resends all log messages to the underlying monad that implements MonadLog.

This can be used to delay logging messages, by accumulating them in WriterLogT, and resending them at the end to the underlying monad.

dumpLogSeq :: MonadLog m => LogSeq -> m ()

Sends all log messages to the a monad that implements MonadLog.

execWriterLog :: MonadLog m => WriterLog a -> m a

Runs a WriterLog, and when it finishes, resends all log messages to the a monad that implements MonadLog.