Ganeti.Luxi

Description

Implementation of the Ganeti LUXI interface.

Documentation

data LuxiOp

Currently supported Luxi operations and JSON serialization.

Constructors

Query ItemType [String] (Filter FilterField)

QueryFields ItemType [String]

QueryNodes [String] [String] Bool

QueryGroups [String] [String] Bool

QueryNetworks [String] [String] Bool

QueryInstances [String] [String] Bool

QueryFilters [String] [String]

ReplaceFilter (Maybe String) (NonNegative Int) [FilterPredicate] FilterAction ReasonTrail

DeleteFilter String

QueryJobs [JobId] [String]

QueryExports [String] Bool

QueryConfigValues [String]

QueryClusterInfo

QueryTags TagKind String

SubmitJob [MetaOpCode]

SubmitJobToDrainedQueue [MetaOpCode]

SubmitManyJobs [[MetaOpCode]]

WaitForJobChange JobId [String] JSValue JSValue Int

PickupJob JobId

ArchiveJob JobId

AutoArchiveJobs Int Int

CancelJob JobId Bool

ChangeJobPriority JobId Int

SetDrainFlag Bool

SetWatcherPause (Maybe ClockTime)

data LuxiReq

Constructors

ReqQuery

ReqQueryFields

ReqQueryNodes

ReqQueryGroups

ReqQueryNetworks

ReqQueryInstances

ReqQueryFilters

ReqReplaceFilter

ReqDeleteFilter

ReqQueryJobs

ReqQueryExports

ReqQueryConfigValues

ReqQueryClusterInfo

ReqQueryTags

ReqSubmitJob

ReqSubmitJobToDrainedQueue

ReqSubmitManyJobs

ReqWaitForJobChange

ReqPickupJob

ReqArchiveJob

ReqAutoArchiveJobs

ReqCancelJob

ReqChangeJobPriority

ReqSetDrainFlag

ReqSetWatcherPause

opToArgs :: LuxiOp -> JSValue

allLuxiCalls :: [String]

List of all defined Luxi calls.

strOfOp :: LuxiOp -> String

The serialisation of LuxiOps into strings in messages.

getLuxiClient :: String -> IO Client

Connects to the master daemon and returns a luxi Client.

getLuxiServer :: Bool -> FilePath -> IO Server

Creates and returns a server endpoint.

decodeLuxiCall :: JSValue -> JSValue -> Result LuxiOp

Converts Luxi call arguments into a LuxiOp data structure. This is used for building a Luxi Handler.

This is currently hand-coded until we make it more uniform so that it can be generated using TH.

callMethod :: LuxiOp -> Client -> IO (ErrorResult JSValue)

Generic luxi method call

submitManyJobs :: Client -> [[MetaOpCode]] -> IO (ErrorResult [JobId])

Specialized submitManyJobs call.

queryJobsStatus :: Client -> [JobId] -> IO (ErrorResult [JobStatus])

Custom queryJobs call.