Ganeti.OpCodes

Description

Implementation of the opcodes.

Documentation

data OpCode

OpCode representation.

We only implement a subset of Ganeti opcodes: those which are actually used in the htools codebase.

Constructors

OpClusterPostInit

OpClusterDestroy

OpClusterQuery

OpClusterVerify

Fields

opDebugSimulateErrors :: Bool

opErrorCodes :: Bool

opSkipChecks :: (ListSet VerifyOptionalChecks)

opIgnoreErrors :: (ListSet CVErrorCode)

opVerbose :: Bool

opOptGroupName :: (Maybe NonEmptyString)

opVerifyClutter :: Bool

OpClusterVerifyConfig

Fields

opDebugSimulateErrors :: Bool

opErrorCodes :: Bool

opIgnoreErrors :: (ListSet CVErrorCode)

opVerbose :: Bool

OpClusterVerifyGroup

Fields

opGroupName :: NonEmptyString

opDebugSimulateErrors :: Bool

opErrorCodes :: Bool

opSkipChecks :: (ListSet VerifyOptionalChecks)

opIgnoreErrors :: (ListSet CVErrorCode)

opVerbose :: Bool

opVerifyClutter :: Bool

OpClusterVerifyDisks

Fields

opOptGroupName :: (Maybe NonEmptyString)

OpGroupVerifyDisks

Fields

opGroupName :: NonEmptyString

OpClusterRepairDiskSizes

Fields

opInstances :: [NonEmptyString]

OpClusterConfigQuery

Fields

opOutputFields :: [NonEmptyString]

OpClusterRename

Fields

opName :: NonEmptyString

OpClusterSetParams

Fields

opForce :: Bool

opHvState :: (Maybe (JSObject JSValue))

opDiskState :: (Maybe (JSObject JSValue))

opVgName :: (Maybe String)

opEnabledHypervisors :: (Maybe [Hypervisor])

opClusterHvParams :: (Maybe (GenericContainer String (JSObject JSValue)))

opClusterBeParams :: (Maybe (JSObject JSValue))

opOsHvp :: (Maybe (GenericContainer String (JSObject JSValue)))

opClusterOsParams :: (Maybe (GenericContainer String (JSObject JSValue)))

opClusterOsParamsPrivate :: (Maybe (GenericContainer String (JSObject (Private JSValue))))

opDiskparams :: (Maybe (GenericContainer DiskTemplate (JSObject JSValue)))

opCandidatePoolSize :: (Maybe (Positive Int))

opMaxRunningJobs :: (Maybe (Positive Int))

opMaxTrackedJobs :: (Maybe (Positive Int))

opUidPool :: (Maybe [(Int, Int)])

opAddUids :: (Maybe [(Int, Int)])

opRemoveUids :: (Maybe [(Int, Int)])

opMaintainNodeHealth :: (Maybe Bool)

opPreallocWipeDisks :: (Maybe Bool)

opNicparams :: (Maybe INicParams)

opgenericNdParams :: (Maybe (JSObject JSValue))

opIpolicy :: (Maybe (JSObject JSValue))

opDrbdHelper :: (Maybe String)

opDefaultIallocator :: (Maybe String)

opDefaultIallocatorParams :: (Maybe (JSObject JSValue))

opNetMacPrefix :: (Maybe NonEmptyString)

opMasterNetdev :: (Maybe String)

opMasterNetmask :: (Maybe (NonNegative Int))

opReservedLvs :: (Maybe [NonEmptyString])

opHiddenOs :: (Maybe [(DdmSimple, NonEmptyString)])

opBlacklistedOs :: (Maybe [(DdmSimple, NonEmptyString)])

opUseExternalMipScript :: (Maybe Bool)

opEnabledDiskTemplates :: (Maybe [DiskTemplate])

opModifyEtcHosts :: (Maybe Bool)

opClusterFileStorageDir :: (Maybe String)

opClusterSharedFileStorageDir :: (Maybe String)

opClusterGlusterStorageDir :: (Maybe String)

opInstallImage :: (Maybe String)

opInstanceCommunicationNetwork :: (Maybe String)

opZeroingImage :: (Maybe String)

opCompressionTools :: (Maybe [NonEmptyString])

opEnabledUserShutdown :: (Maybe Bool)

opEnabledDataCollectors :: (Maybe (GenericContainer String Bool))

opDataCollectorInterval :: (Maybe (GenericContainer String Int))

OpClusterRedistConf

OpClusterActivateMasterIp

OpClusterDeactivateMasterIp

OpClusterRenewCrypto

Fields

opNodeCertificates :: Bool

opRenewSshKeys :: Bool

opSshKeyType :: (Maybe SshKeyType)

opSshKeyBits :: (Maybe (Positive Int))

opVerbose :: Bool

opDebug :: Bool

OpQuery

Fields

opWhat :: QueryTypeOp

opUseLocking :: Bool

opFields :: [NonEmptyString]

opQfilter :: (Maybe [JSValue])

OpQueryFields

Fields

opWhat :: QueryTypeOp

opQueryFieldsFields :: (Maybe [NonEmptyString])

OpOobCommand

Fields

opNodeNames :: [NonEmptyString]

opNodeUuids :: (Maybe [NonEmptyString])

opOobCommand :: OobCommand

opOobTimeout :: Int

opIgnoreStatus :: Bool

opPowerDelay :: Double

OpRestrictedCommand

Fields

opUseLocking :: Bool

opReqNodes :: [NonEmptyString]

opReqNodeUuids :: (Maybe [NonEmptyString])

opRestrictedCommand :: NonEmptyString

OpNodeRemove

Fields

opNodeName :: NonEmptyString

opNodeUuid :: (Maybe NonEmptyString)

OpNodeAdd

Fields

opNodeName :: NonEmptyString

opHvState :: (Maybe (JSObject JSValue))

opDiskState :: (Maybe (JSObject JSValue))

opPrimaryIp :: (Maybe NonEmptyString)

opSecondaryIp :: (Maybe NonEmptyString)

opReadd :: Bool

opGroup :: (Maybe NonEmptyString)

opMasterCapable :: (Maybe Bool)

opVmCapable :: (Maybe Bool)

opgenericNdParams :: (Maybe (JSObject JSValue))

opNodeSetup :: Bool

OpNodeQueryvols

Fields

opOutputFields :: [NonEmptyString]

opNodes :: [NonEmptyString]

OpNodeQueryStorage

Fields

opOutputFields :: [NonEmptyString]

opOptStorageType :: (Maybe StorageType)

opNodes :: [NonEmptyString]

opStorageName :: (Maybe NonEmptyString)

OpNodeModifyStorage

Fields

opNodeName :: NonEmptyString

opNodeUuid :: (Maybe NonEmptyString)

opStorageType :: StorageType

opStorageName :: (Maybe NonEmptyString)

opChanges :: (JSObject JSValue)

OpRepairNodeStorage

Fields

opNodeName :: NonEmptyString

opNodeUuid :: (Maybe NonEmptyString)

opStorageType :: StorageType

opStorageName :: (Maybe NonEmptyString)

opIgnoreConsistency :: Bool

OpNodeSetParams

Fields

opNodeName :: NonEmptyString

opNodeUuid :: (Maybe NonEmptyString)

opForce :: Bool

opHvState :: (Maybe (JSObject JSValue))

opDiskState :: (Maybe (JSObject JSValue))

opMasterCandidate :: (Maybe Bool)

opOffline :: (Maybe Bool)

opDrained :: (Maybe Bool)

opAutoPromote :: Bool

opMasterCapable :: (Maybe Bool)

opVmCapable :: (Maybe Bool)

opSecondaryIp :: (Maybe NonEmptyString)

opgenericNdParams :: (Maybe (JSObject JSValue))

opPowered :: (Maybe Bool)

OpNodePowercycle

Fields

opNodeName :: NonEmptyString

opNodeUuid :: (Maybe NonEmptyString)

opForce :: Bool

OpNodeMigrate

Fields

opNodeName :: NonEmptyString

opNodeUuid :: (Maybe NonEmptyString)

opMigrationMode :: (Maybe MigrationMode)

opOldLiveMode :: (Maybe Bool)

opTargetNode :: (Maybe NonEmptyString)

opTargetNodeUuid :: (Maybe NonEmptyString)

opAllowRuntimeChanges :: Bool

opIgnoreIpolicy :: Bool

opIallocator :: (Maybe NonEmptyString)

OpNodeEvacuate

Fields

opEarlyRelease :: Bool

opNodeName :: NonEmptyString

opNodeUuid :: (Maybe NonEmptyString)

opRemoteNode :: (Maybe NonEmptyString)

opRemoteNodeUuid :: (Maybe NonEmptyString)

opIallocator :: (Maybe NonEmptyString)

opEvacMode :: EvacMode

opIgnoreSoftErrors :: (Maybe Bool)

OpInstanceCreate

Fields

opInstanceName :: String

opForceVariant :: Bool

opWaitForSync :: Bool

opNameCheck :: Bool

opIgnoreIpolicy :: Bool

opOpportunisticLocking :: Bool

opInstBeParams :: (JSObject JSValue)

opinstDisks :: [IDiskParams]

opOptDiskTemplate :: (Maybe DiskTemplate)

opOptGroupName :: (Maybe NonEmptyString)

opFileDriver :: (Maybe FileDriver)

opFileStorageDir :: (Maybe NonEmptyString)

opInstHvParams :: (JSObject JSValue)

opHypervisor :: (Maybe Hypervisor)

opIallocator :: (Maybe NonEmptyString)

opIdentifyDefaults :: Bool

opIpCheck :: Bool

opConflictsCheck :: Bool

opInstCreateMode :: InstCreateMode

opNics :: [INicParams]

opNoInstall :: (Maybe Bool)

opInstOsParams :: (JSObject JSValue)

opOsparamsPrivate :: (Maybe (JSObject (Private JSValue)))

opOsparamsSecret :: (Maybe (JSObject (Secret JSValue)))

opOsType :: (Maybe NonEmptyString)

opPnode :: (Maybe NonEmptyString)

opPnodeUuid :: (Maybe NonEmptyString)

opSnode :: (Maybe NonEmptyString)

opSnodeUuid :: (Maybe NonEmptyString)

opSourceHandshake :: (Maybe [JSValue])

opSourceInstanceName :: (Maybe NonEmptyString)

opSourceShutdownTimeout :: (NonNegative Int)

opSourceX509Ca :: (Maybe NonEmptyString)

opSrcNode :: (Maybe NonEmptyString)

opSrcNodeUuid :: (Maybe NonEmptyString)

opSrcPath :: (Maybe NonEmptyString)

opCompress :: String

opStart :: Bool

opForthcoming :: Bool

opCommit :: Bool

opInstTags :: [NonEmptyString]

opInstanceCommunication :: Bool

opHelperStartupTimeout :: (Maybe Int)

opHelperShutdownTimeout :: (Maybe Int)

OpInstanceMultiAlloc

Fields

opOpportunisticLocking :: Bool

opIallocator :: (Maybe NonEmptyString)

opInstMultiAlloc :: [JSValue]

OpInstanceReinstall

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opForceVariant :: Bool

opOsType :: (Maybe NonEmptyString)

opTempOsParams :: (Maybe (JSObject JSValue))

opOsparamsPrivate :: (Maybe (JSObject (Private JSValue)))

opOsparamsSecret :: (Maybe (JSObject (Secret JSValue)))

OpInstanceRemove

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opShutdownTimeout :: (NonNegative Int)

opIgnoreFailures :: Bool

OpInstanceRename

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opNewName :: NonEmptyString

opNameCheck :: Bool

opIpCheck :: Bool

OpInstanceStartup

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opForce :: Bool

opIgnoreOfflineNodes :: Bool

opTempHvParams :: (JSObject JSValue)

opTempBeParams :: (JSObject JSValue)

opNoRemember :: Bool

opStartupPaused :: Bool

opShutdownTimeout :: (NonNegative Int)

OpInstanceShutdown

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opForce :: Bool

opIgnoreOfflineNodes :: Bool

opInstShutdownTimeout :: (NonNegative Int)

opNoRemember :: Bool

opAdminStateSource :: (Maybe AdminStateSource)

OpInstanceReboot

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opShutdownTimeout :: (NonNegative Int)

opIgnoreSecondaries :: Bool

opRebootType :: RebootType

OpInstanceReplaceDisks

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opEarlyRelease :: Bool

opIgnoreIpolicy :: Bool

opReplaceDisksMode :: ReplaceDisksMode

opReplaceDisksList :: [DiskIndex]

opRemoteNode :: (Maybe NonEmptyString)

opRemoteNodeUuid :: (Maybe NonEmptyString)

opIallocator :: (Maybe NonEmptyString)

OpInstanceFailover

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opShutdownTimeout :: (NonNegative Int)

opIgnoreConsistency :: Bool

opTargetNode :: (Maybe NonEmptyString)

opTargetNodeUuid :: (Maybe NonEmptyString)

opIgnoreIpolicy :: Bool

opMigrationCleanup :: Bool

opIallocator :: (Maybe NonEmptyString)

OpInstanceMigrate

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opMigrationMode :: (Maybe MigrationMode)

opOldLiveMode :: (Maybe Bool)

opTargetNode :: (Maybe NonEmptyString)

opTargetNodeUuid :: (Maybe NonEmptyString)

opAllowRuntimeChanges :: Bool

opIgnoreIpolicy :: Bool

opMigrationCleanup :: Bool

opIallocator :: (Maybe NonEmptyString)

opAllowFailover :: Bool

opIgnoreHvversions :: Bool

OpInstanceMove

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opShutdownTimeout :: (NonNegative Int)

opIgnoreIpolicy :: Bool

opMoveTargetNode :: NonEmptyString

opMoveTargetNodeUuid :: (Maybe NonEmptyString)

opCompress :: String

opIgnoreConsistency :: Bool

OpInstanceConsole

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

OpInstanceActivateDisks

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opIgnoreSize :: Bool

opWaitForSync :: Bool

OpInstanceDeactivateDisks

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opForce :: Bool

OpInstanceRecreateDisks

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opRecreateDisksInfo :: RecreateDisksInfo

opNodes :: [NonEmptyString]

opNodeUuids :: (Maybe [NonEmptyString])

opIallocator :: (Maybe NonEmptyString)

OpInstanceQueryData

Fields

opUseLocking :: Bool

opInstances :: [NonEmptyString]

opStatic :: Bool

OpInstanceSetParams

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opForce :: Bool

opForceVariant :: Bool

opIgnoreIpolicy :: Bool

opInstNicChanges :: (SetParamsMods INicParams)

opInstDiskChanges :: (SetParamsMods IDiskParams)

opInstBeParams :: (JSObject JSValue)

opRuntimeMem :: (Maybe (Positive Int))

opInstHvParams :: (JSObject JSValue)

opOptDiskTemplate :: (Maybe DiskTemplate)

opInstExtParams :: (JSObject JSValue)

opFileDriver :: (Maybe FileDriver)

opFileStorageDir :: (Maybe NonEmptyString)

opPnode :: (Maybe NonEmptyString)

opPnodeUuid :: (Maybe NonEmptyString)

opRemoteNode :: (Maybe NonEmptyString)

opRemoteNodeUuid :: (Maybe NonEmptyString)

opIallocator :: (Maybe NonEmptyString)

opOsName :: (Maybe NonEmptyString)

opInstOsParams :: (JSObject JSValue)

opOsparamsPrivate :: (Maybe (JSObject (Private JSValue)))

opWaitForSync :: Bool

opOffline :: (Maybe Bool)

opConflictsCheck :: Bool

opHotplug :: Bool

opHotplugIfPossible :: Bool

opOptInstanceCommunication :: (Maybe Bool)

OpInstanceGrowDisk

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opWaitForSync :: Bool

opDiskIndex :: DiskIndex

opDiskChgAmount :: (NonNegative Int)

opDiskChkAbsolute :: Bool

opIgnoreIpolicy :: Bool

OpInstanceChangeGroup

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opEarlyRelease :: Bool

opIallocator :: (Maybe NonEmptyString)

opTargetGroups :: (Maybe [NonEmptyString])

OpGroupAdd

Fields

opGroupName :: NonEmptyString

opAllocPolicy :: (Maybe AllocPolicy)

opNdparams :: (Maybe (JSObject JSValue))

opDiskparams :: (Maybe (GenericContainer DiskTemplate (JSObject JSValue)))

opHvState :: (Maybe (JSObject JSValue))

opDiskState :: (Maybe (JSObject JSValue))

opIpolicy :: (Maybe (JSObject JSValue))

OpGroupAssignNodes

Fields

opGroupName :: NonEmptyString

opForce :: Bool

opReqNodes :: [NonEmptyString]

opReqNodeUuids :: (Maybe [NonEmptyString])

OpGroupSetParams

Fields

opGroupName :: NonEmptyString

opAllocPolicy :: (Maybe AllocPolicy)

opNdparams :: (Maybe (JSObject JSValue))

opDiskparams :: (Maybe (GenericContainer DiskTemplate (JSObject JSValue)))

opHvState :: (Maybe (JSObject JSValue))

opDiskState :: (Maybe (JSObject JSValue))

opIpolicy :: (Maybe (JSObject JSValue))

OpGroupRemove

Fields

opGroupName :: NonEmptyString

OpGroupRename

Fields

opGroupName :: NonEmptyString

opNewName :: NonEmptyString

OpGroupEvacuate

Fields

opGroupName :: NonEmptyString

opEarlyRelease :: Bool

opIallocator :: (Maybe NonEmptyString)

opTargetGroups :: (Maybe [NonEmptyString])

opSequential :: Bool

opForceFailover :: Bool

OpOsDiagnose

Fields

opOutputFields :: [NonEmptyString]

opNames :: [NonEmptyString]

OpExtStorageDiagnose

Fields

opOutputFields :: [NonEmptyString]

opNames :: [NonEmptyString]

OpBackupPrepare

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opExportMode :: ExportMode

OpBackupExport

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

opCompress :: String

opShutdownTimeout :: (NonNegative Int)

opExportTarget :: ExportTarget

opExportTargetNodeUuid :: (Maybe NonEmptyString)

opShutdown :: Bool

opRemoveInstance :: Bool

opIgnoreRemoveFailures :: Bool

opExportMode :: ExportMode

opX509KeyName :: (Maybe [JSValue])

opDestinationX509Ca :: (Maybe NonEmptyString)

opZeroFreeSpace :: Bool

opZeroingTimeoutFixed :: (Maybe Int)

opZeroingTimeoutPerMib :: (Maybe Double)

opLongSleep :: Bool

OpBackupRemove

Fields

opInstanceName :: String

opInstanceUuid :: (Maybe NonEmptyString)

OpTagsGet

Fields

opKind :: TagKind

opUseLocking :: Bool

opTagsGetName :: (Maybe String)

OpTagsSearch

Fields

opTagSearchPattern :: NonEmptyString

OpTagsSet

Fields

opKind :: TagKind

opTagsList :: [String]

opTagsGetName :: (Maybe String)

OpTagsDel

Fields

opKind :: TagKind

opTagsList :: [String]

opTagsGetName :: (Maybe String)

OpTestDelay

Fields

opDelayDuration :: Double

opDelayOnMaster :: Bool

opDelayOnNodes :: [NonEmptyString]

opDelayOnNodeUuids :: (Maybe [NonEmptyString])

opDelayRepeat :: (NonNegative Int)

opDelayInterruptible :: Bool

opDelayNoLocks :: Bool

OpTestAllocator

Fields

opIAllocatorDirection :: IAllocatorTestDir

opIAllocatorMode :: IAllocatorMode

opIAllocatorReqName :: NonEmptyString

opIAllocatorNics :: (Maybe [INicParams])

opIAllocatorDisks :: (Maybe [JSValue])

opHypervisor :: (Maybe Hypervisor)

opIallocator :: (Maybe NonEmptyString)

opInstTags :: [NonEmptyString]

opIAllocatorMem :: (Maybe (NonNegative Int))

opIAllocatorVCpus :: (Maybe (NonNegative Int))

opIAllocatorOs :: (Maybe NonEmptyString)

opOptDiskTemplate :: (Maybe DiskTemplate)

opIAllocatorInstances :: (Maybe [NonEmptyString])

opIAllocatorEvacMode :: (Maybe EvacMode)

opTargetGroups :: (Maybe [NonEmptyString])

opIAllocatorSpindleUse :: (NonNegative Int)

opIAllocatorCount :: (NonNegative Int)

opOptGroupName :: (Maybe NonEmptyString)

OpTestJqueue

Fields

opNotifyWaitlock :: Bool

opNotifyExec :: Bool

opLogMessages :: [String]

opJQueueFail :: Bool

OpTestOsParams

Fields

opOsparamsSecret :: (Maybe (JSObject (Secret JSValue)))

OpTestDummy

Fields

opTestDummyResult :: JSValue

opTestDummyMessages :: JSValue

opTestDummyFail :: JSValue

opTestDummySubmitJobs :: JSValue

OpNetworkAdd

Fields

opNetworkName :: NonEmptyString

opNetworkAddress4 :: IPv4Network

opNetworkGateway4 :: (Maybe IPv4Address)

opNetworkAddress6 :: (Maybe IPv6Network)

opNetworkGateway6 :: (Maybe IPv6Address)

opNetMacPrefix :: (Maybe NonEmptyString)

opNetworkAddRsvdIps :: (Maybe [IPv4Address])

opConflictsCheck :: Bool

opInstTags :: [NonEmptyString]

OpNetworkRemove

Fields

opNetworkName :: NonEmptyString

opForce :: Bool

OpNetworkSetParams

Fields

opNetworkName :: NonEmptyString

opNetworkGateway4 :: (Maybe IPv4Address)

opNetworkAddress6 :: (Maybe IPv6Network)

opNetworkGateway6 :: (Maybe IPv6Address)

opNetMacPrefix :: (Maybe NonEmptyString)

opNetworkAddRsvdIps :: (Maybe [IPv4Address])

opNetworkRemoveRsvdIps :: (Maybe [IPv4Address])

OpNetworkConnect

Fields

opGroupName :: NonEmptyString

opNetworkName :: NonEmptyString

opNetworkMode :: NICMode

opNetworkLink :: NonEmptyString

opNetworkVlan :: String

opConflictsCheck :: Bool

OpNetworkDisconnect

Fields

opGroupName :: NonEmptyString

opNetworkName :: NonEmptyString

pyClasses :: [OpCodeDescriptor]

allOpFields :: String -> [String]

opID :: OpCode -> String

Returns the OP_ID for a given opcode value.

allOpIDs :: [String]

A list of all defined/supported opcode IDs.

opReasonSrcID :: OpCode -> String

Convert the opcode name to lowercase with underscores and strip the Op prefix.

opSummary :: OpCode -> String

Computes the summary of the opcode.

data CommonOpParams

Generic/common opcode parameters.

Constructors

CommonOpParams

Fields

opDryRun :: (Maybe Bool)

opDebugLevel :: (Maybe (NonNegative Int))

opPriority :: OpSubmitPriority

opDepends :: (Maybe [JobDependency])

opComment :: (Maybe String)

opReason :: ReasonTrail

defOpParams :: CommonOpParams

Default common parameter values.

data MetaOpCode

The top-level opcode type.

Constructors

MetaOpCode

Fields

metaParams :: CommonOpParams

metaOpCode :: OpCode

resolveDependencies :: Monad m => MetaOpCode -> JobId -> m MetaOpCode

Resolve relative dependencies to absolute ones, given the job Id.

wrapOpCode :: OpCode -> MetaOpCode

Wraps an OpCode with the default parameters to build a MetaOpCode.

setOpComment :: String -> MetaOpCode -> MetaOpCode

Sets the comment on a meta opcode.

setOpPriority :: OpSubmitPriority -> MetaOpCode -> MetaOpCode

Sets the priority on a meta opcode.

Orphan instances

PyValue JSValue

Instance details

Methods

showValue :: JSValue -> String

showValueList :: [JSValue] -> String

PyValue VerifyOptionalChecks

Instance details

Methods

showValue :: VerifyOptionalChecks -> String

showValueList :: [VerifyOptionalChecks] -> String

PyValue CVErrorCode

Instance details

Methods

showValue :: CVErrorCode -> String

showValueList :: [CVErrorCode] -> String

PyValue ExportMode

Instance details

Methods

showValue :: ExportMode -> String

showValueList :: [ExportMode] -> String

PyValue DiskIndex

Instance details

Methods

showValue :: DiskIndex -> String

showValueList :: [DiskIndex] -> String

PyValue INicParams

Instance details

Methods

showValue :: INicParams -> String

showValueList :: [INicParams] -> String

PyValue RecreateDisksInfo

Instance details

Methods

showValue :: RecreateDisksInfo -> String

showValueList :: [RecreateDisksInfo] -> String

PyValue IDiskParams

Instance details

Methods

showValue :: IDiskParams -> String

showValueList :: [IDiskParams] -> String

PyValue a => PyValue (JSObject a)

Instance details

Methods

showValue :: JSObject a -> String

showValueList :: [JSObject a] -> String

PyValue a => PyValue (NonEmpty a)

Instance details

Methods

showValue :: NonEmpty a -> String

showValueList :: [NonEmpty a] -> String

PyValue a => PyValue (NonNegative a)

Instance details

Methods

showValue :: NonNegative a -> String

showValueList :: [NonNegative a] -> String

PyValue a => PyValue (SetParamsMods a)

Instance details

Methods

showValue :: SetParamsMods a -> String

showValueList :: [SetParamsMods a] -> String