Ganeti.Path

Description

Path-related helper functions.

Documentation

dataDir :: IO FilePath

Directory for data.

runDir :: IO FilePath

Directory for runtime files.

logDir :: IO FilePath

Directory for log files.

socketDir :: IO FilePath

Directory for Unix sockets.

livelockDir :: IO FilePath

Directory for the jobs' livelocks.

luxidMessageDir :: IO FilePath

Directory for luxid to write messages to running jobs, like requests to change the priority.

livelockFile :: FilePath -> IO FilePath

A helper for building a job's livelock file. It prepends livelockDir to a given filename.

defaultQuerySocket :: IO FilePath

The default LUXI socket for queries.

defaultWConfdSocket :: IO FilePath

The default WConfD socket for queries.

defaultMetadSocket :: IO FilePath

The default MetaD socket for communication.

confdHmacKey :: IO FilePath

Path to file containing confd's HMAC key.

clusterConfFile :: IO FilePath

Path to cluster configuration file.

lockStatusFile :: IO FilePath

Path to the file representing the lock status.

tempResStatusFile :: IO FilePath

Path to the file representing the lock status.

watcherPauseFile :: IO FilePath

Path to the watcher pause file.

nodedCertFile :: IO FilePath

Path to the noded certificate.

nodedClientCertFile :: IO FilePath

Path to the noded client certificate.

queueDir :: IO FilePath

Job queue directory.

jobQueueSerialFile :: IO FilePath

Job queue serial file.

jobQueueLockFile :: IO FilePath

Job queue lock file

jobQueueDrainFile :: IO FilePath

Job queue drain file

jobQueueArchiveSubDir :: FilePath

Job queue archive directory.

instanceReasonDir :: IO FilePath

Directory containing the reason trails for the last change of status of instances.

getInstReasonFilename :: String -> IO FilePath

The path of the file containing the reason trail for an instance, given the instance name.

jqueueExecutorPy :: IO FilePath

The path to the Python executable for starting jobs.