Ganeti.Query.Language

Description

Implementation of the Ganeti Query2 language.

THH declarations, that require ordering.

data ResultStatus

Status of a query field.

Constructors

RSNormal

RSUnknown

RSNoData

RSUnavail

RSOffline

checkRS :: MonadFail m => ResultStatus -> a -> m a

Check that ResultStatus is success or fail with descriptive message.

data FieldType

Type of a query field.

Constructors

QFTUnknown

QFTText

QFTBool

QFTNumber

QFTNumberFloat

QFTUnit

QFTTimestamp

QFTOther

data QueryTypeOp

Supported items on which Qlang works.

Constructors

QRCluster

QRInstance

QRNode

QRGroup

QROs

QRExport

QRNetwork

queryTypeOpToRaw :: QueryTypeOp -> String

data QueryTypeLuxi

Supported items on which Qlang works.

Constructors

QRLock

QRJob

QRFilter

data ItemType

Overall query type.

Constructors

ItemTypeLuxi QueryTypeLuxi

ItemTypeOpCode QueryTypeOp

Sub data types for query2 queries and responses.

type Fields = [String]

List of requested fields.

data Filter a

Query2 filter expression. It's a parameteric type since we can filter different "things"; e.g. field names, or actual field getters, etc.

Constructors

EmptyFilter
No filter at all

AndFilter [Filter a]
& [expression, ...]

OrFilter [Filter a]
| [expression, ...]

NotFilter (Filter a)
! expression

TrueFilter a
? field

EQFilter a FilterValue
(=|!=) field value

LTFilter a FilterValue
< field value

GTFilter a FilterValue
> field value

LEFilter a FilterValue
<= field value

GEFilter a FilterValue
>= field value

RegexpFilter a FilterRegex
=~ field regexp

ContainsFilter a FilterValue
=[] list-field value

filterArguments :: Filter a -> [a]

Get the "things" a filter talks about. This is useful, e.g., to decide which additional fields to fetch in a query depending on live data.

type FilterField = String

Field name to filter on.

data FilterValue

Value to compare the field value to, for filtering purposes.

Constructors

QuotedString String

NumericValue Integer

showFilterValue :: FilterValue -> JSValue

Serialiser for FilterValue. The Python code just sends this to JSON as-is, so we'll do the same.

data FilterRegex

Regexp to apply to the filter value, for filtering purposes. It holds both the string format, and the "compiled" format, so that we don't re-compile the regex at each match attempt.

mkRegex :: MonadFail m => String -> m FilterRegex

Builder for FilterRegex. We always attempt to compile the regular expression on the initialisation of the data structure; this might fail, if the RE is not well-formed.

type FieldName = String

Name of a field.

type FieldTitle = String

Title of a field, when represented in tabular format.

type FieldDoc = String

Human redable description of a field.

data FieldDefinition

Definition of a field.

Constructors

FieldDefinition

Fields

fdefName :: FieldName

fdefTitle :: FieldTitle

fdefKind :: FieldType

fdefDoc :: FieldDoc

data ResultEntry

Constructors

ResultEntry

Fields

rentryStatus :: ResultStatus

The result status

rentryValue :: Maybe ResultValue

The (optional) result value

type ResultValue = JSValue

Value of a field, in json encoding. (its type will be depending on ResultStatus and FieldType)

Main Qlang queries and responses.

data Query

Query2 query.

Constructors

Query ItemType Fields (Filter FilterField)

data QueryResult

Query2 result.

Constructors

QueryResult

Fields

qresFields :: [FieldDefinition]

qresData :: [ResultRow]

data QueryFields

Query2 Fields query. (to get supported fields names, descriptions, and types)

Constructors

QueryFields ItemType Fields

data QueryFieldsResult

Query2 Fields result.

Constructors

QueryFieldsResult

Fields

qfieldresFields :: [FieldDefinition]