Ganeti.Query.Node

Description

Implementation of the Ganeti Query2 node queries.

Documentation

type Runtime = Either RpcError RpcResultNodeInfo

Runtime is the resulting type for NodeInfo call.

nodeLiveFieldsDefs :: [(FieldName, FieldTitle, FieldType, String, FieldDoc)]

List of node live fields.

getAttrFromStorageInfo :: JSON a => (StorageInfo -> Maybe a) -> Maybe StorageInfo -> JSValue

Helper function to extract an attribute from a maybe StorageType

isStorageInfoOfType :: StorageType -> StorageInfo -> Bool

Check whether the given storage info fits to the given storage type

getStorageInfoForDefault :: [StorageInfo] -> Maybe StorageInfo

Get storage info for the default storage unit

getStorageInfoForType :: [StorageInfo] -> StorageType -> Maybe StorageInfo

Gets the storage info for a storage type FIXME: This needs to be extended when storage pools are implemented, because storage types are not necessarily unique then

nodeLiveFieldExtract :: FieldName -> RpcResultNodeInfo -> JSValue

Map each name to a function that extracts that value from the RPC result.

nodeLiveRpcCall :: FieldName -> Runtime -> Node -> ResultEntry

Helper for extracting field from RPC result.

nodeLiveFieldBuilder :: (FieldName, FieldTitle, FieldType, String, FieldDoc) -> FieldData Node Runtime

Builder for node live fields.

nodeRoleDoc :: String

The docstring for the node role. Note that we use reverse in order to keep the same order as Python.

getNodePower :: ConfigData -> Node -> ResultEntry

Get node powered status.

nodeFields :: FieldList Node Runtime

List of all node fields.

getNumInstances :: (([Instance], [Instance]) -> [Instance]) -> ConfigData -> Node -> Int

Helper function to retrieve the number of (primary or secondary) instances

fieldsMap :: FieldMap Node Runtime

The node fields map.

rpcResultNodeBroken :: Node -> (Node, Runtime)

Create an RPC result for a broken node

storageFields :: [String]

Storage-related query fields

hypervisorFields :: [String]

Hypervisor-related query fields

queryDomainRequired :: [String] -> [String] -> Bool

Check if it is required to include domain-specific entities (for example storage units for storage info, hypervisor specs for hypervisor info) in the node_info call

collectLiveData :: Bool -> ConfigData -> [String] -> [Node] -> IO [(Node, Runtime)]

Collect live data from RPC query if enabled.