Ganeti.Rpc

Description

Implementation of the RPC client.

Base RPC functionality and types

curlOpts :: [CurlOption]

The curl options used for RPC.

data RpcError

Data type for RPC error reporting.

Constructors

CurlLayerError String

JsonDecodeError String

RpcResultError String

OfflineNodeError

explainRpcError :: RpcError -> String

Provide explanation to RPC errors.

type ERpcError = Either RpcError

class ArrayObject a => RpcCall a where

A generic class for RPC calls.

Minimal complete definition

rpcCallName, rpcCallTimeout, rpcCallAcceptOffline

Methods

rpcCallName :: a -> String

Give the (Python) name of the procedure.

rpcCallTimeout :: a -> Int

Calculate the timeout value for the call execution.

rpcCallData :: a -> String

Prepare arguments of the call to be send as POST.

rpcCallAcceptOffline :: a -> Bool

Whether we accept offline nodes when making a call.

class (RpcCall a, JSON b) => Rpc a b | a -> b, b -> a where

Generic class that ensures matching RPC call with its respective result.

Methods

rpcResultFill :: a -> JSValue -> ERpcError b

Create a result based on the received HTTP response.

data HttpClientRequest

Http Request definition.

Constructors

HttpClientRequest

Fields

requestUrl :: String

The actual URL for the node endpoint

requestData :: String

The arguments for the call

requestOpts :: [CurlOption]

The various curl options

isIpV6 :: String -> Bool

Check if a string represented address is IPv6

prepareUrl :: RpcCall a => Int -> Node -> a -> String

Prepare url for the HTTP request.

prepareHttpRequest :: RpcCall a => Int -> [CurlOption] -> Node -> String -> a -> ERpcError HttpClientRequest

Create HTTP request for a given node provided it is online, otherwise create empty response.

parseHttpReply :: Rpc a b => a -> ERpcError (CurlCode, String) -> ERpcError b

Parse an HTTP reply.

parseHttpResponse :: Rpc a b => a -> String -> ERpcError b

Parse a result based on the received HTTP response.

rpcErrors :: [(a, ERpcError b)] -> [(a, RpcError)]

Scan the list of results produced by executeRpcCall and extract all the RPC errors.

logRpcErrors :: (MonadLog m, Show a) => [(a, ERpcError b)] -> m [(a, RpcError)]

Scan the list of results produced by executeRpcCall and log all the RPC errors. Returns the list of errors for further processing.

getOptionsForCall :: Rpc a b => FilePath -> FilePath -> a -> [CurlOption]

Get options for RPC call

getNodedPort :: IO Int

Determine to port to call noded at.

executeRpcCalls :: Rpc a b => [(Node, a)] -> IO [(Node, ERpcError b)]

Execute multiple distinct RPC calls in parallel

executeRpcCalls' :: Rpc a b => [(Node, a, String)] -> IO [(Node, ERpcError b)]

Execute multiple RPC calls in parallel

executeRpcCall :: Rpc a b => [Node] -> a -> IO [(Node, ERpcError b)]

Execute an RPC call for many nodes in parallel. NB this computes the RPC call payload string only once.

sanitizeDictResults :: [(String, Result a)] -> ERpcError [(String, a)]

Helper function that is used to read dictionaries of values.

fromJResultToRes :: Result a -> (a -> b) -> ERpcError b

Helper function to tranform JSON Result to Either RpcError b. Note: For now we really only use it for b s.t. Rpc c b for some c

fromJSValueToRes :: JSON a => JSValue -> (a -> b) -> ERpcError b

Helper function transforming JSValue to Rpc result type.

newtype Compressed

An opaque data type for representing data that might be compressed over the wire.

On Python side it is decompressed by backend._Decompress.

Constructors

Compressed

Fields

getCompressed :: ByteString

packCompressed :: ByteString -> Compressed

toCompressed :: String -> Compressed

RPC calls and results

Instance info

data RpcCallInstanceInfo

Returns information about a single instance

Constructors

RpcCallInstanceInfo

Fields

rpcCallInstInfoInstance :: String

rpcCallInstInfoHname :: Hypervisor

loadRpcCallInstanceInfo :: JSValue -> Result RpcCallInstanceInfo

saveRpcCallInstanceInfo :: RpcCallInstanceInfo -> JSValue

data InstanceState

Constructors

InstanceStateRunning

InstanceStateShutdown

instanceStateFromRaw :: forall m. (Monad m, MonadFail m) => Int -> m InstanceState

instanceStateToRaw :: InstanceState -> Int

_InstanceStateShutdown :: Int

_InstanceStateRunning :: Int

data InstanceInfo

Constructors

InstanceInfo

Fields

instInfoMemory :: Int

instInfoState :: InstanceState

instInfoVcpus :: Int

instInfoTime :: Int

loadInstanceInfo :: JSValue -> Result InstanceInfo

saveInstanceInfo :: InstanceInfo -> JSValue

data RpcResultInstanceInfo

Constructors

RpcResultInstanceInfo

Fields

rpcResInstInfoInstInfo :: (Maybe InstanceInfo)

loadRpcResultInstanceInfo :: JSValue -> Result RpcResultInstanceInfo

saveRpcResultInstanceInfo :: RpcResultInstanceInfo -> JSValue

AllInstancesInfo

data RpcCallAllInstancesInfo

Returns information about all running instances on the given nodes

Constructors

RpcCallAllInstancesInfo

Fields

rpcCallAllInstInfoHypervisors :: [(Hypervisor, HvParams)]

loadRpcCallAllInstancesInfo :: JSValue -> Result RpcCallAllInstancesInfo

saveRpcCallAllInstancesInfo :: RpcCallAllInstancesInfo -> JSValue

data RpcResultAllInstancesInfo

Constructors

RpcResultAllInstancesInfo

Fields

rpcResAllInstInfoInstances :: [(String, InstanceInfo)]

loadRpcResultAllInstancesInfo :: JSValue -> Result RpcResultAllInstancesInfo

saveRpcResultAllInstancesInfo :: RpcResultAllInstancesInfo -> JSValue

InstanceConsoleInfo

data InstanceConsoleInfoParams

Returns information about how to access instances on the given node

Constructors

InstanceConsoleInfoParams

Fields

instConsInfoParamsInstance :: Instance

instConsInfoParamsNode :: Node

instConsInfoParamsGroup :: NodeGroup

instConsInfoParamsHvParams :: HvParams

instConsInfoParamsBeParams :: FilledBeParams

loadInstanceConsoleInfoParams :: JSValue -> Result InstanceConsoleInfoParams

saveInstanceConsoleInfoParams :: InstanceConsoleInfoParams -> JSValue

data RpcCallInstanceConsoleInfo

Constructors

RpcCallInstanceConsoleInfo

Fields

rpcCallInstConsInfoInstanceInfo :: [(String, InstanceConsoleInfoParams)]

loadRpcCallInstanceConsoleInfo :: JSValue -> Result RpcCallInstanceConsoleInfo

saveRpcCallInstanceConsoleInfo :: RpcCallInstanceConsoleInfo -> JSValue

data InstanceConsoleInfo

Constructors

InstanceConsoleInfo

Fields

instConsInfoInstance :: String

instConsInfoKind :: String

instConsInfoMessage :: (Maybe String)

instConsInfoHost :: (Maybe String)

instConsInfoPort :: (Maybe Int)

instConsInfoUser :: (Maybe String)

instConsInfoCommand :: (Maybe [String])

instConsInfoDisplay :: (Maybe String)

loadInstanceConsoleInfo :: JSValue -> Result InstanceConsoleInfo

saveInstanceConsoleInfo :: InstanceConsoleInfo -> JSValue

data RpcResultInstanceConsoleInfo

Constructors

RpcResultInstanceConsoleInfo

Fields

rpcResInstConsInfoInstancesInfo :: [(String, InstanceConsoleInfo)]

loadRpcResultInstanceConsoleInfo :: JSValue -> Result RpcResultInstanceConsoleInfo

saveRpcResultInstanceConsoleInfo :: RpcResultInstanceConsoleInfo -> JSValue

InstanceList

data RpcCallInstanceList

Returns the list of running instances on the given nodes

Constructors

RpcCallInstanceList

Fields

rpcCallInstListHypervisors :: [Hypervisor]

loadRpcCallInstanceList :: JSValue -> Result RpcCallInstanceList

saveRpcCallInstanceList :: RpcCallInstanceList -> JSValue

data RpcResultInstanceList

Constructors

RpcResultInstanceList

Fields

rpcResInstListInstances :: [String]

loadRpcResultInstanceList :: JSValue -> Result RpcResultInstanceList

saveRpcResultInstanceList :: RpcResultInstanceList -> JSValue

NodeInfo

data RpcCallNodeInfo

Returns node information

Constructors

RpcCallNodeInfo

Fields

rpcCallNodeInfoStorageUnits :: [StorageUnit]

rpcCallNodeInfoHypervisors :: [(Hypervisor, HvParams)]

loadRpcCallNodeInfo :: JSValue -> Result RpcCallNodeInfo

saveRpcCallNodeInfo :: RpcCallNodeInfo -> JSValue

data StorageInfo

Constructors

StorageInfo

Fields

storageInfoName :: String

storageInfoType :: String

storageInfoStorageFree :: (Maybe Int)

storageInfoStorageSize :: (Maybe Int)

loadStorageInfo :: JSValue -> Result StorageInfo

saveStorageInfo :: StorageInfo -> JSValue

data HvInfo

Common fields (as described in hv_base.py) are mandatory, other fields are optional.

Constructors

HvInfo

Fields

hvInfoHvVersion :: (Maybe [Int])

hvInfoMemoryTotal :: Int

hvInfoMemoryFree :: Int

hvInfoMemoryDom0 :: Int

hvInfoMemoryHv :: (Maybe Int)

hvInfoCpuTotal :: Int

hvInfoCpuNodes :: Int

hvInfoCpuSockets :: Int

hvInfoCpuDom0 :: Int

loadHvInfo :: JSValue -> Result HvInfo

saveHvInfo :: HvInfo -> JSValue

data RpcResultNodeInfo

Constructors

RpcResultNodeInfo

Fields

rpcResNodeInfoBootId :: String

rpcResNodeInfoStorageInfo :: [StorageInfo]

rpcResNodeInfoHvInfo :: [HvInfo]

loadRpcResultNodeInfo :: JSValue -> Result RpcResultNodeInfo

saveRpcResultNodeInfo :: RpcResultNodeInfo -> JSValue

Version

data RpcCallVersion

Query node version.

Constructors

RpcCallVersion

loadRpcCallVersion :: JSValue -> Result RpcCallVersion

saveRpcCallVersion :: RpcCallVersion -> JSValue

data RpcResultVersion

Query node reply.

Constructors

RpcResultVersion

Fields

rpcResultVersionVersion :: Int

loadRpcResultVersion :: JSValue -> Result RpcResultVersion

saveRpcResultVersion :: RpcResultVersion -> JSValue

StorageList

data RpcCallStorageList

Constructors

RpcCallStorageList

Fields

rpcCallStorageListSuName :: StorageType

rpcCallStorageListSuArgs :: [String]

rpcCallStorageListName :: String

rpcCallStorageListFields :: [StorageField]

loadRpcCallStorageList :: JSValue -> Result RpcCallStorageList

saveRpcCallStorageList :: RpcCallStorageList -> JSValue

data RpcResultStorageList

Constructors

RpcResultStorageList

Fields

rpcResStorageListStorage :: [[(StorageField, JSValue)]]

loadRpcResultStorageList :: JSValue -> Result RpcResultStorageList

saveRpcResultStorageList :: RpcResultStorageList -> JSValue

TestDelay

data RpcCallTestDelay

Call definition for test delay.

Constructors

RpcCallTestDelay

Fields

rpcCallTestDelayDuration :: Double

loadRpcCallTestDelay :: JSValue -> Result RpcCallTestDelay

saveRpcCallTestDelay :: RpcCallTestDelay -> JSValue

data RpcResultTestDelay

Result definition for test delay.

Constructors

RpcResultTestDelay

ExportList

data RpcCallExportList

Call definition for export list.

Constructors

RpcCallExportList

loadRpcCallExportList :: JSValue -> Result RpcCallExportList

saveRpcCallExportList :: RpcCallExportList -> JSValue

data RpcResultExportList

Result definition for export list.

Constructors

RpcResultExportList

Fields

rpcResExportListExports :: [String]

loadRpcResultExportList :: JSValue -> Result RpcResultExportList

saveRpcResultExportList :: RpcResultExportList -> JSValue

Job Queue Replication

data RpcCallJobqueueUpdate

Update a job queue file

Constructors

RpcCallJobqueueUpdate

Fields

rpcCallJobqueueUpdateFileName :: String

rpcCallJobqueueUpdateContent :: String

loadRpcCallJobqueueUpdate :: JSValue -> Result RpcCallJobqueueUpdate

saveRpcCallJobqueueUpdate :: RpcCallJobqueueUpdate -> JSValue

data RpcResultJobQueueUpdate

Constructors

RpcResultJobQueueUpdate

loadRpcResultJobQueueUpdate :: JSValue -> Result RpcResultJobQueueUpdate

saveRpcResultJobQueueUpdate :: RpcResultJobQueueUpdate -> JSValue

data RpcCallJobqueueRename

Rename a file in the job queue

Constructors

RpcCallJobqueueRename

Fields

rpcCallJobqueueRenameRename :: [(String, String)]

loadRpcCallJobqueueRename :: JSValue -> Result RpcCallJobqueueRename

saveRpcCallJobqueueRename :: RpcCallJobqueueRename -> JSValue

data RpcResultJobqueueRename

Constructors

RpcResultJobqueueRename

loadRpcResultJobqueueRename :: JSValue -> Result RpcResultJobqueueRename

saveRpcResultJobqueueRename :: RpcResultJobqueueRename -> JSValue

Watcher Status Update

data RpcCallSetWatcherPause

Set the watcher status

Constructors

RpcCallSetWatcherPause

Fields

rpcCallSetWatcherPauseTime :: (Maybe ClockTime)

loadRpcCallSetWatcherPause :: JSValue -> Result RpcCallSetWatcherPause

saveRpcCallSetWatcherPause :: RpcCallSetWatcherPause -> JSValue

data RpcResultSetWatcherPause

Constructors

RpcResultSetWatcherPause

loadRpcResultSetWatcherPause :: JSValue -> Result RpcResultSetWatcherPause

saveRpcResultSetWatcherPause :: RpcResultSetWatcherPause -> JSValue

Queue drain status

data RpcCallSetDrainFlag

Set the queu drain flag

Constructors

RpcCallSetDrainFlag

Fields

rpcCallSetDrainFlagValue :: Bool

loadRpcCallSetDrainFlag :: JSValue -> Result RpcCallSetDrainFlag

saveRpcCallSetDrainFlag :: RpcCallSetDrainFlag -> JSValue

data RpcResultSetDrainFlag

Constructors

RpcResultSetDrainFlag

loadRpcResultSetDrainFlag :: JSValue -> Result RpcResultSetDrainFlag

saveRpcResultSetDrainFlag :: RpcResultSetDrainFlag -> JSValue

Configuration files upload to nodes

data RpcCallUploadFile

Upload a configuration file to nodes

Constructors

RpcCallUploadFile

Fields

rpcCallUploadFileFileName :: FilePath

rpcCallUploadFileContent :: Compressed

rpcCallUploadFileMode :: (Maybe FileMode)

rpcCallUploadFileUid :: String

rpcCallUploadFileGid :: String

rpcCallUploadFileAtime :: ClockTime

rpcCallUploadFileMtime :: ClockTime

loadRpcCallUploadFile :: JSValue -> Result RpcCallUploadFile

saveRpcCallUploadFile :: RpcCallUploadFile -> JSValue

data RpcResultUploadFile

Constructors

RpcResultUploadFile

loadRpcResultUploadFile :: JSValue -> Result RpcResultUploadFile

saveRpcResultUploadFile :: RpcResultUploadFile -> JSValue

prepareRpcCallUploadFile :: RuntimeEnts -> FilePath -> ResultG RpcCallUploadFile

Reads a file and constructs the corresponding RpcCallUploadFile value.

data RpcCallWriteSsconfFiles

Upload ssconf files to nodes

Constructors

RpcCallWriteSsconfFiles

Fields

rpcCallWriteSsconfFilesValues :: SSConf

loadRpcCallWriteSsconfFiles :: JSValue -> Result RpcCallWriteSsconfFiles

saveRpcCallWriteSsconfFiles :: RpcCallWriteSsconfFiles -> JSValue

data RpcResultWriteSsconfFiles

Constructors

RpcResultWriteSsconfFiles

loadRpcResultWriteSsconfFiles :: JSValue -> Result RpcResultWriteSsconfFiles

saveRpcResultWriteSsconfFiles :: RpcResultWriteSsconfFiles -> JSValue

data RpcCallNodeActivateMasterIp

Activate the master IP address

Constructors

RpcCallNodeActivateMasterIp

Fields

rpcCallNodeActivateMasterIpParams :: MasterNetworkParameters

rpcCallNodeActivateMasterIpEms :: Bool

loadRpcCallNodeActivateMasterIp :: JSValue -> Result RpcCallNodeActivateMasterIp

saveRpcCallNodeActivateMasterIp :: RpcCallNodeActivateMasterIp -> JSValue

data RpcResultNodeActivateMasterIp

Constructors

RpcResultNodeActivateMasterIp

loadRpcResultNodeActivateMasterIp :: JSValue -> Result RpcResultNodeActivateMasterIp

saveRpcResultNodeActivateMasterIp :: RpcResultNodeActivateMasterIp -> JSValue

data RpcCallMasterNodeName

Ask who the node believes is the master.

Constructors

RpcCallMasterNodeName

loadRpcCallMasterNodeName :: JSValue -> Result RpcCallMasterNodeName

saveRpcCallMasterNodeName :: RpcCallMasterNodeName -> JSValue

data RpcResultMasterNodeName

Constructors

RpcResultMasterNodeName

Fields

rpcResultMasterNodeNameMaster :: String

loadRpcResultMasterNodeName :: JSValue -> Result RpcResultMasterNodeName

saveRpcResultMasterNodeName :: RpcResultMasterNodeName -> JSValue