Ganeti.Storage.Diskstats.Types

Description

Diskstats data types

This module holds the definition of the data types describing the status of the disks according to the information contained in procdiskstats.

Documentation

data Diskstats

This is the format of the report produced by each data collector.

Constructors

Diskstats

Fields

dsMajor :: Int

dsMinor :: Int

dsName :: String

dsReadsNum :: Int

dsMergedReads :: Int

dsSecRead :: Int

dsTimeRead :: Int

dsWrites :: Int

dsMergedWrites :: Int

dsSecWritten :: Int

dsTimeWrite :: Int

dsIos :: Int

dsTimeIO :: Int

dsWIOmillis :: Int