Ganeti.Storage.Utils

Description

Implementation of Utility functions for storage

Documentation

getDefaultStorageKey :: ConfigData -> DiskTemplate -> Maybe StorageKey

Get the cluster's default storage unit for a given disk template

getDefaultSpindleSU :: ConfigData -> (StorageType, Maybe StorageKey)

Get the cluster's default spindle storage unit

getClusterStorageUnitRaws :: ConfigData -> [StorageUnitRaw]

Get the cluster's storage units from the configuration

foldSUs :: [(StorageType, Maybe StorageKey)] -> [StorageUnitRaw]

fold the storage unit list by sorting out the ones without keys

getExclusiveStorage :: ConfigData -> Node -> Maybe Bool

Gets the value of the 'exclusive storage' flag of the node

hasExclusiveStorageFlag :: ConfigData -> Node -> Bool

Determines whether a node's config contains an 'exclusive storage' flag

nodesWithValidConfig :: ConfigData -> [Node] -> [Node]

Filter for nodes with a valid config

getStorageUnitsOfNode :: ConfigData -> Node -> [StorageUnit]

Get the storage units of the node