Ganeti.THH.Field

Description

Helpers for creating various kinds of Fields.

They aren't directly needed for the Template Haskell code in Ganeti.THH, so better keep them in a separate module.

Internal functions

numericalReadFn :: JSON a => (String -> Result a) -> [(String, JSValue)] -> JSValue -> Result a

Wrapper around a special parse function, suitable as field-parsing function.

specialNumericalField :: Name -> Field -> Field

Sets the read function to also accept string parsable by the given function.

timeAsDoubleField :: String -> Field

Creates a new mandatory field that reads time as the (floating point) number of seconds since the standard UNIX epoch, and represents it in Haskell as ClockTime.

integralField :: Q Type -> String -> Field

A helper function for creating fields whose Haskell representation is Integral and which are serialized as numbers.

External functions and data types

timeStampFields :: [Field]

Timestamp fields description.

serialFields :: [Field]

Serial number fields description.

uuidFields :: [Field]

UUID fields description.

newtype TagSet

Tag set type.

Constructors

TagSet

Fields

unTagSet :: Set String

emptyTagSet :: TagSet

Empty tag set value.

tagsFields :: [Field]

Tag field description.

Fields related to POSIX data types

fileModeAsIntField :: String -> Field

Creates a new mandatory field that reads a file mode in the standard POSIX file mode representation. The Haskell type of the field is FileMode.

processIdField :: String -> Field

Creates a new mandatory field that contains a POSIX process ID.