Ganeti.THH.PyType

Description

PyType helper for Ganeti Haskell code.

Documentation

data PyType

Represents a Python encoding of types.

Constructors

PTMaybe PyType

PTApp PyType [PyType]

PTOther String

PTAny

PTDictOf

PTListOf

PTNone

PTObject

PTOr

PTSetOf

PTTupleOf

pyType :: Type -> Q PyType

Converts a Haskell type into a Python type.

pyOptionalType :: Bool -> Type -> Q PyType

pyType opt typ converts Haskell type typ into a Python type, where opt determines if the converted type is optional (i.e., Maybe).