Ganeti.THH.RPC

Description

Implements Template Haskell generation of RPC server components from Haskell functions.

Documentation

type RpcServer m = Handler Request m JSValue

data Request

A RPC request consiting of a method and its argument(s).

Constructors

Request

Fields

rMethod :: String

rArgs :: JSValue

dispatch :: Monad m => Map String (RpcFn (ResultT GanetiException m)) -> RpcServer m

mkRpcM

Arguments

:: [Name]
the names of functions to include

-> Q Exp

Takes a list of function names and creates a RPC handler that delegates calls to them.

The functions must conform to (J.JSON i, J.JSON o) => i -> ResultT GanetiException m o. The m monads types of all the functions must unify.

The result expression is of type RpcServer m.