Ganeti.Types

Description

Some common Ganeti types.

This holds types common to both core work, and to htools. Types that are very core specific (e.g. configuration objects) should go in Objects, while types that are specific to htools in-memory representation should go into Types.

Generic types

data NonNegative a

Type that holds a non-negative value.

mkNonNegative :: (Monad m, Num a, Ord a, Show a) => a -> m (NonNegative a)

Smart constructor for NonNegative.

data Positive a

Type that holds a positive value.

mkPositive :: (Monad m, Num a, Ord a, Show a) => a -> m (Positive a)

Smart constructor for Positive.

data Negative a

Type that holds a negative value.

mkNegative :: (Monad m, Num a, Ord a, Show a) => a -> m (Negative a)

Smart constructor for Negative.

data NonEmpty a

Type that holds a non-null list.

mkNonEmpty :: Monad m => [a] -> m (NonEmpty a)

Smart constructor for NonEmpty.

type NonEmptyString = NonEmpty Char

A simple type alias for non-empty strings.

type QueryResultCode = Int

data IPv4Address

mkIPv4Address :: Monad m => String -> m IPv4Address

data IPv4Network

mkIPv4Network :: Monad m => String -> m IPv4Network

data IPv6Address

mkIPv6Address :: Monad m => String -> m IPv6Address

data IPv6Network

mkIPv6Network :: Monad m => String -> m IPv6Network

Ganeti types

data DiskTemplate

Instance disk template type. The disk template is a name for the constructor of the disk configuration DiskLogicalId used for serialization, configuration values, etc.

Constructors

DTDiskless

DTFile

DTSharedFile

DTPlain

DTBlock

DTDrbd8

DTRbd

DTExt

DTGluster

diskTemplateFromRaw :: forall m. Monad m => String -> m DiskTemplate

diskTemplateToRaw :: DiskTemplate -> String

diskTemplateMovable :: DiskTemplate -> Bool

Predicate on disk templates indicating if instances based on this disk template can freely be moved (to any node in the node group).

data TagKind

Data type representing what items the tag operations apply to.

Constructors

TagKindInstance

TagKindNode

TagKindGroup

TagKindCluster

TagKindNetwork

tagKindFromRaw :: forall m. Monad m => String -> m TagKind

tagKindToRaw :: TagKind -> String

data AllocPolicy

The Group allocation policy type.

Note that the order of constructors is important as the automatic Ord instance will order them in the order they are defined, so when changing this data type be careful about the interaction with the desired sorting order.

Constructors

AllocPreferred

AllocLastResort

AllocUnallocable

allocPolicyFromRaw :: forall m. Monad m => String -> m AllocPolicy

allocPolicyToRaw :: AllocPolicy -> String

data InstanceStatus

The Instance real state type.

Constructors

StatusDown

StatusOffline

ErrorDown

ErrorUp

NodeDown

NodeOffline

Running

UserDown

WrongNode

instanceStatusFromRaw :: forall m. Monad m => String -> m InstanceStatus

instanceStatusToRaw :: InstanceStatus -> String

data MigrationMode

Migration mode.

Constructors

MigrationLive

MigrationNonLive

migrationModeToRaw :: MigrationMode -> String

data VerifyOptionalChecks

Verify optional checks.

Constructors

VerifyNPlusOneMem

verifyOptionalChecksToRaw :: VerifyOptionalChecks -> String

data CVErrorCode

Cluster verify error codes.

Constructors

CvECLUSTERCFG

CvECLUSTERCERT

CvECLUSTERCLIENTCERT

CvECLUSTERFILECHECK

CvECLUSTERDANGLINGNODES

CvECLUSTERDANGLINGINST

CvEINSTANCEBADNODE

CvEINSTANCEDOWN

CvEINSTANCELAYOUT

CvEINSTANCEMISSINGDISK

CvEINSTANCEFAULTYDISK

CvEINSTANCEWRONGNODE

CvEINSTANCESPLITGROUPS

CvEINSTANCEPOLICY

CvEINSTANCEUNSUITABLENODE

CvEINSTANCEMISSINGCFGPARAMETER

CvENODEDRBD

CvENODEDRBDVERSION

CvENODEDRBDHELPER

CvENODEFILECHECK

CvENODEHOOKS

CvENODEHV

CvENODELVM

CvENODEN1

CvENODENET

CvENODEOS

CvENODEORPHANINSTANCE

CvENODEORPHANLV

CvENODERPC

CvENODESSH

CvENODEVERSION

CvENODESETUP

CvENODETIME

CvENODEOOBPATH

CvENODEUSERSCRIPTS

CvENODEFILESTORAGEPATHS

CvENODEFILESTORAGEPATHUNUSABLE

CvENODESHAREDFILESTORAGEPATHUNUSABLE

CvENODEGLUSTERSTORAGEPATHUNUSABLE

CvEGROUPDIFFERENTPVSIZE

CvEEXTAGS

cVErrorCodeToRaw :: CVErrorCode -> String

data DdmSimple

Dynamic device modification, just add/remove version.

Constructors

DdmSimpleAdd

DdmSimpleAttach

DdmSimpleRemove

DdmSimpleDetach

data DdmFull

Dynamic device modification, all operations version.

TODO: DDM_SWAP, DDM_MOVE?

Constructors

DdmFullAdd

DdmFullAttach

DdmFullRemove

DdmFullDetach

DdmFullModify

ddmFullToRaw :: DdmFull -> String

data Hypervisor

Hypervisor type definitions.

Constructors

Kvm

XenPvm

Chroot

XenHvm

Lxc

Fake

hypervisorFromRaw :: forall m. Monad m => String -> m Hypervisor

hypervisorToRaw :: Hypervisor -> String

data OobCommand

Oob command type.

Constructors

OobHealth

OobPowerCycle

OobPowerOff

OobPowerOn

OobPowerStatus

oobCommandToRaw :: OobCommand -> String

data OobStatus

Oob command status

Constructors

OobStatusCritical

OobStatusOk

OobStatusUnknown

OobStatusWarning

oobStatusToRaw :: OobStatus -> String

data StorageType

Storage type.

Constructors

StorageFile

StorageSharedFile

StorageGluster

StorageLvmPv

StorageLvmVg

StorageDiskless

StorageBlock

StorageRados

StorageExt

storageTypeToRaw :: StorageType -> String

type StorageKey = String

Storage keys are identifiers for storage units. Their content varies depending on the storage type, for example a storage key for LVM storage is the volume group name.

data StorageUnitRaw

Storage units without storage-type-specific parameters

Constructors

SURaw StorageType StorageKey

data StorageUnit

Full storage unit with storage-type-specific parameters

Constructors

SUFile StorageKey

SUSharedFile StorageKey

SUGluster StorageKey

SULvmPv StorageKey SPExclusiveStorage

SULvmVg StorageKey SPExclusiveStorage

SUDiskless StorageKey

SUBlock StorageKey

SURados StorageKey

SUExt StorageKey

diskTemplateToStorageType :: DiskTemplate -> StorageType

Mapping from disk templates to storage types.

addParamsToStorageUnit :: SPExclusiveStorage -> StorageUnitRaw -> StorageUnit

Equips a raw storage unit with its parameters

data EvacMode

Node evac modes.

This is part of the IAllocator interface and it is used, for example, in RqType. However, it must reside in this module, and not in Types, because it is also used by Constants.

Constructors

ChangePrimary

ChangeSecondary

ChangeAll

evacModeToRaw :: EvacMode -> String

data FileDriver

The file driver type.

Constructors

FileLoop

FileBlktap

FileBlktap2

fileDriverToRaw :: FileDriver -> String

data InstCreateMode

The instance create mode.

Constructors

InstCreate

InstImport

InstRemoteImport

instCreateModeToRaw :: InstCreateMode -> String

data RebootType

Reboot type.

Constructors

RebootSoft

RebootHard

RebootFull

rebootTypeToRaw :: RebootType -> String

data ExportMode

Export modes.

Constructors

ExportModeLocal

ExportModeRemote

exportModeToRaw :: ExportMode -> String

data IAllocatorTestDir

IAllocator run types (OpTestIAllocator).

Constructors

IAllocatorDirIn

IAllocatorDirOut

iAllocatorTestDirToRaw :: IAllocatorTestDir -> String

data IAllocatorMode

IAllocator mode. FIXME: use this in HTools.Backend.IAlloc.

Constructors

IAllocatorAlloc

IAllocatorAllocateSecondary

IAllocatorMultiAlloc

IAllocatorReloc

IAllocatorNodeEvac

IAllocatorChangeGroup

iAllocatorModeToRaw :: IAllocatorMode -> String

data NICMode

Network mode.

Constructors

NMBridged

NMRouted

NMOvs

NMPool

nICModeToRaw :: NICMode -> String

data JobStatus

The JobStatus data type. Note that this is ordered especially such that greater/lesser comparison on values of this type makes sense.

Constructors

JOB_STATUS_QUEUED

JOB_STATUS_WAITING

JOB_STATUS_CANCELING

JOB_STATUS_RUNNING

JOB_STATUS_CANCELED

JOB_STATUS_SUCCESS

JOB_STATUS_ERROR

jobStatusFromRaw :: forall m. Monad m => String -> m JobStatus

jobStatusToRaw :: JobStatus -> String

data FinalizedJobStatus

Finalized job status.

Constructors

JobStatusCanceled

JobStatusSuccessful

JobStatusFailed

finalizedJobStatusToRaw :: FinalizedJobStatus -> String

data JobId

The Ganeti job type.

makeJobId :: Monad m => Int -> m JobId

Builds a job ID.

makeJobIdS :: Monad m => String -> m JobId

Builds a job ID from a string.

type RelativeJobId = Negative Int

Relative job ID type alias.

data JobIdDep

Job ID dependency.

Constructors

JobDepRelative RelativeJobId

JobDepAbsolute JobId

data JobDependency

Job Dependency type.

Constructors

JobDependency JobIdDep [FinalizedJobStatus]

absoluteJobDependency :: Monad m => JobDependency -> JobId -> m JobDependency

From job dependency and job id compute an absolute job dependency.

getJobIdFromDependency :: JobDependency -> [JobId]

From a job dependency get the absolute job id it depends on, if given absolutely.

data OpSubmitPriority

Valid opcode priorities for submit.

Constructors

OpPrioLow

OpPrioNormal

OpPrioHigh

opSubmitPriorityToRaw :: OpSubmitPriority -> Int

parseSubmitPriority :: Monad m => String -> m OpSubmitPriority

Parse submit priorities from a string.

fmtSubmitPriority :: OpSubmitPriority -> String

Format a submit priority as string.

data OpStatus

Our ADT for the OpCode status at runtime (while in a job).

Constructors

OP_STATUS_QUEUED

OP_STATUS_WAITING

OP_STATUS_CANCELING

OP_STATUS_RUNNING

OP_STATUS_CANCELED

OP_STATUS_SUCCESS

OP_STATUS_ERROR

opStatusFromRaw :: forall m. Monad m => String -> m OpStatus

opStatusToRaw :: OpStatus -> String

data ELogType

Type for the job message type.

Constructors

ELogMessage

ELogMessageList

ELogRemoteImport

ELogJqueueTest

ELogDelayTest

eLogTypeToRaw :: ELogType -> String

type ReasonElem = (String, String, Integer)

Type of one element of a reason trail, of form (source, reason, timestamp).

type ReasonTrail = [ReasonElem]

Type representing a reason trail.

data VType

The VTYPES, a mini-type system in Python.

Constructors

VTypeString

VTypeMaybeString

VTypeBool

VTypeSize

VTypeInt

VTypeFloat

vTypeFromRaw :: forall m. Monad m => String -> m VType

vTypeToRaw :: VType -> String

Node role type

data NodeRole

Constructors

NROffline

NRDrained

NRRegular

NRCandidate

NRMaster

nodeRoleToRaw :: NodeRole -> String

roleDescription :: NodeRole -> String

The description of the node role.

Disk types

data DiskMode

Constructors

DiskRdOnly

DiskRdWr

diskModeToRaw :: DiskMode -> String

data BlockDriver

The persistent block driver type. Currently only one type is allowed.

Constructors

BlockDrvManual

blockDriverToRaw :: BlockDriver -> String

Instance types

data AdminState

Constructors

AdminOffline

AdminDown

AdminUp

adminStateFromRaw :: forall m. Monad m => String -> m AdminState

adminStateToRaw :: AdminState -> String

data AdminStateSource

Constructors

AdminSource

UserSource

adminStateSourceFromRaw :: forall m. Monad m => String -> m AdminStateSource

adminStateSourceToRaw :: AdminStateSource -> String

Storage field type

data StorageField

Constructors

SFUsed

SFName

SFAllocatable

SFFree

SFSize

storageFieldToRaw :: StorageField -> String

Disk access protocol

data DiskAccessMode

Constructors

DiskUserspace

DiskKernelspace

diskAccessModeToRaw :: DiskAccessMode -> String

data LocalDiskStatus

Local disk status

Python code depends on: DiskStatusOk < DiskStatusUnknown < DiskStatusFaulty

Constructors

DiskStatusOk

DiskStatusSync

DiskStatusUnknown

DiskStatusFaulty

localDiskStatusFromRaw :: forall m. Monad m => Int -> m LocalDiskStatus

localDiskStatusToRaw :: LocalDiskStatus -> Int

localDiskStatusName :: LocalDiskStatus -> String

data ReplaceDisksMode

Replace disks type.

Constructors

ReplaceOnPrimary

ReplaceOnSecondary

ReplaceNewSecondary

ReplaceAuto

replaceDisksModeToRaw :: ReplaceDisksMode -> String

data RpcTimeout

Basic timeouts for RPC calls.

Constructors

Urgent

Fast

Normal

Slow

FourHours

OneDay

rpcTimeoutFromRaw :: forall m. Monad m => Int -> m RpcTimeout

rpcTimeoutToRaw :: RpcTimeout -> Int

data HotplugAction

Hotplug action.

Constructors

HAAdd

HARemove

HAMod

hotplugActionToRaw :: HotplugAction -> String

data HotplugTarget

Hotplug Device Target.

Constructors

HTDisk

HTNic

hotplugTargetToRaw :: HotplugTarget -> String

data SshKeyType

SSH key type.

Constructors

RSA

DSA

ECDSA

sshKeyTypeToRaw :: SshKeyType -> String

Private type and instances

redacted :: String

newtype Private a

A container for values that should be happy to be manipulated yet refuses to be shown unless explicitly requested.

Constructors

Private

Fields

getPrivate :: a

showPrivateJSObject :: JSON a => [(String, a)] -> JSObject (Private JSValue)

Secret type and instances

newtype Secret a

A container for values that behaves like Private, but doesn't leak the value through showJSON

Constructors

Secret

Fields

getSecret :: a

showSecretJSObject :: JSON a => [(String, a)] -> JSObject (Secret JSValue)

We return "<redacted>" here to satisfy the idempotence of serialization and deserialization, although this will impact the meaningfulness of secret parameters within configuration tests.

revealValInJSObject :: JSObject (Secret JSValue) -> JSObject (Private JSValue)

type HvParams = Container JSValue

The hypervisor parameter type. This is currently a simple map, without type checking on key/value pairs.

type OsParams = Container String

The OS parameters type. This is, and will remain, a string container, since the keys are dynamically declared by the OSes, and the values are always strings.

type OsParamsPrivate = Container (Private String)

class TimeStampObject a where

Class of objects that have timestamps.

Minimal complete definition

cTimeOf, mTimeOf

Methods

cTimeOf :: a -> ClockTime

mTimeOf :: a -> ClockTime

class UuidObject a where

Class of objects that have an UUID.

Minimal complete definition

uuidOf

Methods

uuidOf :: a -> String

class ForthcomingObject a where

Class of objects that can be forthcoming.

Minimal complete definition

isForthcoming

Methods

isForthcoming :: a -> Bool

class SerialNoObject a where

Class of object that have a serial number.

Minimal complete definition

serialOf

Methods

serialOf :: a -> Int

class TagsObject a where

Class of objects that have tags.

Minimal complete definition

tagsOf

Methods

tagsOf :: a -> TagSet