Ganeti.Utils.Statistics

Description

Utility functions for statistical accumulation.

Documentation

type TagTagMap = Map (String, String) Int

Type to store the number of instances for each exclusion and location pair. This is necessary to calculate second component of location score.

data Statistics

Abstract type of statistical accumulations. They behave as if the given statistics were computed on the list of values, but they allow a potentially more efficient update of a given value.

data AggregateComponent

Abstract type of per-node statistics measures. The SimpleNumber is used to construct SumStatistics and StdDevStatistics while SpreadValues is used to construct MapStatistics.

Constructors

SimpleNumber Double

SpreadValues TagTagMap

getSumStatistics :: [AggregateComponent] -> Statistics

Get a statistics that sums up the values.

getStdDevStatistics :: [AggregateComponent] -> Statistics

Get a statistics for the standard deviation.

getMapStatistics :: [AggregateComponent] -> Statistics

Get a statistics for the standard deviation.

getStatisticValue :: Statistics -> Double

Obtain the value of a statistics.

updateStatistics :: Statistics -> (AggregateComponent, AggregateComponent) -> Statistics

In a given statistics replace on value by another. This will only give meaningful results, if the original value was actually part of the statistics.