Ganeti.VCluster

Description

Utilities for virtual clusters.

Documentation

getRootDirectory :: IO (Maybe FilePath)

virtualPath :: FilePath -> FilePath -> FilePath

Pure computation of the virtual path from the original path and the vcluster root

makeVirtualPath :: FilePath -> IO FilePath

Given a path, make it a virtual one, if in a vcluster environment. Otherwise, return unchanged.