Ganeti.WConfd.ConfigVerify

Description

Implementation of functions specific to configuration management.

Configuration checks

keysSet :: Ord k => GenericContainer k v -> Set k

A helper function that returns the key set of a container.

checkUUIDKeys :: (UuidObject a, Show a) => String -> Container a -> ValidationMonad ()

Checks that all objects are indexed by their proper UUID.

checkUUIDRefs :: (UuidObject a, Show a, Foldable f) => String -> String -> (a -> [String]) -> f a -> Container b -> ValidationMonad ()

Checks that all linked UUID of given objects exist.

verifyConfig :: ConfigData -> ValidationMonad ()

Checks consistency of a given configuration.

TODO: Currently this implements only some very basic checks. Evenually all checks from Python ConfigWriter need to be moved here (see issue #759).

verifyConfigErr :: MonadError GanetiException m => ConfigData -> m ()

Checks consistency of a given configuration. If there is an error, throw ConfigVerifyError.