Test.Ganeti.Attoparsec

Description

Unittests for Attoparsec support for unicode

Documentation

part1 :: String

Unicode test string, first part.

part2 :: String

Unicode test string, second part.

simpleParser :: Parser (String, String)

Simple parser able to split a string in two parts, name and value, separated by a '=' sign.

case_unicodeParsing :: Assertion

Tests whether a Unicode string is still Unicode after being parsed.

testAttoparsec :: Test