Test.Ganeti.HTools.Backend.Text

Description

Unittests for ganeti-htools.

Instance text loader tests

toYN :: Bool -> String

prop_Load_Instance :: String -> Int -> Int -> Int -> InstanceStatus -> NonEmptyList Char -> String -> NonNegative Int -> NonNegative Int -> Bool -> DiskTemplate -> Int -> NonNegative Int -> Bool -> [String] -> Property

prop_Load_InstanceFail :: [(String, Int)] -> [String] -> Property

genInstanceNodes :: Gen (Instance, List, NameAssoc)

prop_InstanceLSIdempotent :: Property

prop_Load_Node :: String -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool

prop_Load_NodeFail :: [String] -> Property

prop_Load_NodeSuccess :: String -> NonNegative Int -> NonNegative Int -> NonNegative Int -> NonNegative Int -> NonNegative Int -> NonNegative Int -> Bool -> NonNegative Int -> Bool -> NonNegative Int -> NonNegative Int -> NonNegative Int -> [String] -> Property

prop_NodeLSIdempotent :: Property

prop_ISpecIdempotent :: ISpec -> Property

prop_MultipleMinMaxISpecsIdempotent :: [MinMaxISpecs] -> Property

prop_IPolicyIdempotent :: IPolicy -> Property

prop_CreateSerialise :: Property

This property, while being in the text tests, does more than just test end-to-end the serialisation and loading back workflow; it also tests the Loader.mergeData and the actual Cluster.iterateAlloc (for well-behaving w.r.t. instance allocations, not for the business logic). As such, it's a quite complex and slow test, and that's the reason we restrict it to small cluster sizes.

testHTools_Backend_Text :: Test