Test.Ganeti.HTools.Container

Description

Unittests for ganeti-htools.

Documentation

prop_addTwo :: [Key] -> Int -> Int -> Bool

prop_nameOf :: Node -> Property

prop_findByName :: Property

We test that in a cluster, given a random node, we can find it by its name and alias, as long as all names and aliases are unique, and that we fail to find a non-existing name.

testHTools_Container :: Test