Test.Ganeti.HTools.Graph

Description

Unittests for Ganeti.Htools.Graph

Documentation

data TestableGraph

Constructors

TestableGraph Graph

data TestableClique

Constructors

TestableClique Graph

undirEdges :: Gen (Bounds, [Edge])

Generate node bounds and edges for an undirected graph. A graph is undirected if for every (a, b) edge there is a corresponding (b, a) one.

cliqueEdges :: Gen (Bounds, [Edge])

Generate node bounds and edges for a clique. In a clique all nodes are directly connected to each other.

case_emptyVertColorMapNull :: Assertion

Check that the empty vertex color map is empty.

case_emptyVertColorMapEmpty :: Assertion

Check that the empty vertex color map is zero in size.

anyTwo :: (a -> a -> Bool) -> [a] -> Bool

Check if each two consecutive elements on a list respect a given condition.

prop_verticesByDegreeAscAsc :: TestableGraph -> Bool

Check order of vertices returned by verticesByDegreeAsc.

prop_verticesByDegreeDescDesc :: TestableGraph -> Bool

Check order of vertices returned by verticesByDegreeDesc.

prop_isColorableTestableGraph :: TestableGraph -> Bool

Check that our generated graphs are colorable

prop_isColorableTestableClique :: TestableClique -> Bool

Check that our generated graphs are colorable

prop_colorClique :: (Graph -> VertColorMap) -> TestableClique -> Property

Check that the given algorithm colors a clique with the same number of colors as the vertices number.

prop_colorLFClique :: TestableClique -> Property

Specific check for the LF algorithm.

prop_colorDsaturClique :: TestableClique -> Property

Specific check for the Dsatur algorithm.

prop_colorDcolorClique :: TestableClique -> Property

Specific check for the Dcolor algorithm.

prop_colorAllNodes :: (Graph -> VertColorMap) -> TestableGraph -> Property

prop_colorLFAllNodes :: TestableGraph -> Property

Specific check for the LF algorithm.

prop_colorDsaturAllNodes :: TestableGraph -> Property

Specific check for the Dsatur algorithm.

prop_colorDcolorAllNodes :: TestableGraph -> Property

Specific check for the Dcolor algorithm.

prop_colorProper :: (Graph -> VertColorMap) -> TestableGraph -> Bool

Check that no two vertices sharing the same edge have the same color.

prop_colorLFProper :: TestableGraph -> Bool

Specific check for the LF algorithm.

prop_colorDsaturProper :: TestableGraph -> Bool

Specific check for the Dsatur algorithm.

prop_colorDcolorProper :: TestableGraph -> Bool

Specific check for the Dcolor algorithm.

testHTools_Graph :: Test

List of tests for the Graph module.