Test.Ganeti.Hypervisor.Xen.XmParser

Description

Unittests for xm list --long parser

Arbitraries

genConfig :: Int -> Gen LispConfig

Generator for ListConfig.

A completely arbitrary configuration would contain too many lists and its size would be to big to be actually parsable in reasonable time. This generator builds a random Config that is still of a reasonable size, and it also Avoids generating strings that might be interpreted as numbers.

canBeNumber :: String -> Bool

Determines conservatively whether a string could be a number.

canBeNumberChar :: Char -> Bool

Determines whether a char can be part of the string representation of a number (even in scientific notation).

Helper functions for tests

testDomain :: String -> Map String Domain -> Assertion

Function for testing whether a domain configuration is parsed correctly.

testUptimeInfo :: String -> Map Int UptimeInfo -> Assertion

Function for testing whether a xm uptime output (stored in a file) is parsed correctly.

isAlmostEqual :: LispConfig -> LispConfig -> Property

Determines whether two LispConfig are equal, with the exception of Double values, that just need to be "almost equal".

Meant mainly for testing purposes, given that Double values may be slightly rounded during parsing.

serializeConf :: LispConfig -> String

Function to serialize LispConfigs in such a way that they can be rebuilt again by the lispConfigParser.

serializeUptime :: UptimeInfo -> String

Function to serialize UptimeInfos in such a way that they can be rebuilt againg by the uptimeLineParser.

prop_config :: LispConfig -> Property

Test whether a randomly generated config can be parsed. Implicitly, this also tests that the Show instance of Config is correct.

prop_uptimeInfo :: UptimeInfo -> Property

Test whether a randomly generated UptimeInfo text line can be parsed.

case_xen401list :: Assertion

Test a Xen 4.0.1 xm list --long output.

case_xen401uptime :: Assertion

Test a Xen 4.0.1 xm uptime output.

testHypervisor_Xen_XmParser :: Test

Orphan instances

Arbitrary UptimeInfo
Generates an arbitrary xm uptime output line.

Instance details

Methods

arbitrary :: Gen UptimeInfo

shrink :: UptimeInfo -> [UptimeInfo]

Arbitrary LispConfig
Arbitrary instance for LispConfig using genConfig.

Instance details

Methods

arbitrary :: Gen LispConfig

shrink :: LispConfig -> [LispConfig]