Test.Ganeti.Luxi

Description

Unittests for ganeti-htools.

Luxi tests

prop_CallEncoding :: LuxiOp -> Property

Simple check that encoding/decoding of LuxiOp works.

luxiServerPong :: Client -> IO ()

Server ping-pong helper.

luxiClientPong :: Client -> [String] -> IO [String]

Client ping-pong helper.

prop_ClientServer :: [[DNSChar]] -> Property

Monadic check that, given a server socket, we can connect via a client to it, and that we can send a list of arbitrary messages and get back what we sent.

case_AllDefined :: Assertion

Check that Python and Haskell define the same Luxi requests list.

testLuxi :: Test

Orphan instances

Arbitrary LuxiReq

Instance details

Methods

arbitrary :: Gen LuxiReq

shrink :: LuxiReq -> [LuxiReq]

Arbitrary LuxiOp

Instance details

Methods

arbitrary :: Gen LuxiOp

shrink :: LuxiOp -> [LuxiOp]