Test.Ganeti.Network

Test cases

prop_addressPoolProperties :: Network -> Property

Check that the address pool's properties are calculated correctly.

subsetMaybe :: Maybe BitArray -> Maybe BitArray -> Bool

Checks for the subset relation on Maybe values.

allReservationsSubsumesInternal :: Network -> Bool

Check that all internally reserved ips are included in allReservations.

allReservationsSubsumesExternal :: Network -> Bool

Check that all externally reserved ips are included in allReservations.

checkCounts :: Network -> Property

Check that the counts of free and reserved ips add up.

checkIsFull :: Network -> Property

Check that the detection of a full network works correctly.

checkGetMap :: Network -> Property

Check that the map representation of the network corresponds to the network's reservations.

testNetwork :: Test