Test.Ganeti.OpCodes

Description

Unittests for ganeti-htools.

Arbitrary instances

arbitraryOpTagsGet :: Gen OpCode

arbitraryOpTagsSet :: Gen OpCode

arbitraryOpTagsDel :: Gen OpCode

arbitraryDataCollector :: Gen (GenericContainer String Bool)

arbitraryDataCollectorInterval :: Gen (Maybe (GenericContainer String Int))

Helper functions

emptyJSObject :: JSObject JSValue

Empty JSObject.

emptyMUD :: Gen (Maybe (JSObject JSValue))

Empty maybe unchecked dictionary.

genEmptyContainer :: Ord a => Gen (GenericContainer a b)

Generates an empty container.

genDiskIndices :: Gen [DiskIndex]

Generates list of disk indices.

genNodeNames :: Gen [String]

Generates a list of node names.

genNodeNamesNE :: Gen [NonEmptyString]

Generates a list of node names in non-empty string type.

genNodeNameNE :: Gen NonEmptyString

Gets a node name in non-empty type.

genNameNE :: Gen NonEmptyString

Gets a name (non-fqdn) in non-empty type.

genNamesNE :: Gen [NonEmptyString]

Gets a list of names (non-fqdn) in non-empty type.

genFieldsNE :: Gen [NonEmptyString]

Returns a list of non-empty fields.

genMacPrefix :: Gen NonEmptyString

Generate a 3-byte MAC prefix.

arbitraryPrivateJSObj :: Gen (JSObject (Private JSValue))

JSObject of arbitrary data.

Since JSValue does not implement Arbitrary, I'll simply generate (String, String) objects.

arbitrarySecretJSObj :: Gen (JSObject (Secret JSValue))

JSObject of arbitrary secret data.

isJsonError :: Result a -> Bool

Small helper to check for a failed JSON deserialisation

Test cases

prop_serialization :: OpCode -> Property

Check that opcode serialization is idempotent.

case_AllDefined :: Assertion

Check that Python and Haskell defined the same opcode list.

case_py_compat_types :: Assertion

Custom HUnit test case that forks a Python process and checks correspondence between Haskell-generated OpCodes and their Python decoded, validated and re-encoded version.

Note that we have a strange beast here: since launching Python is expensive, we don't do this via a usual QuickProperty, since that's slow (I've tested it, and it's indeed quite slow). Rather, we use a single HUnit assertion, and in it we manually use QuickCheck to generate 500 opcodes times the number of defined opcodes, which then we pass in bulk to Python. The drawbacks to this method are two fold: we cannot control the number of generated opcodes, since HUnit assertions don't get access to the test options, and for the same reason we can't run a repeatable seed. We should probably find a better way to do this, for example by having a separately-launched Python process (if not running the tests would be skipped).

case_py_compat_fields :: Assertion

Custom HUnit test case that forks a Python process and checks correspondence between Haskell OpCodes fields and their Python equivalent.

prop_setOpComment :: MetaOpCode -> String -> Property

Checks that setOpComment works correctly.

prop_mkDiskIndex_fail :: Positive Int -> Property

Tests wrong (negative) disk index.

case_readRecreateDisks_fail :: Assertion

Tests a few invalid readRecreateDisks cases.

case_readDdmOldChanges_fail :: Assertion

Tests a few invalid readDdmOldChanges cases.

case_readExportTarget_fail :: Assertion

Tests a few invalid readExportTarget cases.

testOpCodes :: Test

Orphan instances

Arbitrary ReplaceDisksMode

Instance details

Methods

arbitrary :: Gen ReplaceDisksMode

shrink :: ReplaceDisksMode -> [ReplaceDisksMode]

Arbitrary DiskIndex

Instance details

Methods

arbitrary :: Gen DiskIndex

shrink :: DiskIndex -> [DiskIndex]

Arbitrary DiskAccess

Instance details

Methods

arbitrary :: Gen DiskAccess

shrink :: DiskAccess -> [DiskAccess]

Arbitrary INicParams

Instance details

Methods

arbitrary :: Gen INicParams

shrink :: INicParams -> [INicParams]

Arbitrary ExportTarget

Instance details

Methods

arbitrary :: Gen ExportTarget

shrink :: ExportTarget -> [ExportTarget]

Arbitrary DdmOldChanges

Instance details

Methods

arbitrary :: Gen DdmOldChanges

shrink :: DdmOldChanges -> [DdmOldChanges]

Arbitrary RecreateDisksInfo

Instance details

Methods

arbitrary :: Gen RecreateDisksInfo

shrink :: RecreateDisksInfo -> [RecreateDisksInfo]

Arbitrary IDiskParams

Instance details

Methods

arbitrary :: Gen IDiskParams

shrink :: IDiskParams -> [IDiskParams]

Arbitrary OpCode

Instance details

Methods

arbitrary :: Gen OpCode

shrink :: OpCode -> [OpCode]

Arbitrary MetaOpCode
Arbitrary instance for MetaOpCode, defined here due to TH ordering.

Instance details

Methods

arbitrary :: Gen MetaOpCode

shrink :: MetaOpCode -> [MetaOpCode]

Arbitrary CommonOpParams

Instance details

Methods

arbitrary :: Gen CommonOpParams

shrink :: CommonOpParams -> [CommonOpParams]

Arbitrary a => Arbitrary (SetParamsMods a)

Instance details

Methods

arbitrary :: Gen (SetParamsMods a)

shrink :: SetParamsMods a -> [SetParamsMods a]