Test.Ganeti.PartialParams

Description

Common tests for PartialParams instances

Documentation

testFillParamsLaw1 :: (PartialParams f p, Show f, Eq f) => f -> p -> Property

Checks for serialisation idempotence.

testToParamsLaw2 :: (PartialParams f p, Show f, Eq f) => f -> f -> Property

Tests that filling partial parameters satisfies the law.

testToFilledLaw3 :: (PartialParams f p, Show f, Eq f) => f -> Property

Tests that converting partial to filled parameters satisfies the law.

testToFilledMonoidLaw1 :: (PartialParams f p, Show f, Eq f, Monoid p) => f -> Property

Tests that the partial params behave correctly as a monoid action.

testToFilledMonoidLaw2 :: (PartialParams f p, Show f, Eq f, Monoid p) => f -> p -> p -> Property

Tests that the partial params behave correctly as a monoid action.