Test.Ganeti.Query.Filter

Description

Unittests for ganeti-htools.

Helpers

checkQueryResults :: ConfigData -> Query -> String -> [[ResultEntry]] -> Property

Run a query and check that we got a specific response.

makeNodeQuery :: Filter FilterField -> Query

Makes a node name query, given a filter.

expectBadQuery :: ConfigData -> Query -> String -> Property

Checks if a given operation failed.

namesToResult :: [String] -> [[ResultEntry]]

A helper to construct a list of results from an expected names list.

genClusterNames :: Int -> Int -> Gen (ConfigData, [String])

Generates a cluster and returns its node names too.

Test cases

prop_node_single_filter :: Property

Tests single node filtering: eq should return it, and (lt and gt) should fail.

prop_node_many_filter :: Property

Tests node filtering based on name equality: many OrFilter should return all results combined, many AndFilter together should return nothing. Note that we need at least 2 nodes so that the AndFilter case breaks.

prop_node_name_ordering :: Property

Tests name ordering consistency: requesting a 'simple filter' results in identical name ordering as the wanted names, requesting a more complex filter results in a niceSort-ed order.

prop_node_regex_filter :: Property

Tests node regex filtering. This is a very basic test :(

prop_node_bad_filter :: String -> Int -> Property

Tests node regex filtering. This is a very basic test :(

prop_makeSimpleFilter :: Property

Tests make simple filter.

testQuery_Filter :: Test