Test.Ganeti.Query.Instance

Description

Unittests for Instance Queries.

Documentation

createInstance :: String -> String -> AdminState -> AdminStateSource -> Instance

Creates an instance with the desired name, pnode uuid, AdminState, and AdminStateSource. All other fields are placeholders.

fakeInstanceInfo :: InstanceInfo

A fake InstanceInfo to be used to check values.

responseError :: String -> (String, ERpcError a)

Erroneous node response - the exact error does not matter.

responseSuccess :: String -> [String] -> (String, ERpcError RpcResultAllInstancesInfo)

Successful response - the error does not really matter.

waldoInstance :: Instance

The instance used for testing. Called Waldo as test cases involve trouble finding information related to it.

case_nodeOffline :: Assertion

Check that an error is thrown when the node is offline

case_nodeOnlineNoInfo :: Assertion

Check that a Right Nothing is returned when the node is online, yet no info is present anywhere in the system.

case_infoOnPrimary :: Assertion

Check the case when the info is on the primary node

case_infoOnSecondary :: Assertion

Check the case when the info is on the primary node

testQuery_Instance :: Test