Test.Ganeti.Query.Language

Description

Unittests for ganeti-htools.

Documentation

genFilter :: Gen (Filter FilterField)

Custom Filter generator (top-level), which enforces a (sane) limit on the depth of the generated filters.

genFilter' :: Int -> Gen (Filter FilterField)

Custom generator for filters that correctly halves the state of the generators at each recursive step, per the QuickCheck documentation, in order not to run out of memory.

genJSValue :: Gen JSValue

Generates an arbitrary JSValue. We do this via a function a not via arbitrary instance since that would require us to define an arbitrary for JSValue, which can be recursive, entering the usual problems with that; so we only generate the base types, not the recursive ones, and not JSNull, which we can't use in a RSNormal ResultEntry.

genResultEntry :: Gen ResultEntry

Generates a ResultEntry value.

prop_filter_serialisation :: Property

Tests that serialisation/deserialisation of filters is idempotent.

prop_filterregex_instances :: FilterRegex -> Property

Tests that filter regexes are serialised correctly.

prop_resultstatus_serialisation :: ResultStatus -> Property

Tests ResultStatus serialisation.

prop_fieldtype_serialisation :: FieldType -> Property

Tests FieldType serialisation.

prop_fielddef_serialisation :: FieldDefinition -> Property

Tests FieldDef serialisation.

prop_resultentry_serialisation :: Property

Tests ResultEntry serialisation. Needed especially as this is done manually, and not via buildObject (different serialisation format).

prop_fieldsresult_serialisation :: Property

Tests FieldDef serialisation. We use a made-up maximum limit of 20 for the generator, since otherwise the lists become too long and we don't care so much about list length but rather structure.

prop_itemtype_serialisation :: ItemType -> Property

Tests ItemType serialisation.

case_filterParsing :: Assertion

Tests basic cases of filter parsing, including legacy ones.

testQuery_Language :: Test

Orphan instances

Arbitrary ResultStatus

Instance details

Methods

arbitrary :: Gen ResultStatus

shrink :: ResultStatus -> [ResultStatus]

Arbitrary FieldType

Instance details

Methods

arbitrary :: Gen FieldType

shrink :: FieldType -> [FieldType]

Arbitrary QueryTypeOp

Instance details

Methods

arbitrary :: Gen QueryTypeOp

shrink :: QueryTypeOp -> [QueryTypeOp]

Arbitrary QueryTypeLuxi

Instance details

Methods

arbitrary :: Gen QueryTypeLuxi

shrink :: QueryTypeLuxi -> [QueryTypeLuxi]

Arbitrary FilterRegex

Instance details

Methods

arbitrary :: Gen FilterRegex

shrink :: FilterRegex -> [FilterRegex]

Arbitrary ItemType

Instance details

Methods

arbitrary :: Gen ItemType

shrink :: ItemType -> [ItemType]

Arbitrary FieldDefinition

Instance details

Methods

arbitrary :: Gen FieldDefinition

shrink :: FieldDefinition -> [FieldDefinition]

Arbitrary QueryFieldsResult

Instance details

Methods

arbitrary :: Gen QueryFieldsResult

shrink :: QueryFieldsResult -> [QueryFieldsResult]

Arbitrary (Filter FilterField)

Instance details

Methods

arbitrary :: Gen (Filter FilterField)

shrink :: Filter FilterField -> [Filter FilterField]