Test.Ganeti.Query.Network

Description

Unittests for Network Queries.

Documentation

prop_getGroupConnection :: NodeGroup -> Property

Check if looking up a valid network ID of a nodegroup yields a non-Nothing result.

prop_getGroupConnection_notFound :: NodeGroup -> String -> Property

Checks if looking up an ID of a non-existing network in a node group yields Nothing.

prop_instIsConnected :: ConfigData -> Property

Checks whether actually connected instances are identified as such.

prop_instIsConnected_notFound :: ConfigData -> String -> Property

Tests whether instances that are not connected to a network are correctly classified as such.

testQuery_Network :: Test