Test.Ganeti.Rpc

Description

Unittests for ganeti-htools.

Documentation

genStorageUnit :: Gen StorageUnit

genStorageUnits :: Gen [StorageUnit]

genHvParams :: Gen HvParams

Generate Hvparams

genHvSpecs :: Gen [(Hypervisor, HvParams)]

Generate hypervisor specifications to be used for the NodeInfo call

genConsoleInfoCallParams :: Gen [(String, InstanceConsoleInfoParams)]

Generates per-instance console info params for the InstanceConsoleInfo call.

genInstanceConsoleInfoParams :: Gen InstanceConsoleInfoParams

Generates parameters for the console info call, consisting of an instance object, node object, HvParams, and FilledBeParams.

runOfflineTest :: (Rpc a b, Eq b, Show b) => a -> Property

Monadic check that, for an offline node and a call that does not support offline nodes, we get a OfflineNodeError response.

prop_noffl_request_allinstinfo :: RpcCallAllInstancesInfo -> Property

prop_noffl_request_instconsinfo :: RpcCallInstanceConsoleInfo -> Property

prop_noffl_request_instlist :: RpcCallInstanceList -> Property

prop_noffl_request_nodeinfo :: RpcCallNodeInfo -> Property

prop_Compressed_serialisation :: Compressed -> Property

Test that the serialisation of Compressed is idempotent.

testRpc :: Test

Orphan instances

Arbitrary Compressed

Instance details

Methods

arbitrary :: Gen Compressed

shrink :: Compressed -> [Compressed]

Arbitrary RpcCallAllInstancesInfo

Instance details

Methods

arbitrary :: Gen RpcCallAllInstancesInfo

shrink :: RpcCallAllInstancesInfo -> [RpcCallAllInstancesInfo]

Arbitrary RpcCallInstanceConsoleInfo

Instance details

Methods

arbitrary :: Gen RpcCallInstanceConsoleInfo

shrink :: RpcCallInstanceConsoleInfo -> [RpcCallInstanceConsoleInfo]

Arbitrary RpcCallInstanceList

Instance details

Methods

arbitrary :: Gen RpcCallInstanceList

shrink :: RpcCallInstanceList -> [RpcCallInstanceList]

Arbitrary RpcCallNodeInfo

Instance details

Methods

arbitrary :: Gen RpcCallNodeInfo

shrink :: RpcCallNodeInfo -> [RpcCallNodeInfo]