Test.Ganeti.Storage.Lvm.LVParser

Description

Unittests for the LV Parser

Documentation

case_lvs_lv :: Assertion

Test parsing a LV lvs output.

serializeLVInfo :: LVInfo -> String

Serialize a LVInfo in the same format that is output by lvs. The "instance" field is not serialized because it's not provided by lvs so it is not part of this test.

serializeLVInfos :: [LVInfo] -> String

Serialize a list of LVInfo in the same format that is output by lvs.

prop_parse_lvs_lv :: [LVInfo] -> Property

Test if a randomly generated LV lvs output is properly parsed.

testStorage_Lvm_LVParser :: Test

Orphan instances

Arbitrary LVInfo
Arbitrary instance for LVInfo. The instance is always Nothing because it is not part of the parsed data: it is added afterwards from a different source.

Instance details

Methods

arbitrary :: Gen LVInfo

shrink :: LVInfo -> [LVInfo]