Test.Ganeti.THH

Description

Unittests for our template-haskell generated code.

Custom types

data TestObj

Type used to test optional field implementation. Equivalent to data TestObj = TestObj { tobjA :: Maybe Int, tobjB :: Maybe Int }.

Constructors

TestObj

Fields

tobjA :: (Maybe Int)

tobjB :: (Maybe Int)

loadTestObj :: JSValue -> Result TestObj

saveTestObj :: TestObj -> JSValue

prop_OptFields :: TestObj -> Property

Tests that serialising an (arbitrary) TestObj instance is correct: fully optional fields are represented in the resulting dictionary only when non-null, optional-but-required fields are always represented (with either null or an actual value).

prop_TestObj_serialization :: TestObj -> Property

Test serialization of TestObj.

prop_TestObj_deserialisationFail :: Property

Test that all superfluous keys will fail to parse.

data UnitObj

A unit-like data type.

Constructors

UnitObj

loadUnitObj :: JSValue -> Result UnitObj

saveUnitObj :: UnitObj -> JSValue

prop_UnitObj_serialization :: UnitObj -> Property

Test serialization of UnitObj.

prop_UnitObj_deserialisationFail :: Property

Test that all superfluous keys will fail to parse.

data FilledTestParams

Constructors

FilledTestParams

Fields

tparamC :: Int

tparamD :: String

data PartialTestParams

Constructors

PartialTestParams

Fields

tparamCP :: (Maybe Int)

tparamDP :: (Maybe String)

allTestParamFields :: [String]

loadPartialTestParams :: JSValue -> Result PartialTestParams

savePartialTestParams :: PartialTestParams -> JSValue

loadFilledTestParams :: JSValue -> Result FilledTestParams

saveFilledTestParams :: FilledTestParams -> JSValue

prop_fillWithPartialParams :: Property

Tests that filling partial parameters works as expected.

prop_fillPartialLaw1 :: FilledTestParams -> PartialTestParams -> Property

Tests that filling partial parameters satisfies the law.

prop_toParams :: Property

Tests that filling partial parameters works as expected.

prop_fillPartialLaw2 :: FilledTestParams -> FilledTestParams -> Property

Tests that filling partial parameters satisfies the law.

prop_fillPartialLaw3 :: FilledTestParams -> Property

Tests that filling partial parameters satisfies the law.

prop_fillPartialMonoidLaw1 :: FilledTestParams -> Property

Tests that the monoid action laws are satisfied.

prop_fillPartialMonoidLaw2 :: FilledTestParams -> PartialTestParams -> PartialTestParams -> Property

Tests that the monoid action laws are satisfied.

testTHH :: Test