Test.Ganeti.TestHTools

Description

Common functionality for htools-related unittests.

Helpers

nullISpec :: ISpec

An ISpec with 0 resources.

nullIPolicy :: IPolicy

Null iPolicy, and by null we mean very liberal.

defGroup :: Group

Default group definition.

defGroupList :: List

Default group, as a (singleton) List.

defGroupAssoc :: Map String Gdx

Default group, as a string map.

createInstance :: Int -> Int -> Int -> Instance

Create an instance given its spec.

makeSmallCluster :: Node -> Int -> List

Create a small cluster by repeating a node spec.

setInstanceSmallerThanNode :: Node -> Instance -> Instance

Update an instance to be smaller than a node.