Test.Ganeti.TestHelper

Description

Unittest helpers for TemplateHaskell components.

Documentation

propPrefix :: String

Test property prefix.

casePrefix :: String

Test case prefix.

case2Pfx :: String

Test case prefix without underscore.

simplifyName :: String -> String -> String

Tries to drop a prefix from a string.

runProp :: Testable prop => String -> prop -> Test

Builds a test from a QuickCheck property.

runCase :: String -> Assertion -> Test

Builds a test for a HUnit test case.

run :: Name -> Q Exp

Runs the correct test provider for a given test, based on its name (not very nice, but...).

mapSlashes :: String -> String

Convert slashes in a name to underscores.

testSuite :: String -> [Name] -> Q [Dec]

Builds a test suite.

mkConsArbitrary :: (Name, [a]) -> Exp

Builds an arbitrary value for a given constructor. This doesn't use the actual types of the fields, since we expect arbitrary instances for all of the types anyway, we only care about the number of fields.

conInfo :: Con -> (Name, [Type])

Extracts the name and the types from a constructor.

mkRegularArbitrary :: Name -> [Con] -> Q [Dec]

Builds an arbitrary instance for a regular data type (i.e. not Bounded).

genArbitrary :: Name -> Q [Dec]

Builds a default Arbitrary instance for a type. This requires that all members are of types that already have Arbitrary instances, and that the arbitrary instances are well behaved (w.r.t. recursive data structures, or similar concerns). In that sense, this is not appropriate for all data types, just those that are simple but very repetitive or have many simple fields.