Test.Ganeti.Types

Description

Unittests for Types.

Arbitrary instance

allDiskTemplates :: [DiskTemplate]

Valid disk templates (depending on configure options).

allStorageTypes :: [StorageType]

Valid storage types.

genReasonElem :: Gen ReasonElem

Generates one element of a reason trail

genReasonTrail :: Gen ReasonTrail

Generates a reason trail

Properties

prop_AllocPolicy_serialisation :: AllocPolicy -> Property

case_AllocPolicy_order :: Assertion

Test AllocPolicy ordering is as expected.

prop_DiskTemplate_serialisation :: DiskTemplate -> Property

prop_InstanceStatus_serialisation :: InstanceStatus -> Property

prop_NonNeg_pass :: NonNegative Int -> Property

Tests building non-negative numbers.

prop_NonNeg_fail :: Positive Int -> Property

Tests building non-negative numbers.

prop_Positive_pass :: Positive Int -> Property

Tests building positive numbers.

prop_Positive_fail :: NonNegative Int -> Property

Tests building positive numbers.

prop_Neg_pass :: Positive Int -> Property

Tests building negative numbers.

prop_Neg_fail :: NonNegative Int -> Property

Tests building negative numbers.

prop_NonEmpty_pass :: NonEmptyList String -> Property

Tests building non-empty lists.

case_NonEmpty_fail :: Assertion

Tests building positive numbers.

prop_MigrationMode_serialisation :: MigrationMode -> Property

Tests migration mode serialisation.

prop_VerifyOptionalChecks_serialisation :: VerifyOptionalChecks -> Property

Tests verify optional checks serialisation.

prop_DdmSimple_serialisation :: DdmSimple -> Property

Tests DdmSimple serialisation.

prop_DdmFull_serialisation :: DdmFull -> Property

Tests DdmFull serialisation.

prop_CVErrorCode_serialisation :: CVErrorCode -> Property

Tests CVErrorCode serialisation.

case_CVErrorCode_pyequiv :: Assertion

Tests equivalence with Python, based on Constants.hs code.

prop_Hypervisor_serialisation :: Hypervisor -> Property

Test Hypervisor serialisation.

prop_OobCommand_serialisation :: OobCommand -> Property

Test OobCommand serialisation.

prop_StorageType_serialisation :: StorageType -> Property

Test StorageType serialisation.

prop_NodeEvacMode_serialisation :: EvacMode -> Property

Test NodeEvacMode serialisation.

prop_FileDriver_serialisation :: FileDriver -> Property

Test FileDriver serialisation.

prop_InstCreateMode_serialisation :: InstCreateMode -> Property

Test InstCreate serialisation.

prop_RebootType_serialisation :: RebootType -> Property

Test RebootType serialisation.

prop_ExportMode_serialisation :: ExportMode -> Property

Test ExportMode serialisation.

prop_IAllocatorTestDir_serialisation :: IAllocatorTestDir -> Property

Test IAllocatorTestDir serialisation.

prop_IAllocatorMode_serialisation :: IAllocatorMode -> Property

Test IAllocatorMode serialisation.

case_IAllocatorMode_pyequiv :: Assertion

Tests equivalence with Python, based on Constants.hs code.

prop_NICMode_serialisation :: NICMode -> Property

Test NICMode serialisation.

prop_OpStatus_serialization :: OpStatus -> Property

Test OpStatus serialisation.

prop_JobStatus_serialization :: JobStatus -> Property

Test JobStatus serialisation.

case_JobStatus_order :: Assertion

Test JobStatus ordering is as expected.

case_NICMode_pyequiv :: Assertion

Tests equivalence with Python, based on Constants.hs code.

prop_FinalizedJobStatus_serialisation :: FinalizedJobStatus -> Property

Test FinalizedJobStatus serialisation.

case_FinalizedJobStatus_pyequiv :: Assertion

Tests equivalence with Python, based on Constants.hs code.

prop_JobId_serialisation :: JobId -> Property

Tests JobId serialisation (both from string and ints).

prop_JobId_fractional :: Property

Tests that fractional job IDs are not accepted.

case_JobId_BadTypes :: Assertion

Tests that a job ID is not parseable from "bad" JSON values.

prop_JobDependency_serialisation :: JobDependency -> Property

Test JobDependency serialisation.

prop_OpSubmitPriority_serialisation :: OpSubmitPriority -> Property

Test OpSubmitPriority serialisation.

prop_OpSubmitPriority_string :: OpSubmitPriority -> Property

Tests string formatting for OpSubmitPriority.

prop_ELogType_serialisation :: ELogType -> Property

Test ELogType serialisation.

testTypes :: Test

Orphan instances

Arbitrary ClockTime

Instance details

Methods

arbitrary :: Gen ClockTime

shrink :: ClockTime -> [ClockTime]

Arbitrary DiskTemplate
Custom Arbitrary instance for DiskTemplate, which needs to handle the case of file storage being disabled at configure time.

Instance details

Methods

arbitrary :: Gen DiskTemplate

shrink :: DiskTemplate -> [DiskTemplate]

Arbitrary TagKind

Instance details

Methods

arbitrary :: Gen TagKind

shrink :: TagKind -> [TagKind]

Arbitrary AllocPolicy

Instance details

Methods

arbitrary :: Gen AllocPolicy

shrink :: AllocPolicy -> [AllocPolicy]

Arbitrary InstanceStatus

Instance details

Methods

arbitrary :: Gen InstanceStatus

shrink :: InstanceStatus -> [InstanceStatus]

Arbitrary MigrationMode

Instance details

Methods

arbitrary :: Gen MigrationMode

shrink :: MigrationMode -> [MigrationMode]

Arbitrary VerifyOptionalChecks

Instance details

Methods

arbitrary :: Gen VerifyOptionalChecks

shrink :: VerifyOptionalChecks -> [VerifyOptionalChecks]

Arbitrary CVErrorCode

Instance details

Methods

arbitrary :: Gen CVErrorCode

shrink :: CVErrorCode -> [CVErrorCode]

Arbitrary DdmSimple

Instance details

Methods

arbitrary :: Gen DdmSimple

shrink :: DdmSimple -> [DdmSimple]

Arbitrary DdmFull

Instance details

Methods

arbitrary :: Gen DdmFull

shrink :: DdmFull -> [DdmFull]

Arbitrary Hypervisor

Instance details

Methods

arbitrary :: Gen Hypervisor

shrink :: Hypervisor -> [Hypervisor]

Arbitrary OobCommand

Instance details

Methods

arbitrary :: Gen OobCommand

shrink :: OobCommand -> [OobCommand]

Arbitrary StorageType
Custom Arbitrary instance for StorageType, which needs to handle the case of file storage being disabled at configure time.

Instance details

Methods

arbitrary :: Gen StorageType

shrink :: StorageType -> [StorageType]

Arbitrary EvacMode

Instance details

Methods

arbitrary :: Gen EvacMode

shrink :: EvacMode -> [EvacMode]

Arbitrary FileDriver

Instance details

Methods

arbitrary :: Gen FileDriver

shrink :: FileDriver -> [FileDriver]

Arbitrary InstCreateMode

Instance details

Methods

arbitrary :: Gen InstCreateMode

shrink :: InstCreateMode -> [InstCreateMode]

Arbitrary RebootType

Instance details

Methods

arbitrary :: Gen RebootType

shrink :: RebootType -> [RebootType]

Arbitrary ExportMode

Instance details

Methods

arbitrary :: Gen ExportMode

shrink :: ExportMode -> [ExportMode]

Arbitrary IAllocatorTestDir

Instance details

Methods

arbitrary :: Gen IAllocatorTestDir

shrink :: IAllocatorTestDir -> [IAllocatorTestDir]

Arbitrary IAllocatorMode

Instance details

Methods

arbitrary :: Gen IAllocatorMode

shrink :: IAllocatorMode -> [IAllocatorMode]

Arbitrary NICMode

Instance details

Methods

arbitrary :: Gen NICMode

shrink :: NICMode -> [NICMode]

Arbitrary JobStatus

Instance details

Methods

arbitrary :: Gen JobStatus

shrink :: JobStatus -> [JobStatus]

Arbitrary FinalizedJobStatus

Instance details

Methods

arbitrary :: Gen FinalizedJobStatus

shrink :: FinalizedJobStatus -> [FinalizedJobStatus]

Arbitrary JobDependency

Instance details

Methods

arbitrary :: Gen JobDependency

shrink :: JobDependency -> [JobDependency]

Arbitrary JobIdDep

Instance details

Methods

arbitrary :: Gen JobIdDep

shrink :: JobIdDep -> [JobIdDep]

Arbitrary JobId

Instance details

Methods

arbitrary :: Gen JobId

shrink :: JobId -> [JobId]

Arbitrary OpSubmitPriority

Instance details

Methods

arbitrary :: Gen OpSubmitPriority

shrink :: OpSubmitPriority -> [OpSubmitPriority]

Arbitrary OpStatus

Instance details

Methods

arbitrary :: Gen OpStatus

shrink :: OpStatus -> [OpStatus]

Arbitrary ELogType

Instance details

Methods

arbitrary :: Gen ELogType

shrink :: ELogType -> [ELogType]

Arbitrary a => Arbitrary (NonEmpty a)

Instance details

Methods

arbitrary :: Gen (NonEmpty a)

shrink :: NonEmpty a -> [NonEmpty a]

(Arbitrary a, Ord a, Num a, Show a) => Arbitrary (Negative a)

Instance details

Methods

arbitrary :: Gen (Negative a)

shrink :: Negative a -> [Negative a]

(Arbitrary a, Ord a, Num a, Show a) => Arbitrary (Positive a)

Instance details

Methods

arbitrary :: Gen (Positive a)

shrink :: Positive a -> [Positive a]

(Arbitrary a, Ord a, Num a, Show a) => Arbitrary (NonNegative a)

Instance details

Methods

arbitrary :: Gen (NonNegative a)

shrink :: NonNegative a -> [NonNegative a]