Test.Ganeti.Utils

Description

Unittests for ganeti-htools.

Documentation

genNonCommaString :: Gen String

Helper to generate a small string that doesn't contain commas.

prop_commaJoinSplit :: Property

If the list is not just an empty element, and if the elements do not contain commas, then join+split should be idempotent.

prop_commaSplitJoin :: String -> Property

Split and join should always be idempotent.

prop_findFirst :: Property

Test findFirst on several possible inputs.

prop_fromObjWithDefault :: Integer -> String -> Bool

fromObjWithDefault, we test using the Maybe monad and an integer value.

prop_if'if :: Bool -> Int -> Int -> Property

Test that functional if' behaves like the syntactic sugar if.

prop_select

Arguments

:: Int
Default result

-> [Int]
List of False values

-> [Int]
List of True values

-> Property
Test result

Test basic select functionality

prop_select_undefd

Arguments

:: [Int]
List of False values

-> NonEmptyList Int
List of True values

-> Property
Test result

Test basic select functionality with undefined default

prop_select_undefv

Arguments

:: [Int]
List of False values

-> NonEmptyList Int
List of True values

-> Property
Test result

Test basic select functionality with undefined list values

prop_parseUnit :: NonNegative Int -> Property

case_niceSort_static :: Assertion

prop_niceSort_single :: Property

Tests single-string behaviour of niceSort.

prop_niceSort_generic :: Property

Tests some generic niceSort properties. Note that the last test must add a non-digit prefix; a digit one might change ordering.

prop_niceSort_numbers :: Property

Tests that niceSorting numbers is identical to actual sorting them (in numeric form).

prop_niceSortKey_equiv :: Property

Tests that niceSort and niceSortKey are equivalent.

prop_rStripSpace :: NonEmptyList Char -> Property

Tests rStripSpace.

case_new_uuid :: Assertion

Tests that the newUUID function produces valid UUIDs.

case_new_uuid_regex :: Assertion

Tests that the newUUID function produces valid UUIDs.

prop_clockTimeToString :: Integer -> Integer -> Property

prop_timediffAdd :: ClockTime -> ClockTime -> ClockTime -> Property

Verify our work-around for ghc bug #2519. Taking diffClockTimes form Time, this test fails with an exception.

prop_chompPrefix_normal :: String -> Property

Test normal operation for chompPrefix.

Any random prefix of a string must be stripped correctly, including the empty prefix, and the whole string.

prop_chompPrefix_last :: Property

Test that chompPrefix correctly allows the last char (the separator) to be absent if the string terminates there.

prop_chompPrefix_empty_string :: Property

Test that chompPrefix on the empty string always returns Nothing for prefixes of length 2 or more.

prop_chompPrefix_nothing :: Property

Test chompPrefix returns Nothing when the prefix doesn't match.

prop_trim :: NonEmptyList Char -> Property

Tests trim.

prop_splitRecombineEithers :: [Either Int Int] -> Property

Tests splitEithers and recombineEithers.

testUtils :: Test

Orphan instances

Arbitrary ClockTime

Instance details

Methods

arbitrary :: Gen ClockTime

shrink :: ClockTime -> [ClockTime]