Ganeti.ConfigReader

Description

Implementation of configuration reader with watching support.

Documentation

type ConfigReader = IO (Result ConfigData)

A type for functions that can return the configuration when executed.

data ReloadModel

Reload model data type.

Constructors

ReloadNotify
We are using notifications

ReloadPoll Int
We are using polling

data ServerState

Server state data type.

Constructors

ServerState

Fields

reloadModel :: ReloadModel

reloadTime :: Integer

Reload time (epoch) in microseconds

reloadFStat :: FStat

maxIdlePollRounds :: Int

Maximum no-reload poll rounds before reverting to inotify.

watchInterval :: Int

Reload timeout in microseconds.

pollInterval :: Int

Ratelimit timeout in microseconds.

reloadRatelimit :: Integer

Ratelimit timeout in microseconds, as an Integer.

initialPoll :: ReloadModel

Initial poll round.

data ConfigReload

Reload status data type.

Constructors

ConfigToDate
No need to reload

ConfigReloaded
Configuration reloaded

ConfigIOError
Error during configuration reload

Configuration handling

Helper functions

moveToPolling :: String -> INotify -> FilePath -> (Result ConfigData -> IO ()) -> MVar ServerState -> IO ReloadModel

Helper function for logging transition into polling mode.

moveToNotify :: IO ReloadModel

Helper function for logging transition into inotify mode.

Configuration loading

updateConfig :: FilePath -> (Result ConfigData -> IO ()) -> IO ()

(Re)loads the configuration.

safeUpdateConfig :: FilePath -> FStat -> (Result ConfigData -> IO ()) -> IO (FStat, ConfigReload)

Wrapper over updateConfig that handles IO errors.

Watcher threads

We have three threads/functions that can mutate the server state:

the long-interval watcher (onWatcherTimer)

the polling watcher (onPollTimer)

the inotify event handler (onInotify)

All of these will mutate the server state under modifyMVar or modifyMVar_, so that server transitions are more or less atomic. The inotify handler remains active during polling mode, but checks for polling mode and doesn't do anything in this case (this check is needed even if we would unregister the event handler due to how events are serialised).

onWatcherTimer :: FilePath -> (Result ConfigData -> IO ()) -> MVar ServerState -> IO ()

Long-interval reload watcher.

This is on top of the inotify-based triggered reload.

onWatcherInner :: FilePath -> (Result ConfigData -> IO ()) -> ServerState -> IO ServerState

Inner onWatcher handler.

This mutates the server state under a modifyMVar_ call. It never changes the reload model, just does a safety reload and tried to re-establish the inotify watcher.

onPollTimer :: IO Bool -> FilePath -> (Result ConfigData -> IO ()) -> MVar ServerState -> IO ()

Short-interval (polling) reload watcher.

This is only active when we're in polling mode; it will automatically exit when it detects that the state has changed to notification.

onPollInner :: IO Bool -> FilePath -> (Result ConfigData -> IO ()) -> ServerState -> IO (ServerState, Bool)

Inner onPoll handler.

This again mutates the state under a modifyMVar call, and also returns whether the thread should continue or not.

addNotifier :: INotify -> FilePath -> (Result ConfigData -> IO ()) -> MVar ServerState -> IO Bool

Setup inotify watcher.

This tries to setup the watch descriptor; in case of any IO errors, it will return False.

onInotify :: INotify -> String -> (Result ConfigData -> IO ()) -> MVar ServerState -> Event -> IO ()

Inotify event handler.

initConfigReader :: (Result ConfigData -> IO ()) -> IO ()