Ganeti.Cpu.LoadParser

Description

procstat file parser

This module holds the definition of the parser that extracts information about the CPU load of the system from the procstat file.

Parser implementation

cpustatParser :: Parser [CPUstat]

The parser for the fragment of CPU status file containing information about the CPU load.