Ganeti.Daemon

Description

Implementation of the generic daemon functionality.

Constants

Data types

data DaemonOptions

Command line options structure.

Constructors

DaemonOptions

Fields

optShowHelp :: Bool

Just show the help

optShowVer :: Bool

Just show the program version

optShowComp :: Bool

Just show the completion info

optDaemonize :: Bool

Whether to daemonize or not

optPort :: Maybe Word16

Override for the network port

optDebug :: Bool

Enable debug messages

optNoUserChecks :: Bool

Ignore user checks

optBindAddress :: Maybe String

Listen on a custom address

optSyslogUsage :: Maybe SyslogUsage

Override for Syslog usage

optForceNode :: Bool

Ignore node checks

optNoVoting :: Bool

skip voting for master

optYesDoIt :: Bool

force dangerous options

defaultOptions :: DaemonOptions

Default values for the command line options.

type OptType = GenericOptType DaemonOptions

Abrreviation for the option type.

type CheckFn a = DaemonOptions -> IO (Either ExitCode a)

Check function type.

type PrepFn a b = DaemonOptions -> a -> IO b

Prepare function type.

type MainFn a b = DaemonOptions -> a -> b -> IO ()

Main execution function type.

Command line options

oNoDaemonize :: OptType

oDebug :: OptType

oNoUserChecks :: OptType

oPort :: Int -> OptType

oBindAddress :: OptType

oSyslogUsage :: OptType

oForceNode :: OptType

oNoVoting :: OptType

oYesDoIt :: OptType

parseArgs :: String -> [OptType] -> IO (DaemonOptions, [String])

Small wrapper over getArgs and parseOpts.

Daemon-related functions

cleanupSocket :: FilePath -> IO ()

Helper function to ensure a socket doesn't exist. Should only be called once we have locked the pid file successfully.

parseAddress

Arguments

:: DaemonOptions
Command line options

-> Int
Default port for this daemon

-> IO (Result (Family, SockAddr))

Based on the options, compute the socket address to use for the daemon.

getFQDN :: IO String

Return the full qualified host name, honoring the vcluster setup.

describeError :: String -> Maybe Handle -> Maybe FilePath -> IO a -> IO a

Run an I/O action that might throw an I/O error, under a handler that will simply annotate and re-throw the exception.

genericMain

Arguments

:: GanetiDaemon
The daemon we're running

-> [OptType]
The available options

-> CheckFn a
Check function

-> PrepFn a b
Prepare function

-> MainFn a b
Execution function

-> IO ()

Generic daemon startup.