Ganeti.DataCollectors.XenCpuLoad

Description

xentop CPU data collector

Documentation

dcName :: String

The name of this data collector.

dcVersion :: DCVersion

The version of this data collector.

dcFormatVersion :: Int

The version number for the data format of this data collector.

dcCategory :: Maybe DCCategory

The category of this data collector.

dcKind :: DCKind

The kind of this data collector.

readXentop :: IO (Result String)

Read xentop output, if this program is available.

parseXentop :: String -> Result (Map String Double)

Parse output of xentop command.

combineWithRollover :: Seq (ClockTime, Double) -> Seq (ClockTime, Double) -> Seq (ClockTime, Double)

Add a new value to a sequence of observations, taking into account counter rollovers. In case of a rollover, we drop the joining interval so that we do not have to make assumptions about the value at which is rolled over, but we do keep the right sequence, appropriately moved.

dcUpdate :: Maybe CollectorData -> IO CollectorData

Updates the given Collector data.

loadAverage :: Seq (ClockTime, Double) -> Maybe Double

From a list of timestamps and cumulative CPU data, compute the average CPU activity in vCPUs.

dcReport :: Maybe CollectorData -> IO DCReport

The data exported by the data collector, taken from the default location.