Ganeti.HTools.Backend.MonD

Description

Monitoring daemon backend

This module holds implements the querying of the monitoring daemons for dynamic utilisation data.

General definitions

data Report

The actual data types for MonD's Data Collectors.

Constructors

CPUavgloadReport CPUavgload

InstanceCpuReport (Map String Double)

data DataCollector

Type describing a data collector basic information.

Constructors

DataCollector

Fields

dName :: String

Name of the data collector

dCategory :: Maybe DCCategory

The name of the category

dMkReport :: DCReport -> Maybe Report

How to parse a monitor report

dUse :: [(Node, Report)] -> (List, List) -> Result (List, List)

How the collector reports are to be used to bring dynamic data into a cluster

Node-total CPU load average data collector

mkCpuReport :: DCReport -> Maybe Report

Parse a DCReport for the node-total CPU collector.

updateNodeCpuFromReport :: (Node, Report) -> Node

Take reports of node CPU values and update a node accordingly.

updateCpuUtilDataFromNode :: List -> Node -> List

Update the instance CPU-utilization data, asuming that each virtual CPU contributes equally to the node CPU load.

useNodeTotalCPU :: [(Node, Report)] -> (List, List) -> Result (List, List)

Update cluster data from node CPU load reports.

totalCPUCollector :: DataCollector

The node-total CPU collector.

Xen instance CPU-usage collector

mkXenCpuReport :: DCReport -> Maybe Report

Parse results of the Xen-Cpu-load data collector.

useInstanceCpuData :: [(Node, Report)] -> (List, List) -> Result (List, List)

Update cluster data based on the per-instance CPU usage reports

xenCPUCollector :: DataCollector

Collector for per-instance CPU data as observed by Xen

Collector choice

collectors :: Options -> [DataCollector]

The list of Data Collectors used by hail and hbal.

Querying infrastructure

mkReport :: DataCollector -> Maybe DCReport -> Maybe Report

Return the data from correct combination of a Data Collector and a DCReport.

type MonDData = (String, [DCReport])

MonDs Data parsed by a mock file. Representing (node name, list of reports produced by MonDs Data Collectors).

type MapMonDData = Map String [DCReport]

A map storing MonDs data.

fromFile :: DataCollector -> Node -> MapMonDData -> Maybe DCReport

Get data report for the specified Data Collector and Node from the map.

getDCCName :: Maybe DCCategory -> String

Get Category Name.

prepareUrl :: DataCollector -> Node -> URLString

Prepare url to query a single collector.

fromCurl :: DataCollector -> Node -> IO (Maybe DCReport)

Query a specified MonD for a Data Collector.

pMonDN :: JSRecord -> Result MonDData

Parse a node's JSON record.

pMonDData :: String -> Result [MonDData]

Parse MonD data file contents.

queryAMonD :: Maybe MapMonDData -> DataCollector -> Node -> IO (Maybe Report)

Query a single MonD for a single Data Collector.

queryAllMonDs :: Maybe MapMonDData -> (List, List) -> DataCollector -> WriterT All IO (List, List)

Query all MonDs for a single Data Collector. Return the updated cluster, as well as a bit inidicating wether the collector succeeded.

queryAllMonDDCs :: ClusterData -> Options -> WriterT All IO ClusterData

Query all MonDs for all Data Collector. Return the cluster enriched by dynamic data, as well as a bit indicating wether all collectors could be queried successfully.