Ganeti.HTools.Backend.Simu

Description

Parsing data from a simulated description of the cluster.

This module holds the code for parsing a cluster description.

Documentation

apolAbbrev :: String -> Result AllocPolicy

Parse a shortened policy string (for command line usage).

parseDesc :: String -> [String] -> Result (AllocPolicy, Int, Int, Int, Int, Int, Bool)

Parse the string description into nodes.

createGroup

Arguments

:: Int
The group index

-> String
The group specification

-> Result (Group, [Node])

Creates a node group with the given specifications.

parseData

Arguments

:: [String]
Cluster description in text format

-> Result ClusterData

Builds the cluster data from node/instance files.

loadData

Arguments

:: [String]
Cluster description in text format

-> IO (Result ClusterData)

Builds the cluster data from node/instance files.