Ganeti.HTools.Container

Description

Module abstracting the node and instance container implementation.

This is currently implemented on top of an IntMap, which seems to give the best performance for our workload.

Documentation

type Key = Key

Our key type.

type Container = IntMap

Our container type.

find :: Key -> Container a -> a

Locate a key in the map (must exist).

add :: Key -> a -> Container a -> Container a

Add or update one element to the map.

addTwo :: Key -> a -> Key -> a -> Container a -> Container a

Add or update two elements of the map.

nameOf :: Element a => Container a -> Key -> String

Compute the name of an element in a container.

findByName :: (Element a, Monad m) => Container a -> String -> m a

Find an element by name in a Container; this is a very slow function.