Ganeti.HTools.Program.Hbal

Description

Cluster rebalancer.

Documentation

options :: IO [OptType]

Options list and functions.

arguments :: [ArgCompletion]

The list of arguments supported by the program.

iterateDepth

Arguments

:: Bool
Whether to print moves

-> AlgorithmOptions
Algorithmic options to apply

-> Table
The starting table

-> Int
Remaining length

-> Int
Max node name len

-> Int
Max instance name len

-> [MoveJob]
Current command list

-> Score
Score at which to stop

-> IO (Table, [MoveJob])
The resulting table and commands

Start computing the solution at the given depth and recurse until we find a valid solution or we exceed the maximum depth.

main :: Options -> [String] -> IO ()

Main function.