Ganeti.HTools.Tags

Description

Tags

This module holds all the tag interpretation done by htools.

Predicates

hasStandbyTag :: Node -> Bool

Predicate of having a standby tag.

Utility functions

Migration restriction tags

getMigRestrictions :: [String] -> [String] -> Set String

Given the cluster tags extract the migration restrictions from a node tag.

getRecvMigRestrictions :: [String] -> [String] -> Set String

Given the cluster tags, extract the set of migration restrictions a node is able to receive from its node tags.

Location tags

getLocations :: [String] -> [String] -> Set String

Given the cluster tags, extract the node location tags from the node tags.