Ganeti.Hs2Py.GenOpCodes

Description

GenOpCodes handles Python opcode generation.

GenOpCodes contains the helper functions that generate the Python opcodes as strings from the Haskell opcode description.

Documentation

showPyClasses :: String

Generates all opcodes as Python classes.