Ganeti.Locking.Allocation

Description

Implementation of lock allocation.

Documentation

data OwnerState

Data type describing the way a lock can be owned.

Constructors

OwnShared

OwnExclusive

type IndirectOwners a b = Map (a, b) OwnerState

Type describing indirect ownership on a lock. We keep the set of all (lock, owner)-pairs for locks that are implied in the given lock, annotated with the type of ownership (shared or exclusive).

data AllocationState a b

The state of a lock that is taken. Besides the state of the lock itself, we also keep track of all other lock allocation that affect the given lock by means of implication.

Constructors

Exclusive b (IndirectOwners a b)

Shared (Set b) (IndirectOwners a b)

indirectOwners :: (Ord a, Ord b) => Map (a, b) OwnerState -> Set b

Compute the set of indirect owners from the information about indirect ownership.

indirectExclusives :: (Ord a, Ord b) => Map (a, b) OwnerState -> Set b

Compute the (zero or one-elment) set of exclusive indirect owners.

data LockAllocation a b

Representation of a Lock allocation

To keep queries for locks efficient, we keep two associations, with the invariant that they fit together: the association from locks to their allocation state, and the association from an owner to the set of locks owned. As we do not export the constructor, the problem of keeping this invariant reduces to only exporting functions that keep the invariant.

Constructors

LockAllocation

Fields

laLocks :: Map a (AllocationState a b)

laOwned :: Map b (Map a OwnerState)

emptyAllocation :: (Ord a, Ord b) => LockAllocation a b

A state with all locks being free.

lockOwners :: Ord b => LockAllocation a b -> [b]

Obtain the list of all owners holding at least a single lock.

listLocks :: Ord b => b -> LockAllocation a b -> Map a OwnerState

Obtain the locks held by a given owner. The locks are reported as a map from the owned locks to the form of ownership (OwnShared or OwnExclusive).

listAllLocks :: Ord b => LockAllocation a b -> [a]

List all locks currently (directly or indirectly) owned by someone.

toOwnersList :: AllocationState a b -> [(b, OwnerState)]

Map an AllocationState to a list of pairs of owners and type of ownership, showing the direct owners only.

listAllLocksOwners :: LockAllocation a b -> [(a, [(b, OwnerState)])]

List all locks currently (directly of indirectly) in use together with the direct owners.

holdsLock :: (Ord a, Ord b) => b -> a -> OwnerState -> LockAllocation a b -> Bool

Returns True if the given owner holds the given lock at the given ownership level or higher. This means that querying for a shared lock returns True of the owner holds the lock in shared or exlusive mode.

data LockRequest a

Data Type describing a change request on a single lock.

Constructors

LockRequest

Fields

lockAffected :: a

lockRequestType :: Maybe OwnerState

requestExclusive :: a -> LockRequest a

Lock request for an exclusive lock.

requestShared :: a -> LockRequest a

Lock request for a shared lock.

requestRelease :: a -> LockRequest a

Request to release a lock.

updateAllocState :: (Ord a, Ord b) => (Maybe (AllocationState a b) -> AllocationState a b) -> LockAllocation a b -> a -> LockAllocation a b

Update the Allocation state of a lock according to a given function.

updateLock :: (Ord a, Ord b) => b -> LockAllocation a b -> LockRequest a -> LockAllocation a b

Internal function to update the state according to a single lock request, assuming all prerequisites are met.

updateIndirectSet :: (Ord a, Ord b) => (IndirectOwners a b -> IndirectOwners a b) -> LockAllocation a b -> a -> LockAllocation a b

Update the set of indirect ownerships of a lock by the given function.

updateIndirects :: (Lock a, Ord b) => b -> LockAllocation a b -> LockRequest a -> LockAllocation a b

Update all indirect onwerships of a given lock.

updateLocks :: (Lock a, Ord b) => b -> [LockRequest a] -> LockAllocation a b -> (LockAllocation a b, Result (Set b))

Update the locks of an owner according to the given request. Return the pair of the new state and the result of the operation, which is the the set of owners on which the operation was blocked on. so an empty set is success, and the state is updated if, and only if, the returned set is emtpy. In that way, it can be used in atomicModifyIORef.

manipulateLocksPredicate :: (Lock a, Ord b) => (a -> LockRequest a) -> (a -> Bool) -> b -> LockAllocation a b -> LockAllocation a b

Manipluate all locks of the owner with a given property.

freeLocksPredicate :: (Lock a, Ord b) => (a -> Bool) -> LockAllocation a b -> b -> LockAllocation a b

Compute the state after an owner releases all its locks that satisfy a certain property.

freeLocks :: (Lock a, Ord b) => LockAllocation a b -> b -> LockAllocation a b

Compute the state after an onwer releases all its locks.

readLockOwnerstate :: JSON a => JSValue -> Result (a, OwnerState)

Read a lock-ownerstate pair from JSON.

readOwnerLock :: (JSON a, JSON b) => JSValue -> Result (b, [(a, OwnerState)])

Read an owner-lock pair from JSON.

toRequest :: (a, OwnerState) -> LockRequest a

Transform a lock-ownerstate pair into a LockRequest.

allocationFromOwners :: (Lock a, Ord b, Show b) => [(b, [(a, OwnerState)])] -> Result (LockAllocation a b)

Obtain a LockAllocation from a given owner-locks list. The obtained allocation is the one obtained if the respective owners requested their locks sequentially.