Ganeti.Query.Instance

Description

Implementation of the Ganeti Query2 instance queries.

Documentation

type Runtime = Either RpcError LiveInfo

Runtime containing the LiveInfo. See the genericQuery function in the Query.hs file for an explanation of the terms used.

fieldsMap :: FieldMap Instance Runtime

The instance fields map.

instanceAliases :: [(FieldName, FieldName)]

The instance aliases.

instanceFields :: FieldList Instance Runtime

The instance fields.

Helper functions for node property retrieval

Various helper functions for property retrieval

Live fields functionality

Functionality related to status and operational status extraction

Helper functions extracting information as necessary for the generic query

getInstanceInfo :: [(String, ERpcError RpcResultAllInstancesInfo)] -> Instance -> ERpcError (Maybe (InstanceInfo, Bool))

Retrieves the instance information if it is present anywhere in the all instances RPC result. Notes if it originates from the primary node. An error is delivered if there is no result, and the primary node is down.

collectLiveData

Arguments

:: Bool
Live queries allowed

-> ConfigData
The cluster config

-> [String]
The requested fields

-> [Instance]
The instance objects

-> IO [(Instance, Runtime)]

Collect live data from RPC query if enabled.