Ganeti.Query.Locks

Description

Implementation of Ganeti Lock field queries

The actual computation of the field values is done by forwarding the request; so only have a minimal field definition here.

Documentation

type RuntimeData = ([(ClientId, OwnerState)], [(ClientId, OwnerState)])

The runtime information for locks. As all information about locks is handled by WConfD, the actual information is obtained as live data. The type represents the information for a single lock, even though all locks are queried simultaneously, ahead of time.

getOwners :: RuntimeData -> a -> ResultEntry

Obtain the owners of a lock from the runtime data.

getMode :: RuntimeData -> a -> ResultEntry

Obtain the mode of a lock from the runtime data.

getPending :: RuntimeData -> a -> ResultEntry

Obtain the pending requests from the runtime data.

lockFields :: FieldList String RuntimeData

List of all lock fields.

fieldsMap :: FieldMap String RuntimeData

The lock fields map.