Ganeti.Query.Network

Description

Implementation of the Ganeti Query2 node group queries.

Documentation

fieldsMap :: FieldMap Network NoDataRuntime

The group fields map.

getGroupConnection :: String -> NodeGroup -> Maybe (String, String, String, String)

Given a network's UUID and a node group, this function assembles a tuple of the group's name, the mode and the link by which the network is connected to the group. Returns Nothing if the network is not connected to the group.

instIsConnected :: String -> Instance -> Bool

Helper function that checks if an instance is linked to the given network.

getNetworkUuid :: ConfigData -> String -> Maybe String

Helper function to look up a network's UUID by its name