Ganeti.Query.Query

Description

Implementation of the Ganeti Query2 functionality.

Helper functions

Main query execution

nameField :: ItemType -> FilterField

Computes the name field for different query types.

uuidField :: ItemType -> FilterField

Computes the uuid field, or the best possible substitute, for different query types.

getRequestedNames :: Query -> [String]

Checks that we have either requested a valid set of names, or we have a more complex filter.

query

Arguments

:: ConfigData
The current configuration

-> Bool
Whether to collect live data

-> Query
The query (item, fields, filter)

-> IO (ErrorResult QueryResult)
Result

Main query execution function.

queryFields :: QueryFields -> ErrorResult QueryFieldsResult

Query fields call.

queryCompat :: QueryResult -> ErrorResult [[JSValue]]

Classic query converter. It gets a standard query result on input and computes the classic style results.