Ganeti.Storage.Lvm.LVParser

Description

Logical Volumer information parser

This module holds the definition of the parser that extracts status information about the logical volumes (LVs) of the system from the output of the lvs command.

Documentation

lvsSeparator :: Char

The separator of the fields returned by lvs

Utility functions

skipSpaces :: Parser ()

Our own space-skipping function, because A.skipSpace also skips newline characters. It skips ZERO or more spaces, so it does not fail if there are no spaces.

bytesP :: Parser Int

A parser recognizing a number of bytes, represented as a number preceeded by a separator and followed by the B character.

intP :: Parser Int

A parser recognizing a number discarding the preceeding separator

stringP :: Parser String

A parser recognizing a string starting with and closed by a separator (both are discarded)

Parser implementation

lvCommand :: String

The command providing the data, in the format the parser expects

lvParams :: [String]

The parameters for getting the data in the format the parser expects

oneLvParser :: Parser LVInfo

The parser for one line of the diskstatus file.

lvParser :: Parser [LVInfo]

The parser for a whole diskstatus file.