Ganeti.WConfd.Monad

Description

All RPC calls are run within this monad.

It encapsulates:

IO operations,

failures,

working with the daemon state.

Code that is specific either to the configuration or the lock management, should go into their corresponding dedicated modules.

Monoid of the locations to flush the configuration to

data DistributionTarget

Data type describing where the configuration has to be distributed to.

Constructors

Everywhere

ToGroups (Set String)

Pure data types used in the monad

data DaemonState

The state of the daemon, capturing both the configuration state and the locking state.

Constructors

DaemonState

Fields

dsConfigState :: ConfigState

dsLockWaiting :: GanetiLockWaiting

dsTempRes :: TempResState

dsTempResL :: Lens' DaemonState TempResState

dsLockWaitingL :: Lens' DaemonState GanetiLockWaiting

dsConfigStateL :: Lens' DaemonState ConfigState

data DaemonHandle

Constructors

DaemonHandle

Fields

dhDaemonState :: IORef DaemonState

The current state of the daemon

dhConfigPath :: FilePath

The configuration file path all static information that doesn't change during the life-time of the daemon should go here; all IDs of threads that do asynchronous work should probably also go here

dhSaveConfigWorker :: AsyncWorker (Any, DistributionTarget) ()

dhSaveLocksWorker :: AsyncWorker () ()

dhSaveTempResWorker :: AsyncWorker () ()

dhLivelock :: Livelock

mkDaemonHandle

Arguments

:: FilePath

-> ConfigState

-> GanetiLockWaiting

-> TempResState

-> (IO ConfigState -> [AsyncWorker DistributionTarget ()] -> ResultG (AsyncWorker (Any, DistributionTarget) ()))
A function that creates a worker that asynchronously saves the configuration to the master file.

-> (IO ConfigState -> ResultG (AsyncWorker DistributionTarget ()))
A function that creates a worker that asynchronously distributes the configuration to master candidates

-> (IO ConfigState -> ResultG (AsyncWorker DistributionTarget ()))
A function that creates a worker that asynchronously distributes SSConf to nodes

-> (IO GanetiLockWaiting -> ResultG (AsyncWorker () ()))
A function that creates a worker that asynchronously saves the lock allocation state.

-> (IO TempResState -> ResultG (AsyncWorker () ()))
A function that creates a worker that asynchronously saves the temporary reservations state.

-> Livelock

-> ResultG DaemonHandle

The monad and its instances

type WConfdMonadIntType = ReaderT DaemonHandle IO

A type alias for easier referring to the actual content of the monad when implementing its instances.

newtype WConfdMonadInt a

The internal part of the monad without error handling.

Constructors

WConfdMonadInt

Fields

getWConfdMonadInt :: WConfdMonadIntType a

runWConfdMonadInt :: WConfdMonadInt a -> DaemonHandle -> IO a

Runs the internal part of the WConfdMonad monad on a given daemon handle.

type WConfdMonad = ResultT GanetiException WConfdMonadInt

The complete monad with error handling.

type AtomicModifyMonad a = ResultT GanetiException WriterLog a

A pure monad that logs and reports errors used for atomic modifications.

Basic functions in the monad

daemonHandle :: WConfdMonad DaemonHandle

Returns the daemon handle.

readConfigState :: WConfdMonad ConfigState

Returns the current configuration, given a handle

unpackConfigResult :: ClockTime -> ConfigState -> (a, ConfigState) -> ((a, Bool, Bool), ConfigState)

From a result of a configuration change, determine if the configuration was changed and if full distribution is needed. If so, also bump the serial number.

modifyConfigStateErrWithImmediate :: (TempResState -> ConfigState -> AtomicModifyMonad (a, ConfigState)) -> WConfdMonad () -> WConfdMonad a

Atomically modifies the configuration state in the WConfdMonad with a computation that can possibly fail; immediately afterwards, while config write is still going on, do the followup action. Return only after replication is finished.

modifyConfigStateErr :: (TempResState -> ConfigState -> AtomicModifyMonad (a, ConfigState)) -> WConfdMonad a

Atomically modifies the configuration state in the WConfdMonad with a computation that can possibly fail.

modifyConfigStateErr_ :: (TempResState -> ConfigState -> AtomicModifyMonad ConfigState) -> WConfdMonad ()

Atomically modifies the configuration state in the WConfdMonad with a computation that can possibly fail.

modifyConfigState :: (ConfigState -> (a, ConfigState)) -> WConfdMonad a

Atomically modifies the configuration state in the WConfdMonad.

modifyConfigStateWithImmediate :: (ConfigState -> (a, ConfigState)) -> WConfdMonad () -> WConfdMonad a

Atomically modifies the configuration state in WConfdMonad; immediately afterwards (while the config write-out is not necessarily finished) do another acation.

forceConfigStateDistribution :: DistributionTarget -> WConfdMonad ()

Force the distribution of configuration without actually modifying it.

We need a separate call for this operation, because modifyConfigState only triggers the distribution when the configuration changes.

modifyConfigDataErr_ :: (TempResState -> ConfigData -> AtomicModifyMonad ConfigData) -> WConfdMonad ()

Atomically modifies the configuration data in the WConfdMonad with a computation that can possibly fail.

modifyTempResStateErr :: (ConfigData -> StateT TempResState ErrorResult a) -> WConfdMonad a

Atomically modifies the state of temporary reservations in WConfdMonad in the presence of possible errors.

modifyTempResState :: (ConfigData -> State TempResState a) -> WConfdMonad a

Atomically modifies the state of temporary reservations in WConfdMonad.

readTempResState :: WConfdMonad (ConfigData, TempResState)

Reads the state of of the configuration and temporary reservations in WConfdMonad.

modifyLockWaiting :: (GanetiLockWaiting -> (GanetiLockWaiting, (a, Set ClientId))) -> WConfdMonad a

Atomically modifies the lock waiting state in WConfdMonad.

modifyLockWaiting_ :: (GanetiLockWaiting -> (GanetiLockWaiting, Set ClientId)) -> WConfdMonad ()

Atomically modifies the lock allocation state in WConfdMonad, not producing any result

readLockWaiting :: WConfdMonad GanetiLockWaiting

Read the lock waiting state in WConfdMonad.

readLockAllocation :: WConfdMonad (LockAllocation GanetiLocks ClientId)

Read the underlying lock allocation.

modifyConfigAndReturnWithLock :: (TempResState -> ConfigState -> AtomicModifyMonad (a, ConfigState)) -> State TempResState () -> WConfdMonad (Maybe a)

Modify the configuration while temporarily acquiring the configuration lock. If the configuration lock is held by someone else, nothing is changed and Nothing is returned.

modifyConfigWithLock :: (TempResState -> ConfigState -> AtomicModifyMonad ConfigState) -> State TempResState () -> WConfdMonad (Maybe ())