Test.Ganeti.HTools.Instance

Description

Unittests for ganeti-htools.

Arbitrary instances

genInstanceSmallerThanNode :: Node -> Gen Instance

Generates an instance smaller than a node.

genInstanceMaybeBiggerThanNode :: Node -> Gen Instance

Generates an instance possibly bigger than a node. In any case, that instance will be bigger than the node's ipolicy's lower bound.

genInstanceOnNodeList :: List -> Gen Instance

Generates an instance with nodes on a node list. The following rules are respected: 1. The instance is never bigger than its primary node 2. If possible the instance has different pnode and snode 3. Else disk templates which require secondary nodes are disabled

genInstanceList :: Gen Instance -> Gen List

Generates an instance list given an instance generator.

Test cases

testHTools_Instance :: Test

Orphan instances

Arbitrary Instance

Instance details

Methods

arbitrary :: Gen Instance

shrink :: Instance -> [Instance]