Test.Ganeti.HTools.Node

Description

Unittests for ganeti-htools.

Arbitrary instances

genNode

Arguments

:: Maybe Int
Minimum node size in terms of units

-> Maybe Int
Maximum node size (when Nothing, bounded just by the max... constants)

-> Gen Node

Generates an arbitrary node based on sizing information.

genOnlineNode :: Gen Node

Helper function to generate a sane node.

genEmptyOnlineNode :: Gen Node

Helper function to generate a sane empty node with consistent internal data.

genNodeList :: Gen Node -> Gen List

Node list generator. Given a node generator, create a random length node list. Note that "real" clusters always have at least one node, so we don't generate empty node lists here.

genUniqueNodeList :: Gen Node -> Gen (List, NameAssoc)

Node list generator where node names are unique

Test cases

testHTools_Node :: Test

List of tests for the Node module.

Orphan instances

Arbitrary Node

Instance details

Methods

arbitrary :: Gen Node

shrink :: Node -> [Node]