Test.Ganeti.Locking.Allocation

Description

Tests for lock allocation.

Documentation

data TestOwner

data TestLock

requestSucceeded :: Ord a => Map a OwnerState -> LockRequest a -> Bool

Verify that a given request is statisfied in list of owned locks

testLocking_Allocation :: Test

Orphan instances

Arbitrary OwnerState

Instance details

Methods

arbitrary :: Gen OwnerState

shrink :: OwnerState -> [OwnerState]

Arbitrary a => Arbitrary (LockRequest a)

Instance details

Methods

arbitrary :: Gen (LockRequest a)

shrink :: LockRequest a -> [LockRequest a]

(Arbitrary a, Lock a, Arbitrary b, Ord b, Show b) => Arbitrary (LockAllocation a b)

Instance details

Methods

arbitrary :: Gen (LockAllocation a b)

shrink :: LockAllocation a b -> [LockAllocation a b]