Test.Ganeti.Objects

Description

Unittests for ganeti-htools.

Arbitrary instances

genInstWithNets :: [String] -> Gen Instance

Generates an instance that is connected to the given networks and possibly some other networks

genInst :: Gen Instance

Generates an instance that is connected to some networks

enhanceInstWithNets :: Instance -> [String] -> Gen Instance

Enhances a given instance with network information, by connecting it to the given networks and possibly some other networks

genDiskWithChildren :: Int -> Gen Disk

genDisk :: Gen Disk

genValidNetwork :: Gen Network

Generates a network instance with minimum netmasks of /24. Generating bigger networks slows down the tests, because long bit strings are generated for the reservations.

genBitString :: Int -> Gen AddressPool

Generate an arbitrary string consisting of '0' and '1' of the given length.

genBitStringMaxLen :: Int -> Gen AddressPool

Generate an arbitrary string consisting of '0' and '1' of the maximum given length.

genEmptyCluster :: Int -> Gen ConfigData

Generator for config data with an empty cluster (no instances), with N defined nodes.

genConfigDataWithNetworks :: ConfigData -> Gen ConfigData

FIXME: make an even simpler base version of creating a cluster.

Generates config data with a couple of networks.

Test properties

prop_fillDict :: [(Int, Int)] -> [(Int, Int)] -> Property

Tests that fillDict behaves correctly

prop_LogicalVolume_serialisation :: LogicalVolume -> Property

prop_LogicalVolume_deserialisationFail :: Property

prop_Disk_serialisation :: Disk -> Property

Test that the serialisation of DiskLogicalId, which is implemented manually, is idempotent. Since we don't have a standalone JSON instance for DiskLogicalId (it's a data type that expands over two fields in a JSObject), we test this by actially testing entire Disk serialisations. So this tests two things at once, basically.

prop_Disk_array_serialisation :: Disk -> Property

prop_Node_serialisation :: Node -> Property

Check that node serialisation is idempotent.

prop_Inst_serialisation :: Instance -> Property

Check that instance serialisation is idempotent.

prop_AddressPool_serialisation :: AddressPool -> Property

Check that address pool serialisation is idempotent.

prop_Network_serialisation :: Network -> Property

Check that network serialisation is idempotent.

prop_FilterAction_serialisation :: FilterAction -> Property

Check that filter action serialisation is idempotent.

prop_FilterPredicate_serialisation :: FilterPredicate -> Property

Check that filter predicate serialisation is idempotent.

prop_Config_serialisation :: Property

Check config serialisation.

casePyCompatNetworks :: Assertion

Custom HUnit test to check the correspondence between Haskell-generated networks and their Python decoded, validated and re-encoded version. For the technical background of this unit test, check the documentation of "case_py_compat_types" of testhsTestGanetiOpcodes.hs

getNetworkProperties :: Network -> (Int, Int, Network)

Creates a tuple of the given network combined with some of its properties to be compared against the same properties generated by the python code.

casePyCompatNodegroups :: Assertion

Tests the compatibility between Haskell-serialized node groups and their python-decoded and encoded version.

genNodeGroup :: Gen NodeGroup

Generates a node group with up to 3 networks. | FIXME: This generates still somewhat completely random data, without normal validation rules.

prop_ip4AddressAsNum :: Ip4Address -> Property

Tests conversions of ip addresses from/to numbers.

prop_ip4AddressToNumber :: Word32 -> Property

Tests that the number produced by ip4AddressToNumber has the correct order of bytes.

caseNextIp4Address :: Assertion

Tests a few simple cases of IPv4 next address.

casePyCompatInstances :: Assertion

Tests the compatibility between Haskell-serialized instances and their python-decoded and encoded version. Note: this can be enhanced with logical validations on the decoded objects

mkLIDPlain :: String -> String -> DiskLogicalId

A helper function for creating LIDPlain values.

caseIncludeLogicalIdPlain :: Assertion

Tests that the logical ID is correctly found in a plain disk

caseIncludeLogicalIdDrbd :: Assertion

Tests that the logical ID is correctly found in a DRBD disk

caseNotIncludeLogicalIdPlain :: Assertion

Tests that the logical ID is correctly NOT found in a plain disk

testObjects :: Test

Orphan instances

IsString Ip4Address
IsString instance for Ip4Address, to help write the tests.

Instance details

Methods

fromString :: String -> Ip4Address

Arbitrary ByteString

Instance details

Methods

arbitrary :: Gen ByteString

shrink :: ByteString -> [ByteString]

Arbitrary DiskMode

Instance details

Methods

arbitrary :: Gen DiskMode

shrink :: DiskMode -> [DiskMode]

Arbitrary BlockDriver

Instance details

Methods

arbitrary :: Gen BlockDriver

shrink :: BlockDriver -> [BlockDriver]

Arbitrary AdminState

Instance details

Methods

arbitrary :: Gen AdminState

shrink :: AdminState -> [AdminState]

Arbitrary AdminStateSource

Instance details

Methods

arbitrary :: Gen AdminStateSource

shrink :: AdminStateSource -> [AdminStateSource]

Arbitrary SshKeyType

Instance details

Methods

arbitrary :: Gen SshKeyType

shrink :: SshKeyType -> [SshKeyType]

Arbitrary OsParams

Instance details

Methods

arbitrary :: Gen OsParams

shrink :: OsParams -> [OsParams]

Arbitrary PartialNicParams

Instance details

Methods

arbitrary :: Gen PartialNicParams

shrink :: PartialNicParams -> [PartialNicParams]

Arbitrary FilledNicParams

Instance details

Methods

arbitrary :: Gen FilledNicParams

shrink :: FilledNicParams -> [FilledNicParams]

Arbitrary PartialNic

Instance details

Methods

arbitrary :: Gen PartialNic

shrink :: PartialNic -> [PartialNic]

Arbitrary PartialBeParams

Instance details

Methods

arbitrary :: Gen PartialBeParams

shrink :: PartialBeParams -> [PartialBeParams]

Arbitrary FilledBeParams

Instance details

Methods

arbitrary :: Gen FilledBeParams

shrink :: FilledBeParams -> [FilledBeParams]

Arbitrary Instance

Instance details

Methods

arbitrary :: Gen Instance

shrink :: Instance -> [Instance]

Arbitrary ForthcomingInstanceData

Instance details

Methods

arbitrary :: Gen ForthcomingInstanceData

shrink :: ForthcomingInstanceData -> [ForthcomingInstanceData]

Arbitrary RealInstanceData

Instance details

Methods

arbitrary :: Gen RealInstanceData

shrink :: RealInstanceData -> [RealInstanceData]

Arbitrary DiskLogicalId

Instance details

Methods

arbitrary :: Gen DiskLogicalId

shrink :: DiskLogicalId -> [DiskLogicalId]

Arbitrary LogicalVolume

Instance details

Methods

arbitrary :: Gen LogicalVolume

shrink :: LogicalVolume -> [LogicalVolume]

Arbitrary Disk
Disk arbitrary instance. Since we don't test disk hierarchy properties, we only generate disks with no children (FIXME), as generating recursive datastructures is a bit more work.

Instance details

Methods

arbitrary :: Gen Disk

shrink :: Disk -> [Disk]

Arbitrary AddressPool

Instance details

Methods

arbitrary :: Gen AddressPool

shrink :: AddressPool -> [AddressPool]

Arbitrary Ip4Network

Instance details

Methods

arbitrary :: Gen Ip4Network

shrink :: Ip4Network -> [Ip4Network]

Arbitrary Ip4Address

Instance details

Methods

arbitrary :: Gen Ip4Address

shrink :: Ip4Address -> [Ip4Address]

Arbitrary Network

Instance details

Methods

arbitrary :: Gen Network

shrink :: Network -> [Network]

Arbitrary PartialISpecParams
FIXME: This generates completely random data, without normal validation rules.

Instance details

Methods

arbitrary :: Gen PartialISpecParams

shrink :: PartialISpecParams -> [PartialISpecParams]

Arbitrary FilledISpecParams

Instance details

Methods

arbitrary :: Gen FilledISpecParams

shrink :: FilledISpecParams -> [FilledISpecParams]

Arbitrary MinMaxISpecs

Instance details

Methods

arbitrary :: Gen MinMaxISpecs

shrink :: MinMaxISpecs -> [MinMaxISpecs]

Arbitrary PartialIPolicy
FIXME: This generates completely random data, without normal validation rules.

Instance details

Methods

arbitrary :: Gen PartialIPolicy

shrink :: PartialIPolicy -> [PartialIPolicy]

Arbitrary FilledIPolicy

Instance details

Methods

arbitrary :: Gen FilledIPolicy

shrink :: FilledIPolicy -> [FilledIPolicy]

Arbitrary PartialNDParams

Instance details

Methods

arbitrary :: Gen PartialNDParams

shrink :: PartialNDParams -> [PartialNDParams]

Arbitrary FilledNDParams

Instance details

Methods

arbitrary :: Gen FilledNDParams

shrink :: FilledNDParams -> [FilledNDParams]

Arbitrary GroupDiskParams
No real arbitrary instance for GroupDiskParams yet.

Instance details

Methods

arbitrary :: Gen GroupDiskParams

shrink :: GroupDiskParams -> [GroupDiskParams]

Arbitrary Node

Instance details

Methods

arbitrary :: Gen Node

shrink :: Node -> [Node]

Arbitrary FilterPredicate

Instance details

Methods

arbitrary :: Gen FilterPredicate

shrink :: FilterPredicate -> [FilterPredicate]

Arbitrary FilterAction

Instance details

Methods

arbitrary :: Gen FilterAction

shrink :: FilterAction -> [FilterAction]

Arbitrary NodeGroup

Instance details

Methods

arbitrary :: Gen NodeGroup

shrink :: NodeGroup -> [NodeGroup]

Arbitrary FilterRule

Instance details

Methods

arbitrary :: Gen FilterRule

shrink :: FilterRule -> [FilterRule]

Arbitrary IpFamily

Instance details

Methods

arbitrary :: Gen IpFamily

shrink :: IpFamily -> [IpFamily]

Arbitrary IAllocatorParams

Instance details

Methods

arbitrary :: Gen IAllocatorParams

shrink :: IAllocatorParams -> [IAllocatorParams]

Arbitrary ClusterNicParams

Instance details

Methods

arbitrary :: Gen ClusterNicParams

shrink :: ClusterNicParams -> [ClusterNicParams]

Arbitrary ClusterOsParamsPrivate

Instance details

Methods

arbitrary :: Gen ClusterOsParamsPrivate

shrink :: ClusterOsParamsPrivate -> [ClusterOsParamsPrivate]

Arbitrary ClusterOsParams

Instance details

Methods

arbitrary :: Gen ClusterOsParams

shrink :: ClusterOsParams -> [ClusterOsParams]

Arbitrary ClusterBeParams

Instance details

Methods

arbitrary :: Gen ClusterBeParams

shrink :: ClusterBeParams -> [ClusterBeParams]

Arbitrary OsHvParams
No real arbitrary instance for OsHvParams yet.

Instance details

Methods

arbitrary :: Gen OsHvParams

shrink :: OsHvParams -> [OsHvParams]

Arbitrary ClusterHvParams
No real arbitrary instance for ClusterHvParams yet.

Instance details

Methods

arbitrary :: Gen ClusterHvParams

shrink :: ClusterHvParams -> [ClusterHvParams]

Arbitrary Cluster

Instance details

Methods

arbitrary :: Gen Cluster

shrink :: Cluster -> [Cluster]

Arbitrary ConfigData

Instance details

Methods

arbitrary :: Gen ConfigData

shrink :: ConfigData -> [ConfigData]

Arbitrary (Container DataCollectorConfig)

Instance details

Methods

arbitrary :: Gen (Container DataCollectorConfig)

shrink :: Container DataCollectorConfig -> [Container DataCollectorConfig]

Arbitrary a => Arbitrary (Private a)

Instance details

Methods

arbitrary :: Gen (Private a)

shrink :: Private a -> [Private a]