Test.Ganeti.Query.Language

Description

Unittests for ganeti-htools.

Documentation

genFilter :: Gen (Filter FilterField)

Custom Filter generator (top-level), which enforces a (sane) limit on the depth of the generated filters.

genJSValue :: Gen JSValue

Generates an arbitrary JSValue. We do this via a function a not via arbitrary instance since that would require us to define an arbitrary for JSValue, which can be recursive, entering the usual problems with that; so we only generate the base types, not the recursive ones, and not JSNull, which we can't use in a RSNormal ResultEntry.

testQuery_Language :: Test

Orphan instances

Arbitrary ResultStatus

Instance details

Methods

arbitrary :: Gen ResultStatus

shrink :: ResultStatus -> [ResultStatus]

Arbitrary FieldType

Instance details

Methods

arbitrary :: Gen FieldType

shrink :: FieldType -> [FieldType]

Arbitrary QueryTypeOp

Instance details

Methods

arbitrary :: Gen QueryTypeOp

shrink :: QueryTypeOp -> [QueryTypeOp]

Arbitrary QueryTypeLuxi

Instance details

Methods

arbitrary :: Gen QueryTypeLuxi

shrink :: QueryTypeLuxi -> [QueryTypeLuxi]

Arbitrary FilterRegex

Instance details

Methods

arbitrary :: Gen FilterRegex

shrink :: FilterRegex -> [FilterRegex]

Arbitrary ItemType

Instance details

Methods

arbitrary :: Gen ItemType

shrink :: ItemType -> [ItemType]

Arbitrary FieldDefinition

Instance details

Methods

arbitrary :: Gen FieldDefinition

shrink :: FieldDefinition -> [FieldDefinition]

Arbitrary QueryFieldsResult

Instance details

Methods

arbitrary :: Gen QueryFieldsResult

shrink :: QueryFieldsResult -> [QueryFieldsResult]

Arbitrary (Filter FilterField)

Instance details

Methods

arbitrary :: Gen (Filter FilterField)

shrink :: Filter FilterField -> [Filter FilterField]