Test.Ganeti.SlotMap

Description

Unittests for the SlotMap.

Documentation

genSlotLimit :: Gen Int

Generates a number typical for the limit of a Slot. Useful for constructing resource bounds when not directly constructing the relevant Slots.

genTestKey :: Gen TestKey

Generates short strings used as SlotMap keys.

We limit ourselves to a small set of key strings with high probability to increase the chance that SlotMaps actually have more than one slot taken.

overfullKeys :: Ord a => SlotMap a -> Set a

Tells which keys of a SlotMap are overfull.

Test cases

testSlotMap :: Test

Orphan instances

Arbitrary Slot

Instance details

Methods

arbitrary :: Gen Slot

shrink :: Slot -> [Slot]