Test.Ganeti.Types

Description

Unittests for Types.

Arbitrary instance

allDiskTemplates :: [DiskTemplate]

Valid disk templates (depending on configure options).

genReasonTrail :: Gen ReasonTrail

Generates a reason trail

Properties

testTypes :: Test

Orphan instances

Arbitrary ClockTime

Instance details

Methods

arbitrary :: Gen ClockTime

shrink :: ClockTime -> [ClockTime]

Arbitrary DiskTemplate
Custom Arbitrary instance for DiskTemplate, which needs to handle the case of file storage being disabled at configure time.

Instance details

Methods

arbitrary :: Gen DiskTemplate

shrink :: DiskTemplate -> [DiskTemplate]

Arbitrary TagKind

Instance details

Methods

arbitrary :: Gen TagKind

shrink :: TagKind -> [TagKind]

Arbitrary AllocPolicy

Instance details

Methods

arbitrary :: Gen AllocPolicy

shrink :: AllocPolicy -> [AllocPolicy]

Arbitrary InstanceStatus

Instance details

Methods

arbitrary :: Gen InstanceStatus

shrink :: InstanceStatus -> [InstanceStatus]

Arbitrary MigrationMode

Instance details

Methods

arbitrary :: Gen MigrationMode

shrink :: MigrationMode -> [MigrationMode]

Arbitrary VerifyOptionalChecks

Instance details

Methods

arbitrary :: Gen VerifyOptionalChecks

shrink :: VerifyOptionalChecks -> [VerifyOptionalChecks]

Arbitrary CVErrorCode

Instance details

Methods

arbitrary :: Gen CVErrorCode

shrink :: CVErrorCode -> [CVErrorCode]

Arbitrary DdmSimple

Instance details

Methods

arbitrary :: Gen DdmSimple

shrink :: DdmSimple -> [DdmSimple]

Arbitrary DdmFull

Instance details

Methods

arbitrary :: Gen DdmFull

shrink :: DdmFull -> [DdmFull]

Arbitrary Hypervisor

Instance details

Methods

arbitrary :: Gen Hypervisor

shrink :: Hypervisor -> [Hypervisor]

Arbitrary OobCommand

Instance details

Methods

arbitrary :: Gen OobCommand

shrink :: OobCommand -> [OobCommand]

Arbitrary StorageType
Custom Arbitrary instance for StorageType, which needs to handle the case of file storage being disabled at configure time.

Instance details

Methods

arbitrary :: Gen StorageType

shrink :: StorageType -> [StorageType]

Arbitrary EvacMode

Instance details

Methods

arbitrary :: Gen EvacMode

shrink :: EvacMode -> [EvacMode]

Arbitrary FileDriver

Instance details

Methods

arbitrary :: Gen FileDriver

shrink :: FileDriver -> [FileDriver]

Arbitrary InstCreateMode

Instance details

Methods

arbitrary :: Gen InstCreateMode

shrink :: InstCreateMode -> [InstCreateMode]

Arbitrary RebootType

Instance details

Methods

arbitrary :: Gen RebootType

shrink :: RebootType -> [RebootType]

Arbitrary ExportMode

Instance details

Methods

arbitrary :: Gen ExportMode

shrink :: ExportMode -> [ExportMode]

Arbitrary IAllocatorTestDir

Instance details

Methods

arbitrary :: Gen IAllocatorTestDir

shrink :: IAllocatorTestDir -> [IAllocatorTestDir]

Arbitrary IAllocatorMode

Instance details

Methods

arbitrary :: Gen IAllocatorMode

shrink :: IAllocatorMode -> [IAllocatorMode]

Arbitrary NICMode

Instance details

Methods

arbitrary :: Gen NICMode

shrink :: NICMode -> [NICMode]

Arbitrary JobStatus

Instance details

Methods

arbitrary :: Gen JobStatus

shrink :: JobStatus -> [JobStatus]

Arbitrary FinalizedJobStatus

Instance details

Methods

arbitrary :: Gen FinalizedJobStatus

shrink :: FinalizedJobStatus -> [FinalizedJobStatus]

Arbitrary JobDependency

Instance details

Methods

arbitrary :: Gen JobDependency

shrink :: JobDependency -> [JobDependency]

Arbitrary JobIdDep

Instance details

Methods

arbitrary :: Gen JobIdDep

shrink :: JobIdDep -> [JobIdDep]

Arbitrary JobId

Instance details

Methods

arbitrary :: Gen JobId

shrink :: JobId -> [JobId]

Arbitrary OpSubmitPriority

Instance details

Methods

arbitrary :: Gen OpSubmitPriority

shrink :: OpSubmitPriority -> [OpSubmitPriority]

Arbitrary OpStatus

Instance details

Methods

arbitrary :: Gen OpStatus

shrink :: OpStatus -> [OpStatus]

Arbitrary ELogType

Instance details

Methods

arbitrary :: Gen ELogType

shrink :: ELogType -> [ELogType]

Arbitrary a => Arbitrary (NonEmpty a)

Instance details

Methods

arbitrary :: Gen (NonEmpty a)

shrink :: NonEmpty a -> [NonEmpty a]

(Arbitrary a, Ord a, Num a, Show a) => Arbitrary (Negative a)

Instance details

Methods

arbitrary :: Gen (Negative a)

shrink :: Negative a -> [Negative a]

(Arbitrary a, Ord a, Num a, Show a) => Arbitrary (Positive a)

Instance details

Methods

arbitrary :: Gen (Positive a)

shrink :: Positive a -> [Positive a]

(Arbitrary a, Ord a, Num a, Show a) => Arbitrary (NonNegative a)

Instance details

Methods

arbitrary :: Gen (NonNegative a)

shrink :: NonNegative a -> [NonNegative a]